Coach Information Search

A coach search function which provides coaching records and school history for a given coach

cfb_coaches(
  first = NULL,
  last = NULL,
  team = NULL,
  year = NULL,
  min_year = NULL,
  max_year = NULL
)

Arguments

first
(String optional): First name for the coach you are trying to look up

last
(String optional): Last name for the coach you are trying to look up

team
(String optional): Team - Select a valid team, D1 football

year
(Integer optional): Year, 4 digit format (YYYY).

min_year
(Integer optional): Minimum Year filter (inclusive), 4 digit format (YYYY).

max_year
(Integer optional): Maximum Year filter (inclusive), 4 digit format (YYYY).

Source

https://api.collegefootballdata.com/coaches

Value

A data frame with coach information with the following columns:

first_name

character. First name of coach.

last_name

character. Last name of coach.

school

character. School of coach.

year

integer. Season of record.

games

integer. Games as coach.

wins

integer. Wins for the season.

losses

integer. Losses for the season.

ties

integer. Ties for the season.

preseason_rank

integer. Preseason rank for the school of coach.

postseason_rank

integer. Postseason rank for the school of coach.

srs

character. Simple Rating System adjustment for team.

sp_overall

character. Bill Connelly's SP+ overall for team.

sp_offense

character. Bill Connelly's SP+ offense for team.

sp_defense

character. Bill Connelly's SP+ defense for team.

Examples

cfb_coaches(first = "Nick", last = "Saban", team = 'alabama')

#> 2020-10-13 00:05:43: Scraping coaches data...

#>    first_name  last_name  school year games wins losses ties preseason_rank
#> 1     Wallace       Wade Alabama 1923    10    7      2    1             NA
#> 2     Wallace       Wade Alabama 1924     9    8      1    0             NA
#> 3     Wallace       Wade Alabama 1925    10   10      0    0             NA
#> 4     Wallace       Wade Alabama 1926    10    9      0    1             NA
#> 5     Wallace       Wade Alabama 1927    10    5      4    1             NA
#> 6     Wallace       Wade Alabama 1928     9    6      3    0             NA
#> 7     Wallace       Wade Alabama 1929     9    6      3    0             NA
#> 8     Wallace       Wade Alabama 1930    10   10      0    0             NA
#> 9       Frank     Thomas Alabama 1931    10    9      1    0             NA
#> 10      Frank     Thomas Alabama 1932    10    8      2    0             NA
#> 11      Frank     Thomas Alabama 1933     9    7      1    1             NA
#> 12      Frank     Thomas Alabama 1934    10   10      0    0             NA
#> 13      Frank     Thomas Alabama 1935     9    6      2    1             NA
#> 14      Frank     Thomas Alabama 1936     9    8      0    1             NA
#> 15      Frank     Thomas Alabama 1937    10    9      1    0             NA
#> 16      Frank     Thomas Alabama 1938     9    7      1    1             NA
#> 17      Frank     Thomas Alabama 1939     9    5      3    1             NA
#> 18      Frank     Thomas Alabama 1940     9    7      2    0             NA
#> 19      Frank     Thomas Alabama 1941    11    9      2    0             NA
#> 20      Frank     Thomas Alabama 1942    11    8      3    0             NA
#> 21      Frank     Thomas Alabama 1944     9    5      2    2             NA
#> 22      Frank     Thomas Alabama 1945    10   10      0    0             NA
#> 23      Frank     Thomas Alabama 1946    11    7      4    0             NA
#> 24        Red       Drew Alabama 1947    11    8      3    0             NA
#> 25        Red       Drew Alabama 1948    11    6      4    1             NA
#> 26        Red       Drew Alabama 1949    10    6      3    1             NA
#> 27        Red       Drew Alabama 1950    11    9      2    0             NA
#> 28        Red       Drew Alabama 1951    11    5      6    0              9
#> 29        Red       Drew Alabama 1952    12   10      2    0             NA
#> 30        Red       Drew Alabama 1953    12    6      3    3              5
#> 31        Red       Drew Alabama 1954    11    4      5    2             14
#> 32       Bear     Bryant Alabama 1958    10    5      4    1             NA
#> 33       Bear     Bryant Alabama 1959    11    7      2    2             NA
#> 34       Bear     Bryant Alabama 1960    11    8      1    2             NA
#> 35       Bear     Bryant Alabama 1961    11   11      0    0              3
#> 36       Bear     Bryant Alabama 1962    11   10      1    0              3
#> 37       Bear     Bryant Alabama 1963    11    9      2    0              3
#> 38       Bear     Bryant Alabama 1964    11   10      1    0              6
#> 39       Bear     Bryant Alabama 1965    11    9      1    1              5
#> 40       Bear     Bryant Alabama 1966    11   11      0    0              1
#> 41       Bear     Bryant Alabama 1967    11    8      2    1              2
#> 42       Bear     Bryant Alabama 1968    11    8      3    0             10
#> 43       Bear     Bryant Alabama 1969    11    6      5    0             13
#> 44       Bear     Bryant Alabama 1970    12    6      5    1             16
#> 45       Bear     Bryant Alabama 1971    12   11      1    0             16
#> 46       Bear     Bryant Alabama 1972    12   10      2    0              7
#> 47       Bear     Bryant Alabama 1973    12   11      1    0              6
#> 48       Bear     Bryant Alabama 1974    12   11      1    0              4
#> 49       Bear     Bryant Alabama 1975    12   11      1    0              2
#> 50       Bear     Bryant Alabama 1976    12    9      3    0              6
#> 51       Bear     Bryant Alabama 1977    12   11      1    0              6
#> 52       Bear     Bryant Alabama 1978    12   11      1    0              1
#> 53       Bear     Bryant Alabama 1979    12   12      0    0              2
#> 54       Bear     Bryant Alabama 1980    12   10      2    0              2
#> 55       Bear     Bryant Alabama 1981    12    9      2    1              4
#> 56       Bear     Bryant Alabama 1982    12    8      4    0              3
#> 57       Bill      Curry Alabama 1987    12    7      5    0             NA
#> 58       Bill      Curry Alabama 1988    12    9      3    0             14
#> 59       Bill      Curry Alabama 1989    12   10      2    0             16
#> 60       Gene  Stallings Alabama 1990    12    7      5    0             12
#> 61       Gene  Stallings Alabama 1991    12   11      1    0             22
#> 62       Gene  Stallings Alabama 1992    13   13      0    0              9
#> 63       Gene  Stallings Alabama 1993    13    9      3    1              2
#> 64       Gene  Stallings Alabama 1994    13   12      1    0             12
#> 65       Gene  Stallings Alabama 1995    11    8      3    0             10
#> 66       Gene  Stallings Alabama 1996    13   10      3    0             15
#> 67     Dennis Franchione Alabama 2001    12    7      5    0             25
#> 68     Dennis Franchione Alabama 2002    13   10      3    0             NA
#> 69       Nick      Saban Alabama 2007    13    7      6    0             NA
#> 70       Nick      Saban Alabama 2008    14   12      2    0             24
#> 71       Nick      Saban Alabama 2009    14   14      0    0              5
#> 72       Nick      Saban Alabama 2010    13   10      3    0              1
#> 73       Nick      Saban Alabama 2011    13   12      1    0              2
#> 74       Nick      Saban Alabama 2012    14   13      1    0              2
#> 75       Nick      Saban Alabama 2013    13   11      2    0              1
#> 76       Nick      Saban Alabama 2014    14   12      2    0              2
#> 77       Nick      Saban Alabama 2015    15   14      1    0              3
#> 78       Nick      Saban Alabama 2016    15   14      1    0              1
#> 79       Nick      Saban Alabama 2017    14   13      1    0              1
#> 80       Nick      Saban Alabama 2018    15   14      1    0              1
#> 81       Nick      Saban Alabama 2019    13   11      2    0              2
#> 82       Nick      Saban Alabama 2020     0    0      0    0              3
#>    postseason_rank  srs sp_overall sp_offense sp_defense
#> 1               NA  3.8       <NA>       <NA>       <NA>
#> 2               NA 22.3       <NA>       <NA>       <NA>
#> 3               NA 21.7       <NA>       <NA>       <NA>
#> 4               NA 21.4       <NA>       <NA>       <NA>
#> 5               NA  2.0       <NA>       <NA>       <NA>
#> 6               NA 13.7       <NA>       <NA>       <NA>
#> 7               NA  9.3       <NA>       <NA>       <NA>
#> 8               NA 27.0       <NA>       <NA>       <NA>
#> 9               NA 20.3       <NA>       <NA>       <NA>
#> 10              NA 15.8       <NA>       <NA>       <NA>
#> 11              NA 11.3       <NA>       <NA>       <NA>
#> 12              NA 30.2       <NA>       <NA>       <NA>
#> 13              NA 19.5       <NA>       <NA>       <NA>
#> 14               4 20.0       <NA>       <NA>       <NA>
#> 15               4 18.3       <NA>       <NA>       <NA>
#> 16              13 13.0       <NA>       <NA>       <NA>
#> 17              NA 15.3       <NA>       <NA>       <NA>
#> 18              NA  9.2       <NA>       <NA>       <NA>
#> 19              20 19.1       <NA>       <NA>       <NA>
#> 20              10 23.4       <NA>       <NA>       <NA>
#> 21              NA 12.4       <NA>       <NA>       <NA>
#> 22               2 36.4       <NA>       <NA>       <NA>
#> 23              NA 14.3       <NA>       <NA>       <NA>
#> 24               6 16.8       <NA>       <NA>       <NA>
#> 25              NA 12.0       <NA>       <NA>       <NA>
#> 26              NA 13.1       <NA>       <NA>       <NA>
#> 27              16 18.0       <NA>       <NA>       <NA>
#> 28              NA  4.1       <NA>       <NA>       <NA>
#> 29               9 24.2       <NA>       <NA>       <NA>
#> 30              13 13.4       <NA>       <NA>       <NA>
#> 31              NA  8.9       <NA>       <NA>       <NA>
#> 32              NA 10.3       <NA>       <NA>       <NA>
#> 33              10 11.0       <NA>       <NA>       <NA>
#> 34               9 15.8       <NA>       <NA>       <NA>
#> 35               1 27.6       <NA>       <NA>       <NA>
#> 36               5 29.0       <NA>       <NA>       <NA>
#> 37               8 20.0       <NA>       <NA>       <NA>
#> 38               1 21.3       <NA>       <NA>       <NA>
#> 39               1 24.5       <NA>       <NA>       <NA>
#> 40               3 28.5       <NA>       <NA>       <NA>
#> 41               8 15.5       <NA>       <NA>       <NA>
#> 42              17 13.8       <NA>       <NA>       <NA>
#> 43              NA 12.5       <NA>       <NA>       <NA>
#> 44              NA 17.4       19.7       39.2       21.3
#> 45               4 33.0         37       42.1        6.5
#> 46               7 31.1       31.8         43       11.6
#> 47               4 31.5       32.6       44.6       13.1
#> 48               5 25.8       29.5       35.7        6.6
#> 49               3 26.6       28.8       36.2        7.2
#> 50              11 21.4         26       35.7       10.7
#> 51               2 28.5       32.5       41.3        9.3
#> 52               1 26.3       33.9         42        8.2
#> 53               1 26.4       27.3       35.6        8.6
#> 54               6 26.8       28.7       36.2        7.7
#> 55               7 18.0       20.2         32       12.1
#> 56              NA 18.6       20.4       37.8       18.4
#> 57              NA 13.1       18.6       32.3       15.3
#> 58              17 13.6       16.6       32.9       17.7
#> 59               9 17.5         23       39.4         18
#> 60              NA  8.2         13       27.1       14.3
#> 61               5 13.1       17.3       29.9       13.9
#> 62               1 20.7       27.2       35.4        8.6
#> 63              14 14.0       18.6         32       14.4
#> 64               5 12.8       19.9       31.1         12
#> 65              21  6.8          9       28.7       20.3
#> 66              11 11.6       17.1       31.2       15.3
#> 67              NA 10.4       11.8       32.9       22.7
#> 68              11 17.7       23.6       37.6       14.4
#> 69              NA  8.3       12.8       30.1       17.2
#> 70               6 17.9         23       34.2       10.9
#> 71               1 27.5       30.2         35        6.5
#> 72              10 24.1       30.1       43.9       14.2
#> 73               1 30.5       32.5       39.3        6.4
#> 74               1 32.2       36.5       45.7       10.7
#> 75               7 25.4       34.1       45.6       12.3
#> 76               4 26.5       34.3       44.5       10.7
#> 77               1 27.3         29       37.7        9.1
#> 78               2 33.6       31.7         40        8.6
#> 79               1 26.3       33.5       38.7        5.6
#> 80              NA 33.5       36.3       50.9       14.2
#> 81               8 28.8       32.4       47.7       14.6
#> 82              NA <NA>       <NA>       <NA>       <NA>