Get results information from games

Get results information from games

cfb_game_info(
  year,
  week = NULL,
  season_type = "regular",
  team = NULL,
  home_team = NULL,
  away_team = NULL,
  conference = NULL,
  game_id = NULL,
  quarter_scores = FALSE
)

Arguments

year
(Integer required): Year, 4 digit format(YYYY)

week
(Integer optional): Week - values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)

season_type
(String default regular): Select Season Type: regular, postseason, or both

team
(String optional): D-I Team

home_team
(String optional): Home D-I Team

away_team
(String optional): Away D-I Team

conference
(String optional): Conference abbreviation - Select a valid FBS conference Conference abbreviations P5: ACC, B12, B1G, SEC, PAC Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

game_id
(Integer optional): Game ID filter for querying a single game Can be found using the `cfb_game_info()` function

quarter_scores
(Logical default FALSE): This is a parameter to return the list columns that give the score at each quarter: home_line_scores and away_line scores. I have defaulted the parameter to false so that you will not have to go to the trouble of dropping it.

Examples

cfb_game_info(2018, week = 1)

#>           id season week season_type               start_date start_time_tbd
#> 1  401013357   2018    1     regular 2018-08-25T21:30:00.000Z             NA
#> 2  401014972   2018    1     regular 2018-08-25T23:00:00.000Z             NA
#> 3  401022510   2018    1     regular 2018-08-25T23:30:00.000Z             NA
#> 4  401013437   2018    1     regular 2018-08-26T02:00:00.000Z             NA
#> 5  401021651   2018    1     regular 2018-08-30T23:00:00.000Z             NA
#> 6  401019470   2018    1     regular 2018-08-30T23:00:00.000Z             NA
#> 7  401013328   2018    1     regular 2018-08-30T23:00:00.000Z             NA
#> 8  401020671   2018    1     regular 2018-08-30T23:00:00.000Z             NA
#> 9  401021655   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 10 401013342   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 11 401013086   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 12 401012675   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 13 401014979   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 14 401013489   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 15 401012258   2018    1     regular 2018-08-31T00:30:00.000Z             NA
#> 16 401022511   2018    1     regular 2018-08-31T02:00:00.000Z             NA
#> 17 401020673   2018    1     regular 2018-08-31T22:00:00.000Z             NA
#> 18 401020672   2018    1     regular 2018-08-31T22:30:00.000Z             NA
#> 19 401012880   2018    1     regular 2018-08-31T23:00:00.000Z             NA
#> 20 401013089   2018    1     regular 2018-08-31T23:00:00.000Z             NA
#> 21 401013354   2018    1     regular 2018-09-01T01:00:00.000Z             NA
#> 22 401022517   2018    1     regular 2018-09-01T01:00:00.000Z             NA
#> 23 401022512   2018    1     regular 2018-09-01T01:00:00.000Z             NA
#> 24 401012680   2018    1     regular 2018-09-01T01:30:00.000Z             NA
#> 25 401012682   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 26 401012872   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 27 401019473   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 28 401013091   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 29 401012820   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 30 401013065   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 31 401013072   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 32 401012804   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 33 401012256   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 34 401014977   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 35 401012253   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 36 401013088   2018    1     regular 2018-09-01T16:20:00.000Z             NA
#> 37 401013090   2018    1     regular 2018-09-01T16:30:00.000Z             NA
#> 38 401013087   2018    1     regular 2018-09-01T17:00:00.000Z             NA
#> 39 401022514   2018    1     regular 2018-09-01T18:00:00.000Z             NA
#> 40 401020675   2018    1     regular 2018-09-01T18:00:00.000Z             NA
#> 41 401012683   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 42 401012251   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 43 401012248   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 44 401012257   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 45 401013092   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 46 401012250   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 47 401013349   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 48 401012862   2018    1     regular 2018-09-01T19:40:00.000Z             NA
#> 49 401012678   2018    1     regular 2018-09-01T20:00:00.000Z             NA
#> 50 401012255   2018    1     regular 2018-09-01T20:00:00.000Z             NA
#> 51 401012247   2018    1     regular 2018-09-01T20:00:00.000Z             NA
#> 52 401012684   2018    1     regular 2018-09-01T20:00:00.000Z             NA
#> 53 401020674   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 54 401021656   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 55 401013094   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 56 401014973   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 57 401021653   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 58 401014975   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 59 401013448   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 60 401019472   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 61 401019474   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 62 401013048   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 63 401021654   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 64 401014974   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 65 401012851   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 66 401012679   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 67 401019476   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 68 401021652   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 69 401013057   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 70 401014978   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 71 401020676   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 72 401012249   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 73 401012888   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 74 401014980   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 75 401012254   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 76 401012259   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 77 401014976   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 78 401013026   2018    1     regular 2018-09-02T00:00:00.000Z             NA
#> 79 401012681   2018    1     regular 2018-09-02T00:00:00.000Z             NA
#> 80 401012246   2018    1     regular 2018-09-02T00:00:00.000Z             NA
#> 81 401022515   2018    1     regular 2018-09-02T00:00:00.000Z             NA
#> 82 401022516   2018    1     regular 2018-09-02T02:00:00.000Z             NA
#> 83 401012677   2018    1     regular 2018-09-02T02:30:00.000Z             NA
#> 84 401012676   2018    1     regular 2018-09-02T02:45:00.000Z             NA
#> 85 401019475   2018    1     regular 2018-09-02T03:00:00.000Z             NA
#> 86 401079833   2018    1     regular 2018-09-02T19:30:00.000Z             NA
#> 87 401013095   2018    1     regular 2018-09-02T23:30:00.000Z             NA
#> 88 401013096   2018    1     regular 2018-09-04T00:00:00.000Z             NA
#>    neutral_site conference_game attendance venue_id
#> 1         FALSE           FALSE       8684     3985
#> 2         FALSE           FALSE      20050     3895
#> 3         FALSE            TRUE      31007     5388
#> 4         FALSE           FALSE      20322     3601
#> 5         FALSE           FALSE          0     4246
#> 6         FALSE            TRUE          0     3892
#> 7         FALSE           FALSE          0     3953
#> 8         FALSE           FALSE          0     3919
#> 9         FALSE           FALSE          0     3817
#> 10        FALSE            TRUE          0     3907
#> 11        FALSE           FALSE          0     4729
#> 12        FALSE           FALSE          0      587
#> 13        FALSE           FALSE          0     3803
#> 14        FALSE           FALSE          0     3646
#> 15        FALSE           FALSE          0     3795
#> 16        FALSE           FALSE          0     3935
#> 17        FALSE           FALSE          0     3980
#> 18        FALSE           FALSE          0     3912
#> 19        FALSE           FALSE          0     3936
#> 20        FALSE           FALSE          0     3982
#> 21        FALSE           FALSE          0      347
#> 22        FALSE           FALSE          0     3816
#> 23        FALSE           FALSE          0     3940
#> 24         TRUE           FALSE          0     3937
#> 25        FALSE           FALSE          0     3861
#> 26         TRUE           FALSE          0     3719
#> 27        FALSE           FALSE          0     3806
#> 28        FALSE           FALSE          0     3670
#> 29        FALSE           FALSE          0     3754
#> 30        FALSE           FALSE          0     3835
#> 31        FALSE           FALSE          0     3616
#> 32        FALSE           FALSE          0     3832
#> 33        FALSE           FALSE          0     3994
#> 34        FALSE           FALSE          0     3895
#> 35         TRUE           FALSE          0     3891
#> 36        FALSE           FALSE          0     3836
#> 37        FALSE           FALSE          0      721
#> 38        FALSE           FALSE          0     3615
#> 39        FALSE           FALSE          0     3713
#> 40        FALSE           FALSE          0     3876
#> 41        FALSE           FALSE          0     3984
#> 42        FALSE           FALSE          0     3683
#> 43         TRUE           FALSE          0     5348
#> 44         TRUE           FALSE          0     3628
#> 45        FALSE           FALSE          0     3752
#> 46        FALSE           FALSE          0     3917
#> 47        FALSE           FALSE     105232     3632
#> 48        FALSE           FALSE          0     3793
#> 49        FALSE           FALSE          0      477
#> 50        FALSE           FALSE          0     3838
#> 51        FALSE           FALSE          0     3887
#> 52        FALSE           FALSE          0     3831
#> 53        FALSE           FALSE          0     3965
#> 54        FALSE           FALSE          0     3975
#> 55        FALSE           FALSE          0     3923
#> 56        FALSE           FALSE          0     4418
#> 57        FALSE           FALSE          0     3608
#> 58        FALSE           FALSE          0     3996
#> 59        FALSE           FALSE          0     3992
#> 60        FALSE           FALSE          0     3886
#> 61        FALSE           FALSE          0     3805
#> 62        FALSE           FALSE          0     3833
#> 63        FALSE           FALSE          0     3666
#> 64        FALSE           FALSE          0     3797
#> 65        FALSE           FALSE          0      218
#> 66        FALSE           FALSE          0     1056
#> 67        FALSE           FALSE          0     3674
#> 68        FALSE           FALSE          0     3625
#> 69        FALSE           FALSE          0     3636
#> 70        FALSE           FALSE          0     3815
#> 71        FALSE           FALSE          0     3739
#> 72        FALSE           FALSE          0     3634
#> 73        FALSE           FALSE          0     3855
#> 74        FALSE           FALSE          0     3946
#> 75        FALSE           FALSE          0     3693
#> 76        FALSE           FALSE          0     3973
#> 77        FALSE           FALSE          0     3825
#> 78        FALSE           FALSE          0     4727
#> 79        FALSE           FALSE          0     3626
#> 80         TRUE           FALSE      57280     4013
#> 81        FALSE           FALSE          0     3971
#> 82        FALSE           FALSE          0     3660
#> 83        FALSE           FALSE          0     3947
#> 84        FALSE           FALSE          0     3619
#> 85        FALSE           FALSE          0     3610
#> 86        FALSE           FALSE          0     3699
#> 87         TRUE           FALSE          0     3687
#> 88        FALSE            TRUE          0     3697
#>                                                       venue home_id
#> 1                             Warren McGuirk Alumni Stadium     113
#> 2                                              Rice Stadium     242
#> 3                                            Canvas Stadium      36
#> 4                                    Aggie Memorial Stadium     166
#> 5                                     Georgia State Stadium    2247
#> 6                                          Rentschler Field      41
#> 7                                          TCF Bank Stadium     135
#> 8                                         Scheumann Stadium    2050
#> 9                                            Malone Stadium    2433
#> 10                                         Ross-Ade Stadium    2509
#> 11                           Benson Field at Yulman Stadium    2655
#> 12                                      Rice-Eccles Stadium     254
#> 13                                             Legion Field       5
#> 14                                    Boone Pickens Stadium     197
#> 15                                               Kyle Field     245
#> 16                                          Spartan Stadium      23
#> 17                                            Waldo Stadium    2711
#> 18                                        Rynearson Stadium    2199
#> 19                                          Spartan Stadium     127
#> 20                                     Wallace Wade Stadium     150
#> 21                                     Camp Randall Stadium     275
#> 22                                           Mackay Stadium    2440
#> 23                                         Stanford Stadium      24
#> 24                      Sports Authority Field at Mile High      36
#> 25                                             Ohio Stadium     194
#> 26                                              FedEx Field     120
#> 27                                  Lincoln Financial Field     218
#> 28                                    Carter-Finley Stadium     152
#> 29                             High Point Solutions Stadium     164
#> 30                 Gaylord Family Oklahoma Memorial Stadium     201
#> 31                                   Amon G. Carter Stadium    2628
#> 32                                         Memorial Stadium     356
#> 33                                   Williams-Brice Stadium    2579
#> 34                                             Rice Stadium     242
#> 35                                              NRG Stadium    2641
#> 36                                         Memorial Stadium     228
#> 37               Bobby Dodd Stadium at Historic Grant Field      59
#> 38                                           Alumni Stadium     103
#> 39                                           Falcon Stadium    2005
#> 40                                            Peden Stadium     195
#> 41                                     War Memorial Stadium    2751
#> 42                                     Commonwealth Stadium      96
#> 43                                    Mercedes-Benz Stadium       2
#> 44                                  Bank of America Stadium    2633
#> 45                                              Heinz Field     221
#> 46                                          Sanford Stadium      61
#> 47                                           Beaver Stadium     213
#> 48                                          Kinnick Stadium    2294
#> 49                            Los Angeles Memorial Coliseum      30
#> 50                                             Faurot Field     142
#> 51 Donald W. Reynolds Razorback Stadium Frank Broyles Field       8
#> 52                              California Memorial Stadium      25
#> 53                                               UB Stadium    2084
#> 54                                Veterans Memorial Stadium    2653
#> 55                                            Scott Stadium     258
#> 56                                 Jerry Richardson Stadium    2429
#> 57                                          Paulson Stadium     290
#> 58                                            Yager Stadium     193
#> 59                                         Williams Stadium    2335
#> 60                                    Raymond James Stadium      58
#> 61                            Liberty Bowl Memorial Stadium     235
#> 62                                         Memorial Stadium    2305
#> 63                                              Cajun Field     309
#> 64                                     Ladd Peebles Stadium       6
#> 65                                              FIU Stadium    2229
#> 66                                                Rose Bowl      26
#> 67                    Skelly Field at H. A. Chapman Stadium     202
#> 68                                  Centennial Bank Stadium    2032
#> 69                      Bill Snyder Family Football Stadium    2306
#> 70                                    M. M. Roberts Stadium    2572
#> 71                                               Glass Bowl    2649
#> 72                                 Ben Hill Griffin Stadium      57
#> 73                                       Notre Dame Stadium      87
#> 74                                         Sun Bowl Stadium    2638
#> 75                                       Davis Wade Stadium     344
#> 76                                       Vanderbilt Stadium     238
#> 77                                           Apogee Stadium     249
#> 78                                           McLane Stadium     239
#> 79                                           Autzen Stadium    2483
#> 80                                    Camping World Stadium     333
#> 81                                       University Stadium     167
#> 82                                          Bulldog Stadium     278
#> 83                                        Sun Devil Stadium       9
#> 84                                          Arizona Stadium      12
#> 85                                            Aloha Stadium      62
#> 86                                    Dowdy-Ficklen Stadium     151
#> 87                                             AT&T Stadium      99
#> 88              Bobby Bowden Field at Doak Campbell Stadium      52
#>                home_team   home_conference home_points   home_post_win_prob
#> 1                  UMass  FBS Independents          63   0.9984542422162311
#> 2                   Rice    Conference USA          31  0.41611197976904707
#> 3         Colorado State     Mountain West          34   0.3317397678892768
#> 4       New Mexico State  FBS Independents           7  0.00427210984632397
#> 5          Georgia State          Sun Belt          24    0.869590187971506
#> 6            Connecticut American Athletic          17 0.004203463433869054
#> 7              Minnesota           Big Ten          48   0.9999638364577065
#> 8             Ball State      Mid-American          42   0.9999720937904223
#> 9       Louisiana Monroe          Sun Belt          34   0.5720859305615573
#> 10                Purdue           Big Ten          27  0.15206580002046932
#> 11                Tulane American Athletic          17   0.2196613212325033
#> 12                  Utah            Pac-12          41   0.9997998688829266
#> 13                   UAB    Conference USA          52    0.999997345694123
#> 14        Oklahoma State            Big 12          58   0.9992698560515085
#> 15             Texas A&M               SEC          59   0.9999928023514462
#> 16        San José State     Mountain West          38   0.2255818432658651
#> 17      Western Michigan      Mid-American          42  0.10069975562117696
#> 18      Eastern Michigan      Mid-American          51   0.9995510164694754
#> 19        Michigan State           Big Ten          38    0.906183590742433
#> 20                  Duke               ACC          34   0.9973846191967987
#> 21             Wisconsin           Big Ten          34   0.9999658482595446
#> 22                Nevada     Mountain West          72   0.9999946772201084
#> 23              Stanford            Pac-12          31   0.9937677403365777
#> 24        Colorado State     Mountain West          13 0.003935607921436278
#> 25            Ohio State           Big Ten          77   0.9999435276188426
#> 26              Maryland           Big Ten          34   0.9131500976054453
#> 27                Temple American Athletic          17  0.08018018010651985
#> 28              NC State               ACC          24    0.839090960760694
#> 29               Rutgers           Big Ten          35   0.9992581309620951
#> 30              Oklahoma            Big 12          63   0.9999928028169979
#> 31                   TCU            Big 12          55   0.9999921756502518
#> 32              Illinois           Big Ten          31   0.8340153107033903
#> 33        South Carolina               SEC          49   0.9999813927842912
#> 34                  Rice    Conference USA          27  0.05005196033053275
#> 35            Texas Tech            Big 12          27 0.007730744808323005
#> 36               Clemson               ACC          48   0.9999840784279569
#> 37          Georgia Tech               ACC          41   0.9999959877845358
#> 38        Boston College               ACC          55   0.9998255691998653
#> 39             Air Force     Mountain West          38   0.9999947029448708
#> 40                  Ohio      Mid-American          38   0.4527258093446935
#> 41               Wyoming     Mountain West          19 0.005358724710402157
#> 42              Kentucky               SEC          35   0.9948967001914114
#> 43                Auburn               SEC          21   0.6081821946558714
#> 44             Tennessee               SEC          14 0.004273317866598812
#> 45            Pittsburgh               ACC          33   0.9997365885384938
#> 46               Georgia               SEC          45    0.999992203255419
#> 47            Penn State           Big Ten          45    0.980509601522245
#> 48                  Iowa           Big Ten          33   0.9987837404630044
#> 49                   USC            Pac-12          43   0.9967409790322357
#> 50              Missouri               SEC          51    0.995968150359603
#> 51              Arkansas               SEC          55   0.9967431808920783
#> 52            California            Pac-12          24   0.8968200177849878
#> 53               Buffalo      Mid-American          48   0.9999142322720109
#> 54                  Troy          Sun Belt          20  0.05923029255824867
#> 55              Virginia               ACC          42     0.99897198899527
#> 56             Charlotte    Conference USA          34    0.999984164464906
#> 57      Georgia Southern          Sun Belt          37   0.9987286659646343
#> 58            Miami (OH)      Mid-American          28   0.2914353097621623
#> 59               Liberty  FBS Independents          52   0.9998402730839259
#> 60         South Florida American Athletic          34   0.9999637417382397
#> 61               Memphis American Athletic          66   0.9999958345824962
#> 62                Kansas            Big 12          23  0.15578688342620697
#> 63             Louisiana          Sun Belt          49   0.9997927217865997
#> 64         South Alabama          Sun Belt          26  0.17346633829827812
#> 65 Florida International    Conference USA          28  0.08816900970878677
#> 66                  UCLA            Pac-12          17   0.2196041237734069
#> 67                 Tulsa American Athletic          38   0.9977510896828313
#> 68        Arkansas State          Sun Belt          48    0.998884580500493
#> 69          Kansas State            Big 12          27   0.5711801005392997
#> 70  Southern Mississippi    Conference USA          55   0.9994952216812634
#> 71                Toledo      Mid-American          66   0.9999873144404265
#> 72               Florida               SEC          53   0.9998257838347568
#> 73            Notre Dame  FBS Independents          24     0.92783043145901
#> 74                  UTEP    Conference USA          10 0.006031035556264542
#> 75     Mississippi State               SEC          63   0.9999955955222241
#> 76            Vanderbilt               SEC          35   0.9996203330946147
#> 77           North Texas    Conference USA          46   0.9987373302064423
#> 78                Baylor            Big 12          55   0.9998295720824828
#> 79                Oregon            Pac-12          58   0.9998393673219237
#> 80               Alabama               SEC          51   0.9999255686466408
#> 81            New Mexico     Mountain West          62   0.9991581186438564
#> 82          Fresno State     Mountain West          79   0.9995402711314478
#> 83         Arizona State            Pac-12          49   0.9999920070738963
#> 84               Arizona            Pac-12          23    0.252378224443892
#> 85               Hawai'i     Mountain West          59   0.9796414129712154
#> 86         East Carolina American Athletic          23   0.4801521413827188
#> 87                   LSU               SEC          33   0.9559552996558128
#> 88         Florida State               ACC           3  0.31246770456927847
#>    away_id                away_team   away_conference away_points
#> 1     2184                 Duquesne              <NA>          15
#> 2     2504             Prairie View              <NA>          28
#> 3       62                  Hawai'i     Mountain West          43
#> 4     2751                  Wyoming     Mountain West          29
#> 5      338           Kennesaw State              <NA>          20
#> 6     2116                      UCF American Athletic          56
#> 7      166         New Mexico State  FBS Independents          10
#> 8     2115      Central Connecticut              <NA>           6
#> 9     2545   Southeastern Louisiana              <NA>          31
#> 10      77             Northwestern           Big Ten          31
#> 11     154              Wake Forest               ACC          23
#> 12    2692              Weber State              <NA>          10
#> 13    2542           Savannah State              <NA>           0
#> 14    2623           Missouri State              <NA>          17
#> 15    2466       Northwestern State              <NA>           7
#> 16     302                 UC Davis              <NA>          44
#> 17     183                 Syracuse               ACC          55
#> 18    2405                 Monmouth              <NA>          17
#> 19     328               Utah State     Mountain West          31
#> 20     349                     Army  FBS Independents          14
#> 21      98         Western Kentucky    Conference USA           3
#> 22    2502           Portland State              <NA>          19
#> 23      21          San Diego State     Mountain West          10
#> 24      38                 Colorado            Pac-12          45
#> 25     204             Oregon State            Pac-12          31
#> 26     251                    Texas            Big 12          29
#> 27     222                Villanova              <NA>          19
#> 28     256            James Madison              <NA>          13
#> 29     326              Texas State          Sun Belt           7
#> 30    2226         Florida Atlantic    Conference USA          14
#> 31    2582                 Southern              <NA>           7
#> 32    2309               Kent State      Mid-American          24
#> 33     324         Coastal Carolina          Sun Belt          15
#> 34     248                  Houston American Athletic          45
#> 35     145                 Ole Miss               SEC          47
#> 36     231                   Furman              <NA>           7
#> 37    2016             Alcorn State              <NA>           0
#> 38     113                    UMass  FBS Independents          21
#> 39    2619              Stony Brook              <NA>           0
#> 40      47                   Howard              <NA>          32
#> 41     265         Washington State            Pac-12          41
#> 42    2117         Central Michigan      Mid-American          20
#> 43     264               Washington            Pac-12          16
#> 44     277            West Virginia            Big 12          40
#> 45     399                   Albany              <NA>           7
#> 46    2046              Austin Peay              <NA>           0
#> 47    2026        Appalachian State          Sun Belt          38
#> 48    2459        Northern Illinois      Mid-American           7
#> 49    2439                     UNLV     Mountain West          21
#> 50    2630                UT Martin              <NA>          14
#> 51    2197         Eastern Illinois              <NA>          20
#> 52     153           North Carolina               ACC          17
#> 53    2169           Delaware State              <NA>          10
#> 54      68              Boise State     Mountain West          56
#> 55     257                 Richmond              <NA>          13
#> 56    2230                  Fordham              <NA>          10
#> 57    2569     South Carolina State              <NA>           6
#> 58     276                 Marshall    Conference USA          35
#> 59     295             Old Dominion    Conference USA          10
#> 60    2210                     Elon              <NA>          14
#> 61    2382                   Mercer              <NA>          14
#> 62    2447                 Nicholls              <NA>          26
#> 63    2755                Grambling              <NA>          17
#> 64    2348           Louisiana Tech    Conference USA          30
#> 65      84                  Indiana           Big Ten          38
#> 66    2132               Cincinnati American Athletic          26
#> 67    2110         Central Arkansas              <NA>          27
#> 68    2546 Southeast Missouri State              <NA>          21
#> 69     233             South Dakota              <NA>          24
#> 70    2296            Jackson State              <NA>           7
#> 71    2678                      VMI              <NA>           3
#> 72    2127      Charleston Southern              <NA>           6
#> 73     130                 Michigan           Big Ten          17
#> 74    2464         Northern Arizona              <NA>          30
#> 75    2617        Stephen F. Austin              <NA>           6
#> 76    2393         Middle Tennessee    Conference USA           7
#> 77    2567                      SMU American Athletic          23
#> 78    2000        Abilene Christian              <NA>          27
#> 79     189            Bowling Green      Mid-American          24
#> 80      97               Louisville               ACC          14
#> 81    2916           Incarnate Word              <NA>          30
#> 82      70                    Idaho              <NA>          13
#> 83    2636           UT San Antonio    Conference USA           7
#> 84     252                      BYU  FBS Independents          28
#> 85    2426                     Navy American Athletic          41
#> 86    2448       North Carolina A&T              <NA>          28
#> 87    2390                    Miami               ACC          17
#> 88     259            Virginia Tech               ACC          24
#>          away_post_win_prob  excitement_index
#> 1      0.001545757783768864  1.04356563705626
#> 2        0.5838880202309529  5.60047334737113
#> 3        0.6682602321107232  3.41476208077519
#> 4         0.995727890153676  1.66126552823916
#> 5         0.130409812028494  5.82077639878451
#> 6         0.995796536566131  1.56456197626456
#> 7    0.00003616354229352137  2.93995772041473
#> 8   0.000027906209577710506  1.88115888090199
#> 9        0.4279140694384427  8.05682292541858
#> 10       0.8479341999795307  7.10225238859315
#> 11       0.7803386787674967  8.79739197306156
#> 12    0.0002001311170733988  4.43127659516895
#> 13 0.0000026543058769812333  1.95260988797199
#> 14    0.0007301439484914685  1.24980300256497
#> 15  0.000007197648553836089  1.67138831446935
#> 16       0.7744181567341348   4.8309356181343
#> 17        0.899300244378823  2.60833231242055
#> 18    0.0004489835305245826  1.40204597863049
#> 19      0.09381640925756696  5.99194648765147
#> 20    0.0026153808032013393  2.23178059939514
#> 21   0.00003415174045540148  2.19417473064598
#> 22  0.000005322779891625551  4.08035648446446
#> 23     0.006232259663422335  4.61789372380646
#> 24       0.9960643920785637  1.66906339105137
#> 25   0.00005647238115735753  1.48460943276462
#> 26       0.0868499023945547   5.0281936334039
#> 27       0.9198198198934802  5.84515711028696
#> 28        0.160909039239306  4.87484268779657
#> 29     0.000741869037904852  1.79676055134555
#> 30  0.000007197183002127794  1.41499938965016
#> 31  0.000007824349748219994   1.2607675959304
#> 32       0.1659846892966097  5.88311152409072
#> 33  0.000018607215708787983  1.30185382925217
#> 34       0.9499480396694673  4.94649067286535
#> 35        0.992269255191677  3.31385903738819
#> 36    0.0000159215720431094  2.00898582775442
#> 37   0.00000401221546419972  1.58127235599564
#> 38   0.00017443080013468393  1.68303687065352
#> 39  0.000005297055129194206  1.64263891361642
#> 40       0.5472741906553065  5.11486368918823
#> 41       0.9946412752895979  5.82338557086884
#> 42    0.0051032998085885906    5.407130834116
#> 43       0.3918178053441286  5.26207864762865
#> 44       0.9957266821334012   3.7954802665032
#> 45    0.0002634114615062133  1.15369069085588
#> 46  0.000007796744580956805  1.01929201570272
#> 47     0.019490398477755022  6.96743176503065
#> 48    0.0012162595369955564  3.47615406329367
#> 49    0.0032590209677643145  4.69112492948641
#> 50     0.004031849640397045  1.24046653086629
#> 51     0.003256819107921749  2.05000353316949
#> 52      0.10317998221501223  3.75944143362848
#> 53   0.00008576772798907228  1.68453201604003
#> 54       0.9407697074417514  2.81056025015645
#> 55     0.001028011004730045  2.98206857835496
#> 56  0.000015835535093988717  3.61417267641785
#> 57    0.0012713340353657232  2.27288533182659
#> 58       0.7085646902378377  2.81815853511343
#> 59   0.00015972691607413214  2.89891504811931
#> 60  0.000036258261760346855  1.49702612211591
#> 61  0.000004165417503809721 0.924128809470888
#> 62        0.844213116573793  7.78761671145787
#> 63   0.00020727821340027486  2.40251314355121
#> 64       0.8265336617017218  4.96335409829418
#> 65       0.9118309902912132  2.94900939227136
#> 66       0.7803958762265931   6.8080457169545
#> 67     0.002248910317168673  7.32206810726285
#> 68    0.0011154194995069977  3.48738448121798
#> 69       0.4288198994607003  7.26181006744773
#> 70    0.0005047783187366317  1.45333628857841
#> 71  0.000012685559573455407  1.71832216329414
#> 72    0.0001742161652431573  1.48686796628544
#> 73      0.07216956854098999  2.82591750321837
#> 74       0.9939689644437355  2.66469555144682
#> 75  0.000004404477775943327  1.63423903361417
#> 76    0.0003796669053852675  2.72260296925641
#> 77    0.0012626697935577136  2.42312831919202
#> 78   0.00017042791751720188  2.33862515413593
#> 79    0.0001606326780763112  2.71526796951292
#> 80   0.00007443135335916118  1.40854492955243
#> 81    0.0008418813561436034  3.13714239259538
#> 82    0.0004597288685521672   1.8898332819208
#> 83  0.000007992926103650433  1.74121825402926
#> 84        0.747621775556108  4.53523527655337
#> 85      0.02035858702878457   1.9151864470078
#> 86       0.5198478586172812  7.48712427141253
#> 87      0.04404470034418717  2.32118897073739
#> 88       0.6875322954307215  2.82950720078561

cfb_game_info(2018, week = 7, conference = 'Ind')

#>          id season week season_type               start_date start_time_tbd
#> 1 401013452   2018    7     regular 2018-10-13T18:00:00.000Z             NA
#> 2 401013148   2018    7     regular 2018-10-13T18:30:00.000Z             NA
#> 3 401013370   2018    7     regular 2018-10-13T19:30:00.000Z             NA
#> 4 401013442   2018    7     regular 2018-10-13T21:00:00.000Z             NA
#> 5 401016408   2018    7     regular 2018-10-14T02:15:00.000Z             NA
#>   neutral_site conference_game attendance venue_id                  venue
#> 1        FALSE           FALSE          0     3992       Williams Stadium
#> 2        FALSE           FALSE          0     3855     Notre Dame Stadium
#> 3         TRUE           FALSE          0     4738         Levi's Stadium
#> 4        FALSE           FALSE          0     3666            Cajun Field
#> 5        FALSE           FALSE          0     3801 LaVell Edwards Stadium
#>   home_id      home_team  home_conference home_points    home_post_win_prob
#> 1    2335        Liberty FBS Independents          22    0.5483899675354477
#> 2      87     Notre Dame FBS Independents          19    0.9661277906146866
#> 3      23 San José State    Mountain West           3 0.0036361807521906077
#> 4     309      Louisiana         Sun Belt          66    0.9983063510069449
#> 5     252            BYU FBS Independents          49    0.9978157683814658
#>   away_id        away_team  away_conference away_points    away_post_win_prob
#> 1    2653             Troy         Sun Belt          16   0.45161003246455234
#> 2     221       Pittsburgh              ACC          14  0.033872209385313434
#> 3     349             Army FBS Independents          52    0.9963638192478094
#> 4     166 New Mexico State FBS Independents          38 0.0016936489930551302
#> 5      62          Hawai'i    Mountain West          23 0.0021842316185342048
#>   excitement_index
#> 1 7.15938486562246
#> 2 6.39074079758295
#> 3 2.66054549025729
#> 4 1.93589609341136
#> 5 1.66763582574363

#7 OTs LSU @ TAMU
cfb_game_info(2018, week = 13, team = "Texas A&M", quarter_scores=TRUE)

#>          id season week season_type               start_date start_time_tbd
#> 1 401012356   2018   13     regular 2018-11-25T00:30:00.000Z             NA
#>   neutral_site conference_game attendance venue_id      venue home_id home_team
#> 1        FALSE            TRUE          0     3795 Kyle Field     245 Texas A&M
#>   home_conference home_points home_scores_Q1 home_scores_Q2 home_scores_Q3
#> 1             SEC          74              7             10              7
#>   home_scores_Q4 home_scores_Q5 home_scores_Q6 home_scores_Q7 home_scores_Q8
#> 1              7              3              7              8              3
#>   home_scores_Q9 home_scores_Q10 home_scores_Q11 home_post_win_prob away_id
#> 1              6               8               8 0.6786258156883023      99
#>   away_team away_conference away_points away_scores_Q1 away_scores_Q2
#> 1       LSU             SEC          72              7              3
#>   away_scores_Q3 away_scores_Q4 away_scores_Q5 away_scores_Q6 away_scores_Q7
#> 1              7             14              3              7              8
#>   away_scores_Q8 away_scores_Q9 away_scores_Q10 away_scores_Q11
#> 1              3              6               8               6
#>    away_post_win_prob excitement_index
#> 1 0.32137418431169773 5.83590584818378