Get Team Statistics by Game

Get Team Statistics by Game

cfb_game_team_stats(
  year,
  week = NULL,
  season_type = "regular",
  team = NULL,
  conference = NULL,
  game_id = NULL,
  rows_per_team = 1
)

Arguments

year
(Integer required): Year, 4 digit format (YYYY)

week
(Integer optional): Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier

season_type
(String default: regular): Select Season Type - regular, postseason, or both

team
(String optional): D-I Team

conference
(String optional): Conference abbreviation - Select a valid FBS conference Conference abbreviations P5: ACC, B12, B1G, SEC, PAC Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

game_id
(Integer optional): Game ID filter for querying a single game Can be found using the `cfb_game_info()` function

rows_per_team
(Integer default 1): Both Teams for each game on one or two row(s), Options: 1 or 2

Examples

cfb_game_team_stats(2018, week = 1, conference = 'B12')

#> # A tibble: 9 x 78
#>       id school conference homeAway points rushingTDs puntReturnYards
#>    <int> <chr>  <chr>      <chr>     <int> <chr>      <chr>          
#> 1 4.01e8 West … Big 12     away         40 0          -1             
#> 2 4.01e8 Texas… Big 12     home         27 2          NA             
#> 3 4.01e8 Oklah… Big 12     home         58 3          -1             
#> 4 4.01e8 Kansa… Big 12     home         27 0          107            
#> 5 4.01e8 Baylor Big 12     home         55 5          38             
#> 6 4.01e8 Oklah… Big 12     home         63 5          46             
#> 7 4.01e8 Texas  Big 12     away         29 2          56             
#> 8 4.01e8 Kansas Big 12     home         23 0          26             
#> 9 4.01e8 TCU    Big 12     home         55 2          113            
#> # … with 71 more variables: puntReturnTDs <chr>, puntReturns <chr>,
#> #   passingTDs <chr>, kickReturnYards <chr>, kickReturnTDs <chr>,
#> #   kickReturns <chr>, kickingPoints <chr>, fumblesRecovered <chr>,
#> #   totalFumbles <chr>, tacklesForLoss <chr>, defensiveTDs <chr>,
#> #   tackles <chr>, sacks <chr>, qbHurries <chr>, passesDeflected <chr>,
#> #   firstDowns <chr>, thirdDownEff <chr>, fourthDownEff <chr>,
#> #   totalYards <chr>, netPassingYards <chr>, completionAttempts <chr>,
#> #   yardsPerPass <chr>, rushingYards <chr>, rushingAttempts <chr>,
#> #   yardsPerRushAttempt <chr>, totalPenaltiesYards <chr>, turnovers <chr>,
#> #   fumblesLost <chr>, interceptions <chr>, possessionTime <chr>,
#> #   interceptionYards <chr>, interceptionTDs <chr>, passesIntercepted <chr>,
#> #   conference_allowed <chr>, homeAway_allowed <chr>, points_allowed <int>,
#> #   rushingTDs_allowed <chr>, puntReturnYards_allowed <chr>,
#> #   puntReturnTDs_allowed <chr>, puntReturns_allowed <chr>,
#> #   passingTDs_allowed <chr>, kickReturnYards_allowed <chr>,
#> #   kickReturnTDs_allowed <chr>, kickReturns_allowed <chr>,
#> #   kickingPoints_allowed <chr>, fumblesRecovered_allowed <chr>,
#> #   totalFumbles_allowed <chr>, tacklesForLoss_allowed <chr>,
#> #   defensiveTDs_allowed <chr>, tackles_allowed <chr>, sacks_allowed <chr>,
#> #   qbHurries_allowed <chr>, passesDeflected_allowed <chr>,
#> #   firstDowns_allowed <chr>, thirdDownEff_allowed <chr>,
#> #   fourthDownEff_allowed <chr>, totalYards_allowed <chr>,
#> #   netPassingYards_allowed <chr>, completionAttempts_allowed <chr>,
#> #   yardsPerPass_allowed <chr>, rushingYards_allowed <chr>,
#> #   rushingAttempts_allowed <chr>, yardsPerRushAttempt_allowed <chr>,
#> #   totalPenaltiesYards_allowed <chr>, turnovers_allowed <chr>,
#> #   fumblesLost_allowed <chr>, interceptions_allowed <chr>,
#> #   possessionTime_allowed <chr>, interceptionYards_allowed <chr>,
#> #   interceptionTDs_allowed <chr>, passesIntercepted_allowed <chr>

cfb_game_team_stats(2019, team = 'LSU')

#> # A tibble: 13 x 78
#>        id school conference homeAway points rushingTDs puntReturnYards
#>     <int> <chr>  <chr>      <chr>     <int> <chr>      <chr>          
#>  1 4.01e8 LSU    SEC        away         45 1          11             
#>  2 4.01e8 LSU    SEC        away         66 2          25             
#>  3 4.01e8 LSU    SEC        home         23 2          7              
#>  4 4.01e8 LSU    SEC        away         46 3          19             
#>  5 4.01e8 LSU    SEC        away         36 0          2              
#>  6 4.01e8 LSU    SEC        away         58 2          NA             
#>  7 4.01e8 LSU    SEC        home         56 5          1              
#>  8 4.01e8 LSU    SEC        home         50 2          NA             
#>  9 4.01e8 LSU    SEC        home         55 2          52             
#> 10 4.01e8 LSU    SEC        home         65 6          108            
#> 11 4.01e8 LSU    SEC        home         42 3          NA             
#> 12 4.01e8 LSU    SEC        home         42 1          10             
#> 13 4.01e8 LSU    SEC        home         37 0          NA             
#> # … with 71 more variables: puntReturnTDs <chr>, puntReturns <chr>,
#> #   passingTDs <chr>, kickingPoints <chr>, interceptionYards <chr>,
#> #   interceptionTDs <chr>, passesIntercepted <chr>, fumblesRecovered <chr>,
#> #   totalFumbles <chr>, tacklesForLoss <chr>, defensiveTDs <chr>,
#> #   tackles <chr>, sacks <chr>, qbHurries <chr>, passesDeflected <chr>,
#> #   possessionTime <chr>, interceptions <chr>, fumblesLost <chr>,
#> #   turnovers <chr>, totalPenaltiesYards <chr>, yardsPerRushAttempt <chr>,
#> #   rushingAttempts <chr>, rushingYards <chr>, yardsPerPass <chr>,
#> #   completionAttempts <chr>, netPassingYards <chr>, totalYards <chr>,
#> #   fourthDownEff <chr>, thirdDownEff <chr>, firstDowns <chr>,
#> #   kickReturnYards <chr>, kickReturnTDs <chr>, kickReturns <chr>,
#> #   conference_allowed <chr>, homeAway_allowed <chr>, points_allowed <int>,
#> #   rushingTDs_allowed <chr>, puntReturnYards_allowed <chr>,
#> #   puntReturnTDs_allowed <chr>, puntReturns_allowed <chr>,
#> #   passingTDs_allowed <chr>, kickingPoints_allowed <chr>,
#> #   interceptionYards_allowed <chr>, interceptionTDs_allowed <chr>,
#> #   passesIntercepted_allowed <chr>, fumblesRecovered_allowed <chr>,
#> #   totalFumbles_allowed <chr>, tacklesForLoss_allowed <chr>,
#> #   defensiveTDs_allowed <chr>, tackles_allowed <chr>, sacks_allowed <chr>,
#> #   qbHurries_allowed <chr>, passesDeflected_allowed <chr>,
#> #   possessionTime_allowed <chr>, interceptions_allowed <chr>,
#> #   fumblesLost_allowed <chr>, turnovers_allowed <chr>,
#> #   totalPenaltiesYards_allowed <chr>, yardsPerRushAttempt_allowed <chr>,
#> #   rushingAttempts_allowed <chr>, rushingYards_allowed <chr>,
#> #   yardsPerPass_allowed <chr>, completionAttempts_allowed <chr>,
#> #   netPassingYards_allowed <chr>, totalYards_allowed <chr>,
#> #   fourthDownEff_allowed <chr>, thirdDownEff_allowed <chr>,
#> #   firstDowns_allowed <chr>, kickReturnYards_allowed <chr>,
#> #   kickReturnTDs_allowed <chr>, kickReturns_allowed <chr>

cfb_game_team_stats(2013, team = "Florida State")

#> # A tibble: 13 x 64
#>        id school conference homeAway points fumblesRecovered rushingTDs
#>     <int> <chr>  <chr>      <chr>     <int> <chr>            <chr>     
#>  1 3.32e8 Flori… ACC        away         41 0                1         
#>  2 3.33e8 Flori… ACC        home         62 0                6         
#>  3 3.33e8 Flori… ACC        home         54 0                4         
#>  4 3.33e8 Flori… ACC        away         48 0                1         
#>  5 3.33e8 Flori… ACC        home         63 1                4         
#>  6 3.33e8 Flori… ACC        away         51 2                2         
#>  7 3.33e8 Flori… ACC        home         49 1                4         
#>  8 3.33e8 Flori… ACC        home         41 0                4         
#>  9 3.33e8 Flori… ACC        away         59 1                3         
#> 10 3.33e8 Flori… ACC        home         59 1                4         
#> 11 3.33e8 Flori… ACC        home         80 0                4         
#> 12 3.33e8 Flori… ACC        away         37 2                1         
#> 13 3.33e8 Flori… ACC        home         45 1                3         
#> # … with 57 more variables: passingTDs <chr>, kickReturnYards <chr>,
#> #   kickReturnTDs <chr>, kickReturns <chr>, kickingPoints <chr>,
#> #   firstDowns <chr>, thirdDownEff <chr>, fourthDownEff <chr>,
#> #   totalYards <chr>, netPassingYards <chr>, completionAttempts <chr>,
#> #   yardsPerPass <chr>, rushingYards <chr>, rushingAttempts <chr>,
#> #   yardsPerRushAttempt <chr>, totalPenaltiesYards <chr>, turnovers <chr>,
#> #   fumblesLost <chr>, interceptions <chr>, possessionTime <chr>,
#> #   puntReturnYards <chr>, puntReturnTDs <chr>, puntReturns <chr>,
#> #   interceptionYards <chr>, interceptionTDs <chr>, passesIntercepted <chr>,
#> #   conference_allowed <chr>, homeAway_allowed <chr>, points_allowed <int>,
#> #   fumblesRecovered_allowed <chr>, rushingTDs_allowed <chr>,
#> #   passingTDs_allowed <chr>, kickReturnYards_allowed <chr>,
#> #   kickReturnTDs_allowed <chr>, kickReturns_allowed <chr>,
#> #   kickingPoints_allowed <chr>, firstDowns_allowed <chr>,
#> #   thirdDownEff_allowed <chr>, fourthDownEff_allowed <chr>,
#> #   totalYards_allowed <chr>, netPassingYards_allowed <chr>,
#> #   completionAttempts_allowed <chr>, yardsPerPass_allowed <chr>,
#> #   rushingYards_allowed <chr>, rushingAttempts_allowed <chr>,
#> #   yardsPerRushAttempt_allowed <chr>, totalPenaltiesYards_allowed <chr>,
#> #   turnovers_allowed <chr>, fumblesLost_allowed <chr>,
#> #   interceptions_allowed <chr>, possessionTime_allowed <chr>,
#> #   puntReturnYards_allowed <chr>, puntReturnTDs_allowed <chr>,
#> #   puntReturns_allowed <chr>, interceptionYards_allowed <chr>,
#> #   interceptionTDs_allowed <chr>, passesIntercepted_allowed <chr>