Get player season averages for Predicted Points Added (PPA)

Get player season averages for Predicted Points Added (PPA)

cfb_metrics_ppa_players_season(
  year = NULL,
  team = NULL,
  conference = NULL,
  position = NULL,
  athlete_id = NULL,
  threshold = NULL,
  excl_garbage_time = FALSE
)

Arguments

year
(Integer required): Year, 4 digit format (YYYY)

team
(String optional): D-I Team

conference
(String optional): Conference abbreviation - S&P+ information by conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

position
(string optional): Position abbreviation of the player you are searching for.
Position Group - options include:
* Offense: QB, RB, FB, TE, OL, G, OT, C, WR
* Defense: DB, CB, S, LB, DE, DT, NT, DL
* Special Teams: K, P, LS, PK

athlete_id
(Integer optional): Athlete ID filter for querying a single athlete
Can be found using the cfbscrapR::cfb_player_info() function.

threshold
(Integer optional): Minimum threshold of plays.

excl_garbage_time
(Logical default FALSE): Select whether to exclude Garbage Time (TRUE or FALSE)

Source

https://api.collegefootballdata.com/ppa/players/season

Value

A data frame with 23 variables:

season

integer.

id

character.

name

character.

position

character.

team

character.

conference

character.

countable_plays

integer.

avg_PPA_all

double.

avg_PPA_pass

double.

avg_PPA_rush

double.

avg_PPA_first_down

double.

avg_PPA_second_down

double.

avg_PPA_third_down

double.

avg_PPA_standard_downs

double.

avg_PPA_passing_downs

double.

total_PPA_all

double.

total_PPA_pass

double.

total_PPA_rush

double.

total_PPA_first_down

double.

total_PPA_second_down

double.

total_PPA_third_down

double.

total_PPA_standard_downs

double.

total_PPA_passing_downs

double.

Examples

cfb_metrics_ppa_players_season(year = 2019, team = 'TCU')

#> 2020-11-03 01:29:25: Scraping CFBData metrics PPA season-level players data...

#>    season      id              name position team conference countable_plays
#> 1    2019 4427105        Max Duggan       QB  TCU     Big 12             456
#> 2    2019 4038533   Darius Anderson       RB  TCU     Big 12             171
#> 3    2019 4038539     Sewo Olonilua       RB  TCU     Big 12             155
#> 4    2019 4241802      Jalen Reagor       WR  TCU     Big 12              91
#> 5    2019 3886634       Alex Delton       QB  TCU     Big 12              63
#> 6    2019 4362430       Taye Barber       WR  TCU     Big 12              48
#> 7    2019 4362442         Pro Wells       TE  TCU     Big 12              28
#> 8    2019 4426635     Darwin Barlow       RB  TCU     Big 12              22
#> 9    2019 4362477      Derius Davis       WR  TCU     Big 12              21
#> 10   2019 4362478   Emari Demercado       RB  TCU     Big 12              17
#> 11   2019 4038556   Artayvious Lynn       TE  TCU     Big 12              15
#> 12   2019 4362489 John Stephens Jr.       WR  TCU     Big 12              15
#> 13   2019 4038536      Dylan Thomas       WR  TCU     Big 12              10
#> 14   2019 4362483   Tevailance Hunt       WR  TCU     Big 12               9
#> 15   2019 3676818  TreVontae Hights       WR  TCU     Big 12               9
#> 16   2019 4427217  Daimarqua Foster       RB  TCU     Big 12               8
#> 17   2019 4426965   Blair Conwright       WR  TCU     Big 12               4
#> 18   2019 4259331       Carter Ware       TE  TCU     Big 12               2
#> 19   2019 4241795   Al'Dontre Davis       WR  TCU     Big 12               1
#>    avg_PPA_all avg_PPA_pass avg_PPA_rush avg_PPA_first_down avg_PPA_second_down
#> 1        0.175        0.079        0.513             -0.009               0.031
#> 2        0.065       -0.179        0.107              0.038               0.001
#> 3        0.126       -0.193        0.205             -0.092               0.070
#> 4        0.296        0.345        0.030              0.397              -0.067
#> 5       -0.020       -0.124        0.475              0.253              -0.549
#> 6        0.586        0.543        0.958              0.504               0.104
#> 7        0.642        0.642           NA              0.058               0.343
#> 8        0.178       -0.881        0.228             -0.170               0.415
#> 9        0.243        0.296       -0.075              0.144              -0.179
#> 10      -0.187           NA       -0.187              0.120              -0.407
#> 11       0.492        0.492           NA              0.273               0.652
#> 12       0.726        0.726           NA              0.658               0.139
#> 13       1.178        1.178           NA             -0.384               0.754
#> 14       0.385        0.385           NA                 NA               0.224
#> 15       0.462        0.462           NA              0.159                  NA
#> 16       0.365           NA        0.365              0.148              -0.381
#> 17       1.068        1.068           NA             -0.020               2.039
#> 18       0.808        0.808           NA             -0.362                  NA
#> 19       3.183        3.183           NA                 NA                  NA
#>    avg_PPA_third_down avg_PPA_standard_downs avg_PPA_passing_downs
#> 1               0.573                  0.040                 0.361
#> 2               1.077                  0.014                 0.294
#> 3               0.823                  0.118                 0.153
#> 4               0.663                  0.287                 0.312
#> 5               0.076                  0.010                -0.057
#> 6               0.922                  0.572                 0.607
#> 7               1.202                  0.092                 1.118
#> 8               0.235                  0.160                 0.209
#> 9               0.580                  0.242                 0.245
#> 10             -0.391                 -0.190                -0.164
#> 11              0.775                  0.392                 0.691
#> 12              1.088                  0.734                 0.717
#> 13              2.952                  0.457                 2.260
#> 14             -0.466                 -0.066                 0.747
#> 15              1.069                  0.159                 1.069
#> 16              2.946                 -0.047                 1.602
#> 17              2.273                 -0.020                 2.156
#> 18              1.979                  0.808                    NA
#> 19              3.183                     NA                 3.183
#>    total_PPA_all total_PPA_pass total_PPA_rush total_PPA_first_down
#> 1         79.725         27.918         51.807               -1.560
#> 2         11.090         -4.486         15.577                3.684
#> 3         19.468         -5.980         25.448               -6.467
#> 4         26.973         26.549          0.424               16.691
#> 5         -1.238         -6.459          5.221                5.561
#> 6         28.145         23.355          4.789                9.584
#> 7         17.965         17.965             NA                0.349
#> 8          3.906         -0.881          4.787               -1.363
#> 9          5.105          5.328         -0.224                1.579
#> 10        -3.185             NA         -3.185                0.839
#> 11         7.374          7.374             NA                1.911
#> 12        10.893         10.893             NA                3.947
#> 13        11.781         11.781             NA               -1.536
#> 14         3.469          3.469             NA                   NA
#> 15         4.162          4.162             NA                0.955
#> 16         2.923             NA          2.923                0.739
#> 17         4.272          4.272             NA               -0.039
#> 18         1.617          1.617             NA               -0.362
#> 19         3.183          3.183             NA                   NA
#>    total_PPA_second_down total_PPA_third_down total_PPA_standard_downs
#> 1                  4.649               72.769                   10.731
#> 2                  0.071                9.689                    1.973
#> 3                  3.707               24.681                   14.108
#> 4                 -1.875               13.269                   16.366
#> 5                -10.435                1.600                    0.355
#> 6                  1.044               17.517                   16.004
#> 7                  3.089               14.418                    1.197
#> 8                  4.565                0.704                    2.237
#> 9                 -0.536                4.062                    3.141
#> 10                -2.851               -1.173                   -2.857
#> 11                 3.913                1.549                    3.920
#> 12                 0.416                6.530                    5.875
#> 13                 1.509               11.808                    2.741
#> 14                 0.896               -1.864                   -0.265
#> 15                    NA                3.208                    0.955
#> 16                -0.762                2.946                   -0.281
#> 17                 2.039                2.273                   -0.039
#> 18                    NA                1.979                    1.617
#> 19                    NA                3.183                       NA
#>    total_PPA_passing_downs
#> 1                   68.994
#> 2                    9.117
#> 3                    5.360
#> 4                   10.607
#> 5                   -1.592
#> 6                   12.141
#> 7                   16.769
#> 8                    1.669
#> 9                    1.963
#> 10                  -0.328
#> 11                   3.453
#> 12                   5.018
#> 13                   9.040
#> 14                   3.734
#> 15                   3.208
#> 16                   3.204
#> 17                   4.311
#> 18                      NA
#> 19                   3.183