Gets player info associated by play

Information describes the players involved in the play this includes passer, receiver, defensive players who create sacks or picks, etc

cfb_play_stats_player(gameId)

Arguments

gameId
Game ID

Details

gameId can be extracted from cfb_game_info

Examples

cfb_play_stats_player(401110723)

#> Warning: `funs()` is deprecated as of dplyr 0.8.0.
#> Please use a list of either functions or lambdas: 
#> 
#>   # Simple named list: 
#>   list(mean = mean, median = median)
#> 
#>   # Auto named with `tibble::lst()`: 
#>   tibble::lst(mean, median)
#> 
#>   # Using lambdas
#>   list(~ mean(., trim = .2), ~ median(., na.rm = TRUE))
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> # A tibble: 120 x 24
#>    playId gameId season  week opponent teamScore opponentScore driveId period
#>    <chr>   <int>  <int> <int> <chr>        <int>         <int> <chr>    <int>
#>  1 40111~ 4.01e8   2019     1 Florida          0             0 401110~      1
#>  2 40111~ 4.01e8   2019     1 Florida          0             0 401110~      1
#>  3 40111~ 4.01e8   2019     1 Florida          0             0 401110~      1
#>  4 40111~ 4.01e8   2019     1 Florida          0             0 401110~      1
#>  5 40111~ 4.01e8   2019     1 Florida          0             0 401110~      1
#>  6 40111~ 4.01e8   2019     1 Florida          0             0 401110~      1
#>  7 40111~ 4.01e8   2019     1 Miami            0             0 401110~      1
#>  8 40111~ 4.01e8   2019     1 Florida          0             0 401110~      1
#>  9 40111~ 4.01e8   2019     1 Miami            0             3 401110~      1
#> 10 40111~ 4.01e8   2019     1 Miami            0             3 401110~      1
#> # ... with 110 more rows, and 15 more variables: secondsRemaining <int>,
#> #   yardsToGoal <int>, down <int>, distance <int>, Sack <chr>, `Sack
#> #   Taken` <chr>, Fumble <chr>, Reception <chr>, Passer <chr>, Target <chr>,
#> #   Interception <chr>, `Interception Thrown` <chr>, Rush <chr>,
#> #   Touchdown <chr>, `Fumble Recovered` <chr>