College Football Mapping for Play Stats Types

This function identifies all play stats types identified in the play by play data. This can be used to filter out play stats types when calling functions before hand.

cfb_play_stats_types()

Format

A data frame with 22 rows and 2 variables:

id

Referencing play id

name

Type of play stats

...
Examples

cfb_play_stats_types()

#>    id                name
#> 1   1        Incompletion
#> 2   2              Target
#> 3   3        Pass Breakup
#> 4   4          Completion
#> 5   5           Reception
#> 6   6              Tackle
#> 7   7                Rush
#> 8   8              Fumble
#> 9   9       Fumble Forced
#> 10 10    Fumble Recovered
#> 11 11          Sack Taken
#> 12 12                Sack
#> 13 13             Kickoff
#> 14 14         Onside Kick
#> 15 15      Kickoff Return
#> 16 16                Punt
#> 17 17          Punt Block
#> 18 18  FG Attempt Blocked
#> 19 19    Field Goal Block
#> 20 20 Interception Thrown
#> 21 21        Interception
#> 22 22           Touchdown