Player Information Search

A player usage function with Year as a required input.

cfb_player_usage(
  year = 2019,
  team = NULL,
  conference = NULL,
  position = NULL,
  athlete_id = NULL,
  excl_garbage_time = FALSE
)

Arguments

year
(Integer required, default 2019): Year, 4 digit format (YYYY).

team
(String optional): Team - Select a valid team, D1 football

conference
(String optional): Conference abbreviation - Select a valid FBS conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

position
(string optional): Position of the player you are searching for.
Position Group - options include:
* Offense: QB, RB, FB, TE, OL, G, OT, C, WR
* Defense: DB, CB, S, LB, DE, DT, NT, DL
* Special Teams: K, P, LS, PK

athlete_id
(Integer optional): Athlete ID filter for querying a single athlete Can be found using the cfbscrapR::cfb_player_info() function.

excl_garbage_time
(Logical default FALSE): Select whether to exclude Garbage Time (TRUE/FALSE)

Source

https://api.collegefootballdata.com/player/usage

Value

A data frame with 14 variables:

season

integer.

athlete_id

character.

name

character.

position

character.

team

character.

conference

character.

usg_overall

double.

usg_pass

double.

usg_rush

double.

usg_1st_down

double.

usg_2nd_down

double.

usg_3rd_down

double.

usg_standard_downs

double.

usg_passing_downs

double.

Examples

cfb_player_usage(year = 2019, position = 'WR', team = 'Florida State')

#> 2020-11-03 01:34:33: Scraping player usage data...

#>   season athlete_id               name position          team conference
#> 1   2019    4240028    Tamorrion Terry       WR Florida State        ACC
#> 2   2019    4240034      D.J. Matthews       WR Florida State        ACC
#> 3   2019    4363046 Tre'Shaun Harrison       WR Florida State        ACC
#> 4   2019    4240033     Ontaria Wilson       WR Florida State        ACC
#> 5   2019    4363032    Keyshawn Helton       WR Florida State        ACC
#> 6   2019    4035628        Keith Gavin       WR Florida State        ACC
#> 7   2019    4363030    Warren Thompson       WR Florida State        ACC
#> 8   2019    4363044       Jordan Young       WR Florida State        ACC
#> 9   2019    4360473       Adarius Dent       WR Florida State        ACC
#>   usg_overall usg_pass usg_rush usg_1st_down usg_2nd_down usg_3rd_down
#> 1      0.1184   0.2181   0.0102       0.1402        0.099        0.112
#> 2      0.0604   0.1090   0.0077       0.0485        0.076        0.071
#> 3      0.0459   0.0812   0.0077       0.0404        0.046        0.065
#> 4      0.0374   0.0673   0.0051       0.0377        0.034        0.047
#> 5      0.0278   0.0487   0.0051       0.0296        0.027        0.029
#> 6      0.0254   0.0487   0.0000       0.0216        0.030        0.029
#> 7      0.0085   0.0162   0.0000       0.0081        0.015        0.000
#> 8      0.0072   0.0116   0.0026       0.0054        0.008        0.012
#> 9      0.0012   0.0023   0.0000       0.0027        0.000        0.000
#>   usg_standard_downs usg_passing_downs
#> 1              0.113             0.128
#> 2              0.046             0.089
#> 3              0.048             0.043
#> 4              0.038             0.036
#> 5              0.024             0.036
#> 6              0.020             0.036
#> 7              0.007             0.011
#> 8              0.007             0.007
#> 9              0.002             0.000