CFB Recruiting Information - Position Groups

If only start_year is provided, function will get CFB recruiting information based on position groups during that year for all FBS teams.

cfb_recruiting_position(
  start_year = NULL,
  end_year = NULL,
  team = NULL,
  conference = NULL
)

Arguments

start_year
(Integer optional): Start Year, 4 digit format (YYYY). Note: 2000 is the minimum value

end_year
(Integer optional): End Year, 4 digit format (YYYY). Note: 2020 is the maximum value currently

team
(String optional): Team - Select a valid team, D-I football

conference
(String optional): Conference name - select a valid FBS conference Conference names P5: ACC, Big 12, Big Ten, SEC, Pac-12 Conference names G5 and FBS Independents: Conference USA, Mid-American, Mountain West, FBS Independents, American Athletic

Details

If you would like CFB recruiting information for players, please see the cfbscrapR::cfb_recruiting_player() function

If you would like CFB recruiting information for teams, please see the cfbscrapR::cfb_recruiting_team() function

Examples

cfb_recruiting_position(2018, team="Texas")

#> 2020-09-05 13:30:39: Scraping position group recruiting data...

#>    team conference position_group avg_rating total_rating commits avg_stars
#> 1 Texas     Big 12 Defensive Back  0.9480923      12.3252      13  4.076923
#> 2 Texas     Big 12 Defensive Line  0.9050300       9.0503      10  3.500000
#> 3 Texas     Big 12     Linebacker  0.9040250       7.2322       8  3.500000
#> 4 Texas     Big 12 Offensive Line  0.9099364      10.0093      11  3.909091
#> 5 Texas     Big 12    Quarterback  0.9323400       4.6617       5  4.000000
#> 6 Texas     Big 12       Receiver  0.9238545      10.1624      11  3.727273
#> 7 Texas     Big 12   Running Back  0.9463667       2.8391       3  4.333333
#> 8 Texas     Big 12  Special Teams  0.8869875       7.0959       8  3.625000
#> 9 Texas     Big 12  All Positions  0.9184942      63.3761      69  3.797101

cfb_recruiting_position(2016, 2020, team="Virginia")

#> 2020-09-05 13:30:40: Scraping position group recruiting data...

#>       team conference position_group avg_rating total_rating commits avg_stars
#> 1 Virginia        ACC Defensive Back  0.8332929      11.6661      14  2.857143
#> 2 Virginia        ACC Defensive Line  0.8512125      13.6194      16  3.125000
#> 3 Virginia        ACC     Linebacker  0.8461056      15.2299      18  3.000000
#> 4 Virginia        ACC Offensive Line  0.8463294      14.3876      17  3.058824
#> 5 Virginia        ACC    Quarterback  0.8487200       4.2436       5  3.000000
#> 6 Virginia        ACC       Receiver  0.8392842      15.9464      19  2.947368
#> 7 Virginia        ACC   Running Back  0.8453714       5.9176       7  3.000000
#> 8 Virginia        ACC  Special Teams  0.8361111       7.5250       9  2.888889
#> 9 Virginia        ACC  All Positions  0.8431962      88.5356     105  2.990476

cfb_recruiting_position(2015, 2020, conference = "SEC")

#> 2020-09-05 13:30:40: Scraping position group recruiting data...

#>                  team conference position_group avg_rating total_rating commits
#> 1             Alabama        SEC Defensive Back  0.9456044      21.7489      23
#> 2             Alabama        SEC Defensive Line  0.9521448      27.6122      29
#> 3             Alabama        SEC     Linebacker  0.9449375      15.1190      16
#> 4             Alabama        SEC Offensive Line  0.9320348      21.4368      23
#> 5             Alabama        SEC    Quarterback  0.9061222       8.1551       9
#> 6             Alabama        SEC       Receiver  0.9244810      19.4141      21
#> 7             Alabama        SEC   Running Back  0.9296071      13.0145      14
#> 8             Alabama        SEC  Special Teams  0.8858300       8.8583      10
#> 9            Arkansas        SEC Defensive Back  0.8603522      19.7881      23
#> 10           Arkansas        SEC Defensive Line  0.8792360      21.9809      25
#> 11           Arkansas        SEC     Linebacker  0.8497611      15.2957      18
#> 12           Arkansas        SEC Offensive Line  0.8514273      18.7314      22
#> 13           Arkansas        SEC    Quarterback  0.8728000       6.9824       8
#> 14           Arkansas        SEC       Receiver  0.8796667      15.8340      18
#> 15           Arkansas        SEC   Running Back  0.8772857       6.1410       7
#> 16           Arkansas        SEC  Special Teams  0.8520625       6.8165       8
#> 17             Auburn        SEC Defensive Back  0.8866111      15.9590      18
#> 18             Auburn        SEC Defensive Line  0.9238227      20.3241      22
#> 19             Auburn        SEC     Linebacker  0.8956500      17.9130      20
#> 20             Auburn        SEC Offensive Line  0.8970000      15.2490      17
#> 21             Auburn        SEC    Quarterback  0.8895400       4.4477       5
#> 22             Auburn        SEC       Receiver  0.9000333      21.6008      24
#> 23             Auburn        SEC   Running Back  0.9126500       9.1265      10
#> 24             Auburn        SEC  Special Teams  0.8959950      17.9199      20
#> 25            Florida        SEC Defensive Back  0.9001296      24.3035      27
#> 26            Florida        SEC Defensive Line  0.9037500      19.8825      22
#> 27            Florida        SEC     Linebacker  0.8848714      12.3882      14
#> 28            Florida        SEC Offensive Line  0.8845846      22.9992      26
#> 29            Florida        SEC    Quarterback  0.8857143       6.2000       7
#> 30            Florida        SEC       Receiver  0.9088958      21.8135      24
#> 31            Florida        SEC   Running Back  0.8956125       7.1649       8
#> 32            Florida        SEC  Special Teams  0.8576000       4.2880       5
#> 33            Georgia        SEC Defensive Back  0.9096519      24.5606      27
#> 34            Georgia        SEC Defensive Line  0.9257000      26.8453      29
#> 35            Georgia        SEC     Linebacker  0.9416125      15.0658      16
#> 36            Georgia        SEC Offensive Line  0.9210885      23.9483      26
#> 37            Georgia        SEC    Quarterback  0.9089714       6.3628       7
#> 38            Georgia        SEC       Receiver  0.9174240      22.9356      25
#> 39            Georgia        SEC   Running Back  0.9415500       7.5324       8
#> 40            Georgia        SEC  Special Teams  0.8711818       9.5830      11
#> 41           Kentucky        SEC Defensive Back  0.8643550      17.2871      20
#> 42           Kentucky        SEC Defensive Line  0.8644333      20.7464      24
#> 43           Kentucky        SEC     Linebacker  0.8672545       9.5398      11
#> 44           Kentucky        SEC Offensive Line  0.8630727      18.9876      22
#> 45           Kentucky        SEC    Quarterback  0.8507000       5.9549       7
#> 46           Kentucky        SEC       Receiver  0.8564571      17.9856      21
#> 47           Kentucky        SEC   Running Back  0.8686875       6.9495       8
#> 48           Kentucky        SEC  Special Teams  0.8501357      11.9019      14
#> 49                LSU        SEC Defensive Back  0.9350048      19.6351      21
#> 50                LSU        SEC Defensive Line  0.9136714      25.5828      28
#> 51                LSU        SEC     Linebacker  0.9198417      11.0381      12
#> 52                LSU        SEC Offensive Line  0.9072435      20.8666      23
#> 53                LSU        SEC    Quarterback  0.8723375       6.9787       8
#> 54                LSU        SEC       Receiver  0.9117375      21.8817      24
#> 55                LSU        SEC   Running Back  0.9235500       9.2355      10
#> 56                LSU        SEC  Special Teams  0.8536000      11.0968      13
#> 57  Mississippi State        SEC Defensive Back  0.8745700      17.4914      20
#> 58  Mississippi State        SEC Defensive Line  0.8814954      19.3929      22
#> 59  Mississippi State        SEC     Linebacker  0.8978857       6.2852       7
#> 60  Mississippi State        SEC Offensive Line  0.8597611      15.4757      18
#> 61  Mississippi State        SEC    Quarterback  0.8763167       5.2579       6
#> 62  Mississippi State        SEC       Receiver  0.8560118      14.5522      17
#> 63  Mississippi State        SEC   Running Back  0.8868833       5.3213       6
#> 64  Mississippi State        SEC  Special Teams  0.8602647      14.6245      17
#> 65           Missouri        SEC Defensive Back  0.8480762      17.8096      21
#> 66           Missouri        SEC Defensive Line  0.8596533      12.8948      15
#> 67           Missouri        SEC     Linebacker  0.8507400      12.7611      15
#> 68           Missouri        SEC Offensive Line  0.8571368      16.2856      19
#> 69           Missouri        SEC    Quarterback  0.8803400       4.4017       5
#> 70           Missouri        SEC       Receiver  0.8580182      18.8764      22
#> 71           Missouri        SEC   Running Back  0.8418875       6.7351       8
#> 72           Missouri        SEC  Special Teams  0.8499800       8.4998      10
#> 73           Ole Miss        SEC Defensive Back  0.8611700      17.2234      20
#> 74           Ole Miss        SEC Defensive Line  0.8740850      17.4817      20
#> 75           Ole Miss        SEC     Linebacker  0.8562667      12.8440      15
#> 76           Ole Miss        SEC Offensive Line  0.8751600      21.8790      25
#> 77           Ole Miss        SEC    Quarterback  0.8837667       7.9539       9
#> 78           Ole Miss        SEC       Receiver  0.9107737      17.3047      19
#> 79           Ole Miss        SEC   Running Back  0.8931778       8.0386       9
#> 80           Ole Miss        SEC  Special Teams  0.8872000       3.5488       4
#> 81     South Carolina        SEC Defensive Back  0.8678158      16.4885      19
#> 82     South Carolina        SEC Defensive Line  0.8892520      22.2313      25
#> 83     South Carolina        SEC     Linebacker  0.8679385      11.2832      13
#> 84     South Carolina        SEC Offensive Line  0.8643846      22.4740      26
#> 85     South Carolina        SEC    Quarterback  0.9167714       6.4174       7
#> 86     South Carolina        SEC       Receiver  0.8702875      20.8869      24
#> 87     South Carolina        SEC   Running Back  0.8800222       7.9202       9
#> 88     South Carolina        SEC  Special Teams  0.8657727       9.5235      11
#> 89          Tennessee        SEC Defensive Back  0.8828609      20.3058      23
#> 90          Tennessee        SEC Defensive Line  0.8997958      21.5951      24
#> 91          Tennessee        SEC     Linebacker  0.8929667      13.3945      15
#> 92          Tennessee        SEC Offensive Line  0.8971682      19.7377      22
#> 93          Tennessee        SEC    Quarterback  0.8852556       7.9673       9
#> 94          Tennessee        SEC       Receiver  0.8760783      20.1498      23
#> 95          Tennessee        SEC   Running Back  0.8682900       8.6829      10
#> 96          Tennessee        SEC  Special Teams  0.8586667      12.8800      15
#> 97          Texas A&M        SEC Defensive Back  0.9171850      18.3437      20
#> 98          Texas A&M        SEC Defensive Line  0.9104000      21.8496      24
#> 99          Texas A&M        SEC     Linebacker  0.8793500      15.8283      18
#> 100         Texas A&M        SEC Offensive Line  0.8868846      23.0590      26
#> 101         Texas A&M        SEC    Quarterback  0.8889500       7.1116       8
#> 102         Texas A&M        SEC       Receiver  0.9113115      23.6941      26
#> 103         Texas A&M        SEC   Running Back  0.8864333      10.6372      12
#> 104         Texas A&M        SEC  Special Teams  0.8424000       8.4240      10
#> 105        Vanderbilt        SEC Defensive Back  0.8577158      16.2966      19
#> 106        Vanderbilt        SEC Defensive Line  0.8307143      11.6300      14
#> 107        Vanderbilt        SEC     Linebacker  0.8474154      11.0164      13
#> 108        Vanderbilt        SEC Offensive Line  0.8457091      18.6056      22
#> 109        Vanderbilt        SEC    Quarterback  0.8758167       5.2549       6
#> 110        Vanderbilt        SEC       Receiver  0.8459190      17.7643      21
#> 111        Vanderbilt        SEC   Running Back  0.8433333       5.0600       6
#> 112        Vanderbilt        SEC  Special Teams  0.8228400       8.2284      10
#> 113           Alabama        SEC  All Positions  0.9335097     135.3590     145
#> 114          Arkansas        SEC  All Positions  0.8648837     111.5700     129
#> 115            Auburn        SEC  All Positions  0.9010294     122.5400     136
#> 116           Florida        SEC  All Positions  0.8950361     119.0400     133
#> 117           Georgia        SEC  All Positions  0.9183477     136.8340     149
#> 118          Kentucky        SEC  All Positions  0.8610457     109.3530     127
#> 119               LSU        SEC  All Positions  0.9087432     126.3150     139
#> 120 Mississippi State        SEC  All Positions  0.8708062      98.4011     113
#> 121          Missouri        SEC  All Positions  0.8544704      98.2641     115
#> 122          Ole Miss        SEC  All Positions  0.8782983     106.2740     121
#> 123    South Carolina        SEC  All Positions  0.8748134     117.2250     134
#> 124         Tennessee        SEC  All Positions  0.8844901     124.7130     141
#> 125         Texas A&M        SEC  All Positions  0.8954687     128.9470     144
#> 126        Vanderbilt        SEC  All Positions  0.8455514      93.8562     111
#>     avg_stars
#> 1    4.000000
#> 2    4.172414
#> 3    4.062500
#> 4    3.956522
#> 5    3.888889
#> 6    3.809524
#> 7    3.928571
#> 8    3.500000
#> 9    3.217391
#> 10   3.320000
#> 11   3.055556
#> 12   3.045455
#> 13   3.500000
#> 14   3.388889
#> 15   3.142857
#> 16   3.125000
#> 17   3.500000
#> 18   3.772727
#> 19   3.650000
#> 20   3.470588
#> 21   3.600000
#> 22   3.541667
#> 23   3.900000
#> 24   3.600000
#> 25   3.481481
#> 26   3.818182
#> 27   3.357143
#> 28   3.461538
#> 29   3.428571
#> 30   3.541667
#> 31   3.500000
#> 32   3.000000
#> 33   3.666667
#> 34   3.931034
#> 35   4.000000
#> 36   3.769231
#> 37   3.857143
#> 38   3.800000
#> 39   4.000000
#> 40   3.454545
#> 41   3.050000
#> 42   3.166667
#> 43   3.363636
#> 44   3.181818
#> 45   3.000000
#> 46   3.047619
#> 47   3.125000
#> 48   3.142857
#> 49   3.952381
#> 50   3.642857
#> 51   3.833333
#> 52   3.521739
#> 53   3.375000
#> 54   3.791667
#> 55   3.900000
#> 56   3.076923
#> 57   3.150000
#> 58   3.409091
#> 59   3.285714
#> 60   3.111111
#> 61   3.500000
#> 62   3.058824
#> 63   3.500000
#> 64   3.000000
#> 65   3.047619
#> 66   3.200000
#> 67   3.000000
#> 68   3.105263
#> 69   3.400000
#> 70   3.045455
#> 71   3.000000
#> 72   2.900000
#> 73   3.050000
#> 74   3.250000
#> 75   3.066667
#> 76   3.200000
#> 77   3.444444
#> 78   3.684211
#> 79   3.555556
#> 80   3.250000
#> 81   3.105263
#> 82   3.520000
#> 83   3.230769
#> 84   3.115385
#> 85   3.714286
#> 86   3.208333
#> 87   3.111111
#> 88   3.363636
#> 89   3.304348
#> 90   3.541667
#> 91   3.533333
#> 92   3.545455
#> 93   3.444444
#> 94   3.260870
#> 95   3.200000
#> 96   3.066667
#> 97   3.750000
#> 98   3.666667
#> 99   3.333333
#> 100  3.461538
#> 101  3.500000
#> 102  3.653846
#> 103  3.416667
#> 104  3.000000
#> 105  3.105263
#> 106  2.785714
#> 107  3.000000
#> 108  3.000000
#> 109  3.166667
#> 110  3.047619
#> 111  3.000000
#> 112  2.700000
#> 113  3.958621
#> 114  3.217054
#> 115  3.617647
#> 116  3.511278
#> 117  3.805369
#> 118  3.133858
#> 119  3.661871
#> 120  3.203540
#> 121  3.069565
#> 122  3.289256
#> 123  3.268657
#> 124  3.375887
#> 125  3.520833
#> 126  2.981982