CFB Recruiting Information - Team Rankings

Gets CFB team recruiting ranks with filters available for year and team. At least one of year or team must be specified for the function to run

cfb_recruiting_team(year = NULL, team = NULL)

Arguments

year
(Integer optional): Recruiting Class Year, 4 digit format (YYYY). Note: 2000 is the minimum value

team
(String optional): Team - Select a valid team, D1 football

Source

https://api.collegefootballdata.com/recruiting/teams

Value

A data frame with 4 variables:

year

integer.

rank

integer.

team

character.

points

character.

Details

If you would like CFB recruiting information for players, please see the cfbscrapR::cfb_recruiting_player() function

If you would like to get CFB recruiting information based on position groups during a time period for all FBS teams, please see the cfbscrapR::cfb_recruiting_position() function.

Examples

cfb_recruiting_team(2018, team = "Texas")

#> 2020-11-30 13:55:09: Scraping team recruiting data...

#>   year rank  team points
#> 1 2018    3 Texas 300.06

cfb_recruiting_team(2016, team = "Virginia")

#> 2020-11-30 13:55:09: Scraping team recruiting data...

#>   year rank     team points
#> 1 2016   63 Virginia 164.95

cfb_recruiting_team(2016, team = "Texas A&M")

#> 2020-11-30 13:55:10: Scraping team recruiting data...

#>   year rank      team points
#> 1 2016   18 Texas A&M 238.94

cfb_recruiting_team(2011)

#> 2020-11-30 13:55:10: Scraping team recruiting data...

#>     year rank                  team points
#> 1   2011    1               Alabama 298.50
#> 2   2011    2         Florida State 297.03
#> 3   2011    3                   USC 286.99
#> 4   2011    4                 Texas 284.05
#> 5   2011    5                Auburn 281.02
#> 6   2011    6            Ohio State 278.47
#> 7   2011    7               Georgia 278.18
#> 8   2011    8                   LSU 272.65
#> 9   2011    9            Notre Dame 271.38
#> 10  2011   10               Clemson 269.69
#> 11  2011   11               Florida 258.24
#> 12  2011   12                Oregon 257.09
#> 13  2011   13              Oklahoma 256.79
#> 14  2011   14             Tennessee 253.78
#> 15  2011   15            California 249.70
#> 16  2011   16              Nebraska 245.60
#> 17  2011   17        South Carolina 244.09
#> 18  2011   18        North Carolina 230.93
#> 19  2011   19            Texas Tech 230.54
#> 20  2011   20              Ole Miss 226.45
#> 21  2011   21              Arkansas 226.22
#> 22  2011   22              Stanford 221.85
#> 23  2011   23              Virginia 220.78
#> 24  2011   24            Washington 219.90
#> 25  2011   25        Oklahoma State 212.82
#> 26  2011   26                  Iowa 208.64
#> 27  2011   27            Louisville 208.34
#> 28  2011   28                   TCU 207.48
#> 29  2011   29               Rutgers 205.88
#> 30  2011   30              Michigan 205.16
#> 31  2011   31            Penn State 201.40
#> 32  2011   32        Michigan State 198.83
#> 33  2011   33                 Miami 198.68
#> 34  2011   34             Texas A&M 193.44
#> 35  2011   35         Virginia Tech 190.02
#> 36  2011   36              Kentucky 184.89
#> 37  2011   37              Illinois 183.92
#> 38  2011   38                  Utah 182.11
#> 39  2011   39        Boston College 180.89
#> 40  2011   40                Kansas 178.77
#> 41  2011   41     Mississippi State 178.10
#> 42  2011   42          Oregon State 175.84
#> 43  2011   43          Georgia Tech 174.84
#> 44  2011   44             Wisconsin 174.08
#> 45  2011   45                  UCLA 172.23
#> 46  2011   46                Baylor 169.94
#> 47  2011   47            Cincinnati 168.77
#> 48  2011   48               Arizona 166.94
#> 49  2011   49            Vanderbilt 164.00
#> 50  2011   50              Maryland 162.97
#> 51  2011   51          Kansas State 161.75
#> 52  2011   52         West Virginia 160.31
#> 53  2011   53           Boise State 159.98
#> 54  2011   54                   UCF 158.74
#> 55  2011   55               Indiana 158.62
#> 56  2011   56              Missouri 158.06
#> 57  2011   57             Minnesota 157.15
#> 58  2011   58              Syracuse 155.37
#> 59  2011   59            Iowa State 153.86
#> 60  2011   60              Marshall 152.84
#> 61  2011   61          Northwestern 150.59
#> 62  2011   62            Pittsburgh 150.28
#> 63  2011   63                  Duke 149.33
#> 64  2011   64              Colorado 147.70
#> 65  2011   65         Arizona State 146.68
#> 66  2011   66           Wake Forest 144.89
#> 67  2011   67                   SMU 144.59
#> 68  2011   68      Washington State 144.31
#> 69  2011   69                   BYU 143.97
#> 70  2011   70         South Florida 143.19
#> 71  2011   71              NC State 139.33
#> 72  2011   72         East Carolina 138.07
#> 73  2011   73               Houston 133.64
#> 74  2011   74                Toledo 132.90
#> 75  2011   75                Purdue 129.88
#> 76  2011   76        Colorado State 122.79
#> 77  2011   77               Memphis 122.48
#> 78  2011   78       San Diego State 118.44
#> 79  2011   79           Connecticut 117.60
#> 80  2011   80                 Tulsa 113.38
#> 81  2011   81                Tulane 110.80
#> 82  2011   82      Western Michigan  94.12
#> 83  2011   83                Temple  93.88
#> 84  2011   84             Louisiana  91.52
#> 85  2011   85      Western Kentucky  85.81
#> 86  2011   86  Southern Mississippi  83.64
#> 87  2011   87                  Navy  81.56
#> 88  2011   88                  Army  77.89
#> 89  2011   89             Air Force  77.86
#> 90  2011   90         Bowling Green  77.59
#> 91  2011   91         South Alabama  76.36
#> 92  2011   92 Florida International  76.24
#> 93  2011   93      Central Michigan  75.64
#> 94  2011   94               Hawai'i  73.77
#> 95  2011   95     Northern Illinois  67.82
#> 96  2011   96                  Ohio  67.80
#> 97  2011   97                  Rice  67.71
#> 98  2011   98          Fresno State  67.30
#> 99  2011   99      Middle Tennessee  66.67
#> 100 2011  100               Buffalo  61.88
#> 101 2011  101                   UAB  61.54
#> 102 2011  102            Ball State  60.53
#> 103 2011  103      Florida Atlantic  60.49
#> 104 2011  104               Wyoming  59.82
#> 105 2011  105            New Mexico  57.85
#> 106 2011  106      Louisiana Monroe  57.72
#> 107 2011  107            Miami (OH)  57.32
#> 108 2011  108            Utah State  56.53
#> 109 2011  109           North Texas  55.16
#> 110 2011  110                  UNLV  53.88
#> 111 2011  111      Eastern Michigan  53.69
#> 112 2011  112                  Troy  51.83
#> 113 2011  113        Louisiana Tech  51.59
#> 114 2011  114                 Akron  50.30
#> 115 2011  115                Nevada  50.28
#> 116 2011  116        San José State  48.83
#> 117 2011  117                 Idaho  48.78
#> 118 2011  118        Arkansas State  48.25
#> 119 2011  119            Kent State  45.57
#> 120 2011  120                  UTEP  35.04
#> 121 2011  121      New Mexico State  30.13
#> 122 2011  122         Northern Iowa  28.92
#> 123 2011  123             Villanova  24.24
#> 124 2011  124  South Carolina State  18.10
#> 125 2011  125               Liberty  13.33
#> 126 2011  126               Harvard  13.29
#> 127 2011  127           Texas State  12.55
#> 128 2011  128            Holy Cross  11.46
#> 129 2011  129          Old Dominion   8.32
#> 130 2011  130                 UMass   6.67
#> 131 2011  131        UT San Antonio   6.67
#> 132 2011  132       Bethune-Cookman   5.99
#> 133 2011  133     Appalachian State   4.32
#> 134 2011  134      Western Carolina   0.00
#> 135 2011  135      Eastern Illinois   0.00
#> 136 2011  136    North Dakota State   0.00
#> 137 2011  137         Indiana State   0.00