Team Info Lookup Lists all teams in conference or all D-I teams if conference is left NULL Current support only for D-I

Team Info Lookup Lists all teams in conference or all D-I teams if conference is left NULL Current support only for D-I

cfb_team_info(conference = NULL, only_fbs = TRUE, year = NULL)

Arguments

conference
(String optional): Conference abbreviation - Select a valid FBS conference Conference abbreviations P5: ACC, B12, B1G, SEC, PAC, Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

only_fbs
(Logical default TRUE): Filter for only returning FBS teams for a given year. If year is left blank while only_fbs is TRUE, then will return values for most current year

year
(Integer optional): Year, 4 digit format (YYYY). Filter for getting a list of major division team for a given year

Examples

cfb_team_info(conference = "SEC")

#>      id            school       mascot abbreviation       alt_name1 alt_name2
#> 1   333           Alabama Crimson Tide          ALA            <NA>       ALA
#> 2     8          Arkansas   Razorbacks          ARK            <NA>       ARK
#> 3     2            Auburn       Tigers          AUB            <NA>       AUB
#> 4    57           Florida       Gators          FLA            <NA>       FLA
#> 5    61           Georgia     Bulldogs          UGA            <NA>       UGA
#> 6    96          Kentucky     Wildcats           UK            <NA>        UK
#> 7    99               LSU       Tigers          LSU Louisiana State       LSU
#> 8   344 Mississippi State     Bulldogs         MSST            <NA>      MSST
#> 9   142          Missouri       Tigers          MIZ            <NA>       MIZ
#> 10  145          Ole Miss       Rebels         MISS            <NA>      MISS
#> 11 2579    South Carolina    Gamecocks           SC            <NA>        SC
#> 12 2633         Tennessee   Volunteers         TENN            <NA>      TENN
#> 13  245         Texas A&M       Aggies         TA&M            <NA>      TA&M
#> 14  238        Vanderbilt   Commodores          VAN            <NA>       VAN
#>         alt_name3 conference division   color alt_color
#> 1         Alabama        SEC     West #690014   #f1f2f3
#> 2        Arkansas        SEC     West #9c1831   #000000
#> 3          Auburn        SEC     West #03244d   #f1f2f3
#> 4         Florida        SEC     East #0021A5   #0021a5
#> 5         Georgia        SEC     East #CC0000   #000000
#> 6        Kentucky        SEC     East #005DAA   #ffffff
#> 7             LSU        SEC     West #22005c   #fdd023
#> 8         Miss St        SEC     West #762123   #c8c8c8
#> 9        Missouri        SEC     East #000000   #000000
#> 10       Ole Miss        SEC     West #001148   #00205b
#> 11 South Carolina        SEC     East #670010   #000000
#> 12      Tennessee        SEC     East #EE9627   #ffffff
#> 13      Texas A&M        SEC     West #5C0025   #ffffff
#> 14     Vanderbilt        SEC     East #000000   #231f20
#>                                                                                                          logos
#> 1    http://a.espncdn.com/i/teamlogos/ncaa/500/333.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/333.png
#> 2        http://a.espncdn.com/i/teamlogos/ncaa/500/8.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/8.png
#> 3        http://a.espncdn.com/i/teamlogos/ncaa/500/2.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2.png
#> 4      http://a.espncdn.com/i/teamlogos/ncaa/500/57.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/57.png
#> 5      http://a.espncdn.com/i/teamlogos/ncaa/500/61.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/61.png
#> 6      http://a.espncdn.com/i/teamlogos/ncaa/500/96.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/96.png
#> 7      http://a.espncdn.com/i/teamlogos/ncaa/500/99.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/99.png
#> 8    http://a.espncdn.com/i/teamlogos/ncaa/500/344.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/344.png
#> 9    http://a.espncdn.com/i/teamlogos/ncaa/500/142.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/142.png
#> 10   http://a.espncdn.com/i/teamlogos/ncaa/500/145.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/145.png
#> 11 http://a.espncdn.com/i/teamlogos/ncaa/500/2579.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2579.png
#> 12 http://a.espncdn.com/i/teamlogos/ncaa/500/2633.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2633.png
#> 13   http://a.espncdn.com/i/teamlogos/ncaa/500/245.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/245.png
#> 14   http://a.espncdn.com/i/teamlogos/ncaa/500/238.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/238.png

cfb_team_info(conference = "Ind")

#>     id           school         mascot abbreviation     alt_name1 alt_name2
#> 1  349             Army  Black Knights         ARMY          <NA>      ARMY
#> 2  252              BYU        Cougars          BYU Brigham Young       BYU
#> 3 2335          Liberty         Flames          LIB          <NA>       LIB
#> 4  166 New Mexico State         Aggies         NMSU          <NA>      NMSU
#> 5   87       Notre Dame Fighting Irish           ND          <NA>        ND
#> 6  113            UMass      Minutemen        UMASS          <NA>     UMASS
#>       alt_name3       conference division   color alt_color
#> 1          Army FBS Independents       NA #ce9c00   #231f20
#> 2           BYU FBS Independents       NA #001E4C   #ffffff
#> 3       Liberty FBS Independents       NA #071740   #a61f21
#> 4 New Mexico St FBS Independents       NA #891216   #000000
#> 5    Notre Dame FBS Independents       NA #00122b   #ae9142
#> 6         UMass FBS Independents       NA #880007      <NA>
#>                                                                                                         logos
#> 1   http://a.espncdn.com/i/teamlogos/ncaa/500/349.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/349.png
#> 2   http://a.espncdn.com/i/teamlogos/ncaa/500/252.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/252.png
#> 3 http://a.espncdn.com/i/teamlogos/ncaa/500/2335.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2335.png
#> 4   http://a.espncdn.com/i/teamlogos/ncaa/500/166.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/166.png
#> 5     http://a.espncdn.com/i/teamlogos/ncaa/500/87.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/87.png
#> 6   http://a.espncdn.com/i/teamlogos/ncaa/500/113.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/113.png

cfb_team_info(year = 2019)

#>       id                school           mascot abbreviation
#> 1   2005             Air Force          Falcons          AFA
#> 2   2006                 Akron             Zips          AKR
#> 3    333               Alabama     Crimson Tide          ALA
#> 4   2026     Appalachian State     Mountaineers          APP
#> 5     12               Arizona         Wildcats         ARIZ
#> 6      9         Arizona State       Sun Devils          ASU
#> 7      8              Arkansas       Razorbacks          ARK
#> 8   2032        Arkansas State       Red Wolves         ARST
#> 9    349                  Army    Black Knights         ARMY
#> 10     2                Auburn           Tigers          AUB
#> 11  2050            Ball State        Cardinals         BALL
#> 12   239                Baylor            Bears          BAY
#> 13    68           Boise State          Broncos          BSU
#> 14   103        Boston College           Eagles           BC
#> 15   189         Bowling Green          Falcons         BGSU
#> 16  2084               Buffalo            Bulls         BUFF
#> 17   252                   BYU          Cougars          BYU
#> 18    25            California     Golden Bears          CAL
#> 19  2117      Central Michigan        Chippewas          CMU
#> 20  2429             Charlotte            49ers         CHAR
#> 21  2132            Cincinnati         Bearcats          CIN
#> 22   228               Clemson           Tigers         CLEM
#> 23   324      Coastal Carolina     Chanticleers          CCU
#> 24    38              Colorado        Buffaloes         COLO
#> 25    36        Colorado State             Rams          CSU
#> 26    41           Connecticut          Huskies         CONN
#> 27   150                  Duke      Blue Devils         DUKE
#> 28   151         East Carolina          Pirates          ECU
#> 29  2199      Eastern Michigan           Eagles          EMU
#> 30    57               Florida           Gators          FLA
#> 31  2226      Florida Atlantic             Owls          FAU
#> 32  2229 Florida International  Golden Panthers          FIU
#> 33    52         Florida State        Seminoles          FSU
#> 34   278          Fresno State         Bulldogs         FRES
#> 35    61               Georgia         Bulldogs          UGA
#> 36   290      Georgia Southern           Eagles         GASO
#> 37  2247         Georgia State         Panthers         GAST
#> 38    59          Georgia Tech   Yellow Jackets           GT
#> 39    62               Hawai'i Rainbow Warriors          HAW
#> 40   248               Houston          Cougars          HOU
#> 41   356              Illinois  Fighting Illini          ILL
#> 42    84               Indiana         Hoosiers          IND
#> 43  2294                  Iowa         Hawkeyes         IOWA
#> 44    66            Iowa State         Cyclones          ISU
#> 45  2305                Kansas         Jayhawks           KU
#> 46  2306          Kansas State         Wildcats          KSU
#> 47  2309            Kent State   Golden Flashes         KENT
#> 48    96              Kentucky         Wildcats           UK
#> 49  2335               Liberty           Flames          LIB
#> 50   309             Louisiana    Ragin' Cajuns          ULL
#> 51  2433      Louisiana Monroe         Warhawks          ULM
#> 52  2348        Louisiana Tech         Bulldogs           LT
#> 53    97            Louisville        Cardinals          LOU
#> 54    99                   LSU           Tigers          LSU
#> 55   276              Marshall  Thundering Herd         MRSH
#> 56   120              Maryland        Terrapins           MD
#> 57   235               Memphis           Tigers          MEM
#> 58  2390                 Miami       Hurricanes        MIAMI
#> 59   193            Miami (OH)         RedHawks         M-OH
#> 60   130              Michigan       Wolverines         MICH
#> 61   127        Michigan State         Spartans          MSU
#> 62  2393      Middle Tennessee     Blue Raiders         MTSU
#> 63   135             Minnesota   Golden Gophers         MINN
#> 64   344     Mississippi State         Bulldogs         MSST
#> 65   142              Missouri           Tigers          MIZ
#> 66  2426                  Navy       Midshipmen         NAVY
#> 67   152              NC State         Wolfpack         NCST
#> 68   158              Nebraska      Cornhuskers          NEB
#> 69  2440                Nevada        Wolf Pack          NEV
#> 70   167            New Mexico            Lobos          UNM
#> 71   166      New Mexico State           Aggies         NMSU
#> 72   153        North Carolina        Tar Heels          UNC
#> 73  2459     Northern Illinois          Huskies          NIU
#> 74   249           North Texas       Mean Green          UNT
#> 75    77          Northwestern         Wildcats           NW
#> 76    87            Notre Dame   Fighting Irish           ND
#> 77   195                  Ohio          Bobcats         OHIO
#> 78   194            Ohio State         Buckeyes          OSU
#> 79   201              Oklahoma          Sooners         OKLA
#> 80   197        Oklahoma State          Cowboys         OKST
#> 81   295          Old Dominion         Monarchs          ODU
#> 82   145              Ole Miss           Rebels         MISS
#> 83  2483                Oregon            Ducks          ORE
#> 84   204          Oregon State          Beavers         ORST
#> 85   213            Penn State    Nittany Lions          PSU
#> 86   221            Pittsburgh         Panthers         PITT
#> 87  2509                Purdue     Boilermakers          PUR
#> 88   242                  Rice             Owls         RICE
#> 89   164               Rutgers  Scarlet Knights         RUTG
#> 90    21       San Diego State           Aztecs         SDSU
#> 91    23        San José State         Spartans         SJSU
#> 92  2567                   SMU         Mustangs          SMU
#> 93     6         South Alabama          Jaguars          USA
#> 94  2579        South Carolina        Gamecocks           SC
#> 95  2572  Southern Mississippi    Golden Eagles          USM
#> 96    58         South Florida            Bulls          USF
#> 97    24              Stanford         Cardinal         STAN
#> 98   183              Syracuse           Orange          SYR
#> 99  2628                   TCU     Horned Frogs          TCU
#> 100  218                Temple             Owls          TEM
#> 101 2633             Tennessee       Volunteers         TENN
#> 102  251                 Texas        Longhorns          TEX
#> 103  245             Texas A&M           Aggies         TA&M
#> 104  326           Texas State          Bobcats         TXST
#> 105 2641            Texas Tech      Red Raiders          TTU
#> 106 2649                Toledo          Rockets          TOL
#> 107 2653                  Troy          Trojans         TROY
#> 108 2655                Tulane       Green Wave         TULN
#> 109  202                 Tulsa Golden Hurricane         TLSA
#> 110    5                   UAB          Blazers          UAB
#> 111 2116                   UCF          Knights          UCF
#> 112   26                  UCLA           Bruins         UCLA
#> 113  113                 UMass        Minutemen        UMASS
#> 114 2439                  UNLV           Rebels         UNLV
#> 115   30                   USC          Trojans          USC
#> 116  254                  Utah             Utes         UTAH
#> 117  328            Utah State           Aggies          USU
#> 118 2638                  UTEP           Miners         UTEP
#> 119 2636        UT San Antonio      Roadrunners         UTSA
#> 120  238            Vanderbilt       Commodores          VAN
#> 121  258              Virginia        Cavaliers          UVA
#> 122  259         Virginia Tech           Hokies           VT
#> 123  154           Wake Forest    Demon Deacons         WAKE
#> 124  264            Washington          Huskies         WASH
#> 125  265      Washington State          Cougars          WSU
#> 126   98      Western Kentucky      Hilltoppers          WKU
#> 127 2711      Western Michigan          Broncos          WMU
#> 128  277         West Virginia     Mountaineers          WVU
#> 129  275             Wisconsin          Badgers          WIS
#> 130 2751               Wyoming          Cowboys          WYO
#>                  alt_name1 alt_name2       alt_name3        conference division
#> 1                     <NA>       AFA       Air Force     Mountain West Mountain
#> 2                     <NA>       AKR           Akron      Mid-American     East
#> 3                     <NA>       ALA         Alabama               SEC     West
#> 4                     <NA>       APP  Appalachian St          Sun Belt     <NA>
#> 5                     <NA>      ARIZ         Arizona            Pac-12    South
#> 6                     <NA>       ASU   Arizona State            Pac-12    South
#> 7                     <NA>       ARK        Arkansas               SEC     West
#> 8                     <NA>      ARST  Arkansas State          Sun Belt     <NA>
#> 9                     <NA>      ARMY            Army  FBS Independents     <NA>
#> 10                    <NA>       AUB          Auburn               SEC     West
#> 11                    <NA>      BALL      Ball State      Mid-American     West
#> 12                    <NA>       BAY          Baylor            Big 12     <NA>
#> 13                    <NA>       BSU     Boise State     Mountain West Mountain
#> 14                    <NA>        BC  Boston College               ACC Atlantic
#> 15     Bowling Green State      BGSU   Bowling Green      Mid-American     East
#> 16                    <NA>      BUFF         Buffalo      Mid-American     East
#> 17           Brigham Young       BYU             BYU  FBS Independents     <NA>
#> 18                    <NA>       CAL             Cal            Pac-12    North
#> 19                    <NA>       CMU   Cent Michigan      Mid-American     West
#> 20                    <NA>      CHAR       Charlotte    Conference USA     East
#> 21                    <NA>       CIN      Cincinnati American Athletic     East
#> 22                    <NA>      CLEM         Clemson               ACC Atlantic
#> 23                    <NA>       CCU     C. Carolina          Sun Belt     <NA>
#> 24                    <NA>      COLO        Colorado            Pac-12    South
#> 25                    <NA>       CSU  Colorado State     Mountain West Mountain
#> 26                    <NA>      CONN           UConn American Athletic     East
#> 27                    <NA>      DUKE            Duke               ACC  Coastal
#> 28                    <NA>       ECU             ECU American Athletic     East
#> 29                    <NA>       EMU      E Michigan      Mid-American     West
#> 30                    <NA>       FLA         Florida               SEC     East
#> 31                    <NA>       FAU             FAU    Conference USA     East
#> 32                    <NA>       FIU             FIU    Conference USA     East
#> 33                    <NA>       FSU             FSU               ACC Atlantic
#> 34                    <NA>      FRES    Fresno State     Mountain West     West
#> 35                    <NA>       UGA         Georgia               SEC     East
#> 36                    <NA>      GASO     Ga Southern          Sun Belt     <NA>
#> 37                    <NA>      GAST   Georgia State          Sun Belt     <NA>
#> 38                    <NA>        GT    Georgia Tech               ACC  Coastal
#> 39                    <NA>       HAW         Hawai'i     Mountain West     West
#> 40                    <NA>       HOU         Houston American Athletic     West
#> 41                    <NA>       ILL        Illinois           Big Ten     West
#> 42                    <NA>       IND         Indiana           Big Ten     East
#> 43                    <NA>      IOWA            Iowa           Big Ten     West
#> 44                    <NA>       ISU      Iowa State            Big 12     <NA>
#> 45                    <NA>        KU          Kansas            Big 12     <NA>
#> 46                    <NA>       KSU    Kansas State            Big 12     <NA>
#> 47                    <NA>      KENT      Kent State      Mid-American     East
#> 48                    <NA>        UK        Kentucky               SEC     East
#> 49                    <NA>       LIB         Liberty  FBS Independents     <NA>
#> 50            UL Lafayette       ULL       Louisiana          Sun Belt     <NA>
#> 51              La.-Monroe       ULM       UL Monroe          Sun Belt     <NA>
#> 52                    <NA>        LT         LA Tech    Conference USA     West
#> 53                    <NA>       LOU      Louisville               ACC Atlantic
#> 54         Louisiana State       LSU             LSU               SEC     West
#> 55                    <NA>      MRSH        Marshall    Conference USA     East
#> 56                    <NA>        MD        Maryland           Big Ten     East
#> 57                    <NA>       MEM         Memphis American Athletic     West
#> 58              Miami (FL)     MIAMI           Miami               ACC  Coastal
#> 59                    <NA>      M-OH      Miami (OH)      Mid-American     East
#> 60                    <NA>      MICH        Michigan           Big Ten     East
#> 61                    <NA>       MSU  Michigan State           Big Ten     East
#> 62  Middle Tennessee State      MTSU   Mid Tennessee    Conference USA     East
#> 63                    <NA>      MINN       Minnesota           Big Ten     West
#> 64                    <NA>      MSST         Miss St               SEC     West
#> 65                    <NA>       MIZ        Missouri               SEC     East
#> 66                    <NA>      NAVY            Navy American Athletic     West
#> 67      North Carolina St.      NCST        NC State               ACC Atlantic
#> 68                    <NA>       NEB        Nebraska           Big Ten     West
#> 69                    <NA>       NEV          Nevada     Mountain West     West
#> 70                    <NA>       UNM      New Mexico     Mountain West Mountain
#> 71                    <NA>      NMSU   New Mexico St  FBS Independents     <NA>
#> 72                    <NA>       UNC             UNC               ACC  Coastal
#> 73                    <NA>       NIU      N Illinois      Mid-American     West
#> 74                    <NA>       UNT     North Texas    Conference USA     West
#> 75                    <NA>        NW    Northwestern           Big Ten     West
#> 76                    <NA>        ND      Notre Dame  FBS Independents     <NA>
#> 77                    <NA>      OHIO            Ohio      Mid-American     East
#> 78                    <NA>       OSU             OSU           Big Ten     East
#> 79                    <NA>      OKLA        Oklahoma            Big 12     <NA>
#> 80                    <NA>      OKST  Oklahoma State            Big 12     <NA>
#> 81                    <NA>       ODU    Old Dominion    Conference USA     East
#> 82                    <NA>      MISS        Ole Miss               SEC     West
#> 83                    <NA>       ORE          Oregon            Pac-12    North
#> 84                    <NA>      ORST    Oregon State            Pac-12    North
#> 85                    <NA>       PSU      Penn State           Big Ten     East
#> 86                    Pitt      PITT            Pitt               ACC  Coastal
#> 87                    <NA>       PUR          Purdue           Big Ten     West
#> 88                    <NA>      RICE            Rice    Conference USA     West
#> 89                    <NA>      RUTG         Rutgers           Big Ten     East
#> 90                    <NA>      SDSU San Diego State     Mountain West     West
#> 91            San Jose St.      SJSU  San Jose State     Mountain West     West
#> 92      Southern Methodist       SMU             SMU American Athletic     West
#> 93                    <NA>       USA   South Alabama          Sun Belt     <NA>
#> 94                    <NA>        SC  South Carolina               SEC     East
#> 95                    <NA>       USM   Southern Miss    Conference USA     West
#> 96                    <NA>       USF             USF American Athletic     East
#> 97                    <NA>      STAN        Stanford            Pac-12    North
#> 98                    <NA>       SYR        Syracuse               ACC Atlantic
#> 99         Texas Christian       TCU             TCU            Big 12     <NA>
#> 100                   <NA>       TEM          Temple American Athletic     East
#> 101                   <NA>      TENN       Tennessee               SEC     East
#> 102                   <NA>       TEX           Texas            Big 12     <NA>
#> 103                   <NA>      TA&M       Texas A&M               SEC     West
#> 104                   <NA>      TXST     Texas State          Sun Belt     <NA>
#> 105                   <NA>       TTU      Texas Tech            Big 12     <NA>
#> 106                   <NA>       TOL          Toledo      Mid-American     West
#> 107                   <NA>      TROY            Troy          Sun Belt     <NA>
#> 108                   <NA>      TULN          Tulane American Athletic     West
#> 109                   <NA>      TLSA           Tulsa American Athletic     West
#> 110     Alabama-Birmingham       UAB             UAB    Conference USA     West
#> 111        Central Florida       UCF             UCF American Athletic     East
#> 112                   <NA>      UCLA            UCLA            Pac-12    South
#> 113                   <NA>     UMASS           UMass  FBS Independents     <NA>
#> 114       Nevada-Las Vegas      UNLV            UNLV     Mountain West     West
#> 115                   <NA>       USC             USC            Pac-12    South
#> 116                   <NA>      UTAH            Utah            Pac-12    South
#> 117                   <NA>       USU      Utah State     Mountain West Mountain
#> 118          Texas-El Paso      UTEP            UTEP    Conference USA     West
#> 119      Texas-San Antonio      UTSA            UTSA    Conference USA     West
#> 120                   <NA>       VAN      Vanderbilt               SEC     East
#> 121                   <NA>       UVA             UVA               ACC  Coastal
#> 122                   <NA>        VT              VT               ACC  Coastal
#> 123                   <NA>      WAKE     Wake Forest               ACC Atlantic
#> 124                   <NA>      WASH      Washington            Pac-12    North
#> 125                   <NA>       WSU   Washington St            Pac-12    North
#> 126                   <NA>       WKU      W Kentucky    Conference USA     East
#> 127                   <NA>       WMU      W Michigan      Mid-American     West
#> 128                   <NA>       WVU   West Virginia            Big 12     <NA>
#> 129                   <NA>       WIS       Wisconsin           Big Ten     West
#> 130                   <NA>       WYO         Wyoming     Mountain West Mountain
#>       color alt_color
#> 1   #004a7b   #ffffff
#> 2   #00285e   #84754e
#> 3   #690014   #f1f2f3
#> 4   #000000   #ffcd00
#> 5   #002449   #00205b
#> 6   #942139   #f1f2f3
#> 7   #9c1831   #000000
#> 8   #e81018   #000000
#> 9   #ce9c00   #231f20
#> 10  #03244d   #f1f2f3
#> 11  #DA0000   #ffffff
#> 12  #004834   #ffb81c
#> 13  #09347A   #d8d9da
#> 14  #88001a   #a39161
#> 15  #2b1000   #492000
#> 16  #041A9B   #ebebeb
#> 17  #001E4C   #ffffff
#> 18  #031522   #ffc423
#> 19  #6a0032   #ffffff
#> 20  #0C562D   #cfab7a
#> 21  #000000   #717073
#> 22  #F66733   #522d80
#> 23  #007073   #876447
#> 24  #d1c57e   #ffd200
#> 25  #004537   #ffc425
#> 26  #001d40   #f1f2f3
#> 27  #001A57   #f1f2f3
#> 28  #4b1869   #f0907b
#> 29  #00331b   #f0f0f0
#> 30  #0021A5   #0021a5
#> 31  #004B85   #bb2f4c
#> 32  #001538   #c5960c
#> 33  #782F40   #ceb888
#> 34  #c41230   #231f20
#> 35  #CC0000   #000000
#> 36  #003775   #f0f0f0
#> 37  #1e539a   #ebebeb
#> 38  #00223e   #002c56
#> 39  #003420   #ffffff
#> 40  #C90822   #231f20
#> 41  #f77329   #fa6300
#> 42  #7D110C   #eeedeb
#> 43  #000000   #ffe100
#> 44  #660015   #830b2c
#> 45  #0022B4   #e8000d
#> 46  #633194   #e7d2ad
#> 47  #002445   #f0b510
#> 48  #005DAA   #ffffff
#> 49  #071740   #a61f21
#> 50  #ce2842      <NA>
#> 51  #231F20   #b18445
#> 52  #002d65   #d3313a
#> 53  #ad000a   #cccccc
#> 54  #22005c   #fdd023
#> 55  #186329   #be854c
#> 56  #D5002B   #ffcd00
#> 57  #002447   #231f20
#> 58  #004325   #f0f0f0
#> 59  #a4000c   #f0f0f0
#> 60  #00274c   #ffcb05
#> 61  #18453B   #ffffff
#> 62  #0079C2   #c3c5c8
#> 63  #7F011B   #7a0019
#> 64  #762123   #c8c8c8
#> 65  #000000   #000000
#> 66  #131630   #b6a77c
#> 67  #EF1216   #231f20
#> 68  #F20017   #f5f1e7
#> 69  #153E5F   #8e9295
#> 70  #000000   #231f20
#> 71  #891216   #000000
#> 72  #99bfe5   #13294b
#> 73  #F1122C   #cc0000
#> 74  #00853D   #000000
#> 75  #372286   #d6cac1
#> 76  #00122b   #ae9142
#> 77  #295A29   #e4bb85
#> 78  #DE3121   #666666
#> 79  #7b0000   #cccccc
#> 80  #FF6500   #ff9900
#> 81  #00507d   #a1d2f1
#> 82  #001148   #00205b
#> 83  #044520   #fee123
#> 84  #c34500   #dea076
#> 85  #00265D   #002e5c
#> 86  #003263   #231f20
#> 87  #B89D29   #a4a9ad
#> 88  #003D7D   #d1d5d8
#> 89  #d21034   #ffffff
#> 90  #BF2C37   #a8996e
#> 91  #005893   #fdba31
#> 92  #E32F38   #ecdcb9
#> 93  #003E7E      <NA>
#> 94  #670010   #000000
#> 95  #FFAA3C   #ffc423
#> 96  #004A36   #231f20
#> 97  #A80532   #ffffff
#> 98  #F37321   #0d1d37
#> 99  #3C377D   #f1f2f3
#> 100 #A80532   #a7a9ac
#> 101 #EE9627   #ffffff
#> 102 #EE7524   #f0f0f0
#> 103 #5C0025   #ffffff
#> 104 #4e1719   #b4975a
#> 105 #C80025   #231f20
#> 106 #00488F   #ffd51d
#> 107 #AE0210   #88898c
#> 108 #005837   #3baf29
#> 109 #004371   #ee3b33
#> 110 #003b28   #ffc845
#> 111 #000000   #231f20
#> 112 #005C8E   #ffc72c
#> 113 #880007      <NA>
#> 114 #000000   #c2c4c6
#> 115 #AE2531   #ffc72c
#> 116 #890012   #7e8083
#> 117 #003263   #949ca1
#> 118 #002843   #ff6f12
#> 119 #002A5C   #f47321
#> 120 #000000   #231f20
#> 121 #00204e   #002d62
#> 122 #74232D   #c2c1ba
#> 123 #9E7E38   #000000
#> 124 #2B2F64   #e8e3d3
#> 125 #94022a   #6a747c
#> 126 #F32026   #b3b5b8
#> 127 #B59966   #bfa979
#> 128 #FFC600   #eaaa00
#> 129 #A00001   #f7f7f7
#> 130 #533B22   #ffffff
#>                                                                                                           logos
#> 1   http://a.espncdn.com/i/teamlogos/ncaa/500/2005.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2005.png
#> 2   http://a.espncdn.com/i/teamlogos/ncaa/500/2006.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2006.png
#> 3     http://a.espncdn.com/i/teamlogos/ncaa/500/333.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/333.png
#> 4   http://a.espncdn.com/i/teamlogos/ncaa/500/2026.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2026.png
#> 5       http://a.espncdn.com/i/teamlogos/ncaa/500/12.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/12.png
#> 6         http://a.espncdn.com/i/teamlogos/ncaa/500/9.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/9.png
#> 7         http://a.espncdn.com/i/teamlogos/ncaa/500/8.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/8.png
#> 8   http://a.espncdn.com/i/teamlogos/ncaa/500/2032.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2032.png
#> 9     http://a.espncdn.com/i/teamlogos/ncaa/500/349.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/349.png
#> 10        http://a.espncdn.com/i/teamlogos/ncaa/500/2.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2.png
#> 11  http://a.espncdn.com/i/teamlogos/ncaa/500/2050.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2050.png
#> 12    http://a.espncdn.com/i/teamlogos/ncaa/500/239.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/239.png
#> 13      http://a.espncdn.com/i/teamlogos/ncaa/500/68.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/68.png
#> 14    http://a.espncdn.com/i/teamlogos/ncaa/500/103.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/103.png
#> 15    http://a.espncdn.com/i/teamlogos/ncaa/500/189.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/189.png
#> 16  http://a.espncdn.com/i/teamlogos/ncaa/500/2084.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2084.png
#> 17    http://a.espncdn.com/i/teamlogos/ncaa/500/252.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/252.png
#> 18      http://a.espncdn.com/i/teamlogos/ncaa/500/25.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/25.png
#> 19  http://a.espncdn.com/i/teamlogos/ncaa/500/2117.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2117.png
#> 20  http://a.espncdn.com/i/teamlogos/ncaa/500/2429.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2429.png
#> 21  http://a.espncdn.com/i/teamlogos/ncaa/500/2132.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2132.png
#> 22    http://a.espncdn.com/i/teamlogos/ncaa/500/228.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/228.png
#> 23    http://a.espncdn.com/i/teamlogos/ncaa/500/324.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/324.png
#> 24      http://a.espncdn.com/i/teamlogos/ncaa/500/38.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/38.png
#> 25      http://a.espncdn.com/i/teamlogos/ncaa/500/36.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/36.png
#> 26      http://a.espncdn.com/i/teamlogos/ncaa/500/41.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/41.png
#> 27    http://a.espncdn.com/i/teamlogos/ncaa/500/150.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/150.png
#> 28    http://a.espncdn.com/i/teamlogos/ncaa/500/151.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/151.png
#> 29  http://a.espncdn.com/i/teamlogos/ncaa/500/2199.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2199.png
#> 30      http://a.espncdn.com/i/teamlogos/ncaa/500/57.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/57.png
#> 31  http://a.espncdn.com/i/teamlogos/ncaa/500/2226.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2226.png
#> 32  http://a.espncdn.com/i/teamlogos/ncaa/500/2229.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2229.png
#> 33      http://a.espncdn.com/i/teamlogos/ncaa/500/52.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/52.png
#> 34    http://a.espncdn.com/i/teamlogos/ncaa/500/278.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/278.png
#> 35      http://a.espncdn.com/i/teamlogos/ncaa/500/61.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/61.png
#> 36    http://a.espncdn.com/i/teamlogos/ncaa/500/290.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/290.png
#> 37  http://a.espncdn.com/i/teamlogos/ncaa/500/2247.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2247.png
#> 38      http://a.espncdn.com/i/teamlogos/ncaa/500/59.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/59.png
#> 39      http://a.espncdn.com/i/teamlogos/ncaa/500/62.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/62.png
#> 40    http://a.espncdn.com/i/teamlogos/ncaa/500/248.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/248.png
#> 41    http://a.espncdn.com/i/teamlogos/ncaa/500/356.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/356.png
#> 42      http://a.espncdn.com/i/teamlogos/ncaa/500/84.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/84.png
#> 43  http://a.espncdn.com/i/teamlogos/ncaa/500/2294.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2294.png
#> 44      http://a.espncdn.com/i/teamlogos/ncaa/500/66.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/66.png
#> 45  http://a.espncdn.com/i/teamlogos/ncaa/500/2305.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2305.png
#> 46  http://a.espncdn.com/i/teamlogos/ncaa/500/2306.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2306.png
#> 47  http://a.espncdn.com/i/teamlogos/ncaa/500/2309.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2309.png
#> 48      http://a.espncdn.com/i/teamlogos/ncaa/500/96.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/96.png
#> 49  http://a.espncdn.com/i/teamlogos/ncaa/500/2335.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2335.png
#> 50    http://a.espncdn.com/i/teamlogos/ncaa/500/309.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/309.png
#> 51  http://a.espncdn.com/i/teamlogos/ncaa/500/2433.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2433.png
#> 52  http://a.espncdn.com/i/teamlogos/ncaa/500/2348.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2348.png
#> 53      http://a.espncdn.com/i/teamlogos/ncaa/500/97.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/97.png
#> 54      http://a.espncdn.com/i/teamlogos/ncaa/500/99.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/99.png
#> 55    http://a.espncdn.com/i/teamlogos/ncaa/500/276.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/276.png
#> 56    http://a.espncdn.com/i/teamlogos/ncaa/500/120.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/120.png
#> 57    http://a.espncdn.com/i/teamlogos/ncaa/500/235.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/235.png
#> 58  http://a.espncdn.com/i/teamlogos/ncaa/500/2390.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2390.png
#> 59    http://a.espncdn.com/i/teamlogos/ncaa/500/193.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/193.png
#> 60    http://a.espncdn.com/i/teamlogos/ncaa/500/130.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/130.png
#> 61    http://a.espncdn.com/i/teamlogos/ncaa/500/127.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/127.png
#> 62  http://a.espncdn.com/i/teamlogos/ncaa/500/2393.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2393.png
#> 63    http://a.espncdn.com/i/teamlogos/ncaa/500/135.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/135.png
#> 64    http://a.espncdn.com/i/teamlogos/ncaa/500/344.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/344.png
#> 65    http://a.espncdn.com/i/teamlogos/ncaa/500/142.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/142.png
#> 66  http://a.espncdn.com/i/teamlogos/ncaa/500/2426.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2426.png
#> 67    http://a.espncdn.com/i/teamlogos/ncaa/500/152.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/152.png
#> 68    http://a.espncdn.com/i/teamlogos/ncaa/500/158.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/158.png
#> 69  http://a.espncdn.com/i/teamlogos/ncaa/500/2440.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2440.png
#> 70    http://a.espncdn.com/i/teamlogos/ncaa/500/167.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/167.png
#> 71    http://a.espncdn.com/i/teamlogos/ncaa/500/166.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/166.png
#> 72    http://a.espncdn.com/i/teamlogos/ncaa/500/153.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/153.png
#> 73  http://a.espncdn.com/i/teamlogos/ncaa/500/2459.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2459.png
#> 74    http://a.espncdn.com/i/teamlogos/ncaa/500/249.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/249.png
#> 75      http://a.espncdn.com/i/teamlogos/ncaa/500/77.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/77.png
#> 76      http://a.espncdn.com/i/teamlogos/ncaa/500/87.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/87.png
#> 77    http://a.espncdn.com/i/teamlogos/ncaa/500/195.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/195.png
#> 78    http://a.espncdn.com/i/teamlogos/ncaa/500/194.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/194.png
#> 79    http://a.espncdn.com/i/teamlogos/ncaa/500/201.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/201.png
#> 80    http://a.espncdn.com/i/teamlogos/ncaa/500/197.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/197.png
#> 81    http://a.espncdn.com/i/teamlogos/ncaa/500/295.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/295.png
#> 82    http://a.espncdn.com/i/teamlogos/ncaa/500/145.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/145.png
#> 83  http://a.espncdn.com/i/teamlogos/ncaa/500/2483.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2483.png
#> 84    http://a.espncdn.com/i/teamlogos/ncaa/500/204.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/204.png
#> 85    http://a.espncdn.com/i/teamlogos/ncaa/500/213.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/213.png
#> 86    http://a.espncdn.com/i/teamlogos/ncaa/500/221.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/221.png
#> 87  http://a.espncdn.com/i/teamlogos/ncaa/500/2509.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2509.png
#> 88    http://a.espncdn.com/i/teamlogos/ncaa/500/242.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/242.png
#> 89    http://a.espncdn.com/i/teamlogos/ncaa/500/164.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/164.png
#> 90      http://a.espncdn.com/i/teamlogos/ncaa/500/21.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/21.png
#> 91      http://a.espncdn.com/i/teamlogos/ncaa/500/23.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/23.png
#> 92  http://a.espncdn.com/i/teamlogos/ncaa/500/2567.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2567.png
#> 93        http://a.espncdn.com/i/teamlogos/ncaa/500/6.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/6.png
#> 94  http://a.espncdn.com/i/teamlogos/ncaa/500/2579.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2579.png
#> 95  http://a.espncdn.com/i/teamlogos/ncaa/500/2572.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2572.png
#> 96      http://a.espncdn.com/i/teamlogos/ncaa/500/58.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/58.png
#> 97      http://a.espncdn.com/i/teamlogos/ncaa/500/24.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/24.png
#> 98    http://a.espncdn.com/i/teamlogos/ncaa/500/183.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/183.png
#> 99  http://a.espncdn.com/i/teamlogos/ncaa/500/2628.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2628.png
#> 100   http://a.espncdn.com/i/teamlogos/ncaa/500/218.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/218.png
#> 101 http://a.espncdn.com/i/teamlogos/ncaa/500/2633.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2633.png
#> 102   http://a.espncdn.com/i/teamlogos/ncaa/500/251.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/251.png
#> 103   http://a.espncdn.com/i/teamlogos/ncaa/500/245.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/245.png
#> 104   http://a.espncdn.com/i/teamlogos/ncaa/500/326.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/326.png
#> 105 http://a.espncdn.com/i/teamlogos/ncaa/500/2641.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2641.png
#> 106 http://a.espncdn.com/i/teamlogos/ncaa/500/2649.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2649.png
#> 107 http://a.espncdn.com/i/teamlogos/ncaa/500/2653.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2653.png
#> 108 http://a.espncdn.com/i/teamlogos/ncaa/500/2655.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2655.png
#> 109   http://a.espncdn.com/i/teamlogos/ncaa/500/202.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/202.png
#> 110       http://a.espncdn.com/i/teamlogos/ncaa/500/5.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/5.png
#> 111 http://a.espncdn.com/i/teamlogos/ncaa/500/2116.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2116.png
#> 112     http://a.espncdn.com/i/teamlogos/ncaa/500/26.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/26.png
#> 113   http://a.espncdn.com/i/teamlogos/ncaa/500/113.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/113.png
#> 114 http://a.espncdn.com/i/teamlogos/ncaa/500/2439.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2439.png
#> 115     http://a.espncdn.com/i/teamlogos/ncaa/500/30.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/30.png
#> 116   http://a.espncdn.com/i/teamlogos/ncaa/500/254.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/254.png
#> 117   http://a.espncdn.com/i/teamlogos/ncaa/500/328.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/328.png
#> 118 http://a.espncdn.com/i/teamlogos/ncaa/500/2638.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2638.png
#> 119 http://a.espncdn.com/i/teamlogos/ncaa/500/2636.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2636.png
#> 120   http://a.espncdn.com/i/teamlogos/ncaa/500/238.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/238.png
#> 121   http://a.espncdn.com/i/teamlogos/ncaa/500/258.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/258.png
#> 122   http://a.espncdn.com/i/teamlogos/ncaa/500/259.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/259.png
#> 123   http://a.espncdn.com/i/teamlogos/ncaa/500/154.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/154.png
#> 124   http://a.espncdn.com/i/teamlogos/ncaa/500/264.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/264.png
#> 125   http://a.espncdn.com/i/teamlogos/ncaa/500/265.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/265.png
#> 126     http://a.espncdn.com/i/teamlogos/ncaa/500/98.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/98.png
#> 127 http://a.espncdn.com/i/teamlogos/ncaa/500/2711.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2711.png
#> 128   http://a.espncdn.com/i/teamlogos/ncaa/500/277.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/277.png
#> 129   http://a.espncdn.com/i/teamlogos/ncaa/500/275.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/275.png
#> 130 http://a.espncdn.com/i/teamlogos/ncaa/500/2751.png, http://a.espncdn.com/i/teamlogos/ncaa/500-dark/2751.png