Get matchup history between two teams.

Get matchup history between two teams.

cfb_team_matchup(team1, team2, min_year = NULL, max_year = NULL)

Arguments

team1
(String required): D-I Team 1

team2
(String required): D-I Team 2

min_year
(Integer optional): Minimum of year range, 4 digit format (YYYY)

max_year
(Integer optional): Maximum of year range, 4 digit format (YYYY)

Examples

cfb_team_matchup('Texas','Oklahoma')

#>     season week season_type                     date neutral_site
#> 1     1902    4     regular 1902-10-04T00:00:00.000Z        FALSE
#> 2     1903   10     regular 1903-11-13T00:00:00.000Z        FALSE
#> 3     1903    6     regular 1903-10-17T00:00:00.000Z        FALSE
#> 4     1904   10     regular 1904-11-16T00:00:00.000Z        FALSE
#> 5     1905    8     regular 1905-11-03T00:00:00.000Z        FALSE
#> 6     1906    7     regular 1906-11-02T00:00:00.000Z        FALSE
#> 7     1907    9     regular 1907-11-15T00:00:00.000Z        FALSE
#> 8     1908    9     regular 1908-11-13T00:00:00.000Z        FALSE
#> 9     1909   10     regular 1909-11-19T00:00:00.000Z        FALSE
#> 10    1910   11     regular 1910-11-24T00:00:00.000Z        FALSE
#> 11    1911   10     regular 1911-11-25T00:00:00.000Z        FALSE
#> 12    1912    5     regular 1912-10-19T00:00:00.000Z        FALSE
#> 13    1913    8     regular 1913-11-10T00:00:00.000Z        FALSE
#> 14    1914    6     regular 1914-10-24T00:00:00.000Z        FALSE
#> 15    1915    6     regular 1915-10-23T00:00:00.000Z        FALSE
#> 16    1916    6     regular 1916-10-21T00:00:00.000Z        FALSE
#> 17    1917    6     regular 1917-10-20T00:00:00.000Z        FALSE
#> 18    1919    5     regular 1919-10-18T00:00:00.000Z        FALSE
#> 19    1922    9     regular 1922-11-18T00:00:00.000Z        FALSE
#> 20    1923    9     regular 1923-11-17T00:00:00.000Z        FALSE
#> 21    1929    5     regular 1929-10-19T00:00:00.000Z        FALSE
#> 22    1930    6     regular 1930-10-18T00:00:00.000Z        FALSE
#> 23    1931    6     regular 1931-10-17T00:00:00.000Z        FALSE
#> 24    1932    5     regular 1932-10-15T00:00:00.000Z        FALSE
#> 25    1933    5     regular 1933-10-14T00:00:00.000Z        FALSE
#> 26    1934    5     regular 1934-10-13T00:00:00.000Z        FALSE
#> 27    1935    4     regular 1935-10-12T00:00:00.000Z        FALSE
#> 28    1936    4     regular 1936-10-10T00:00:00.000Z        FALSE
#> 29    1937    4     regular 1937-10-09T00:00:00.000Z        FALSE
#> 30    1938    4     regular 1938-10-08T00:00:00.000Z        FALSE
#> 31    1939    5     regular 1939-10-14T00:00:00.000Z        FALSE
#> 32    1940    5     regular 1940-10-12T00:00:00.000Z        FALSE
#> 33    1941    4     regular 1941-10-11T00:00:00.000Z        FALSE
#> 34    1942    5     regular 1942-10-10T00:00:00.000Z        FALSE
#> 35    1943    5     regular 1943-10-09T00:00:00.000Z        FALSE
#> 36    1944    8     regular 1944-10-14T00:00:00.000Z        FALSE
#> 37    1945    6     regular 1945-10-13T00:00:00.000Z        FALSE
#> 38    1946    5     regular 1946-10-12T00:00:00.000Z        FALSE
#> 39    1947    5     regular 1947-10-11T00:00:00.000Z        FALSE
#> 40    1948    5     regular 1948-10-09T00:00:00.000Z        FALSE
#> 41    1949    5     regular 1949-10-08T00:00:00.000Z        FALSE
#> 42    1950    5     regular 1950-10-14T00:00:00.000Z        FALSE
#> 43    1951    5     regular 1951-10-13T00:00:00.000Z        FALSE
#> 44    1952    5     regular 1952-10-11T00:00:00.000Z        FALSE
#> 45    1953    4     regular 1953-10-10T00:00:00.000Z        FALSE
#> 46    1954    4     regular 1954-10-09T00:00:00.000Z        FALSE
#> 47    1955    4     regular 1955-10-08T00:00:00.000Z        FALSE
#> 48    1956    5     regular 1956-10-13T00:00:00.000Z        FALSE
#> 49    1957    5     regular 1957-10-12T00:00:00.000Z        FALSE
#> 50    1958    5     regular 1958-10-11T00:00:00.000Z        FALSE
#> 51    1959    5     regular 1959-10-10T00:00:00.000Z        FALSE
#> 52    1960    5     regular 1960-10-08T00:00:00.000Z        FALSE
#> 53    1961    5     regular 1961-10-14T00:00:00.000Z        FALSE
#> 54    1962    5     regular 1962-10-13T00:00:00.000Z        FALSE
#> 55    1963    5     regular 1963-10-12T00:00:00.000Z        FALSE
#> 56    1964    5     regular 1964-10-10T00:00:00.000Z        FALSE
#> 57    1965    5     regular 1965-10-09T00:00:00.000Z        FALSE
#> 58    1966    5     regular 1966-10-08T00:00:00.000Z        FALSE
#> 59    1967    6     regular 1967-10-14T00:00:00.000Z        FALSE
#> 60    1968    5     regular 1968-10-12T00:00:00.000Z        FALSE
#> 61    1969    5     regular 1969-10-11T00:00:00.000Z        FALSE
#> 62    1970    7     regular 1970-10-10T00:00:00.000Z        FALSE
#> 63    1971    5     regular 1971-10-09T00:00:00.000Z        FALSE
#> 64    1972    6     regular 1972-10-14T00:00:00.000Z        FALSE
#> 65    1973    7     regular 1973-10-13T00:00:00.000Z        FALSE
#> 66    1974    6     regular 1974-10-12T00:00:00.000Z        FALSE
#> 67    1975    6     regular 1975-10-11T00:00:00.000Z        FALSE
#> 68    1976    6     regular 1976-10-09T00:00:00.000Z        FALSE
#> 69    1977    6     regular 1977-10-08T00:00:00.000Z        FALSE
#> 70    1978    6     regular 1978-10-07T00:00:00.000Z        FALSE
#> 71    1979    7     regular 1979-10-13T00:00:00.000Z        FALSE
#> 72    1980    7     regular 1980-10-11T00:00:00.000Z        FALSE
#> 73    1981    6     regular 1981-10-10T00:00:00.000Z        FALSE
#> 74    1982    6     regular 1982-10-09T00:00:00.000Z        FALSE
#> 75    1983    7     regular 1983-10-08T00:00:00.000Z        FALSE
#> 76    1984    8     regular 1984-10-13T00:00:00.000Z        FALSE
#> 77    1985    7     regular 1985-10-12T00:00:00.000Z        FALSE
#> 78    1986    7     regular 1986-10-11T00:00:00.000Z        FALSE
#> 79    1987    7     regular 1987-10-10T00:00:00.000Z        FALSE
#> 80    1988    7     regular 1988-10-08T00:00:00.000Z        FALSE
#> 81    1989    7     regular 1989-10-14T00:00:00.000Z        FALSE
#> 82    1990    8     regular 1990-10-13T00:00:00.000Z        FALSE
#> 83    1991    7     regular 1991-10-12T00:00:00.000Z        FALSE
#> 84    1992    7     regular 1992-10-10T00:00:00.000Z        FALSE
#> 85    1993    7     regular 1993-10-09T00:00:00.000Z        FALSE
#> 86    1994    7     regular 1994-10-08T00:00:00.000Z        FALSE
#> 87    1995    8     regular 1995-10-14T00:00:00.000Z        FALSE
#> 88    1996    8     regular 1996-10-12T00:00:00.000Z        FALSE
#> 89    1997    8     regular 1997-10-11T00:00:00.000Z        FALSE
#> 90    1998    7     regular 1998-10-10T00:00:00.000Z        FALSE
#> 91    1999    7     regular 1999-10-09T00:00:00.000Z        FALSE
#> 92    2000    7     regular 2000-10-07T00:00:00.000Z        FALSE
#> 93    2001    7     regular 2001-10-06T15:30:00.000Z        FALSE
#> 94    2002    8     regular 2002-10-12T15:30:00.000Z        FALSE
#> 95    2003    8     regular 2003-10-11T15:30:00.000Z        FALSE
#> 96    2004    7     regular 2004-10-09T12:00:00.000Z        FALSE
#> 97    2005    6     regular 2005-10-08T13:00:00.000Z        FALSE
#> 98    2006    6     regular 2006-10-07T15:30:00.000Z        FALSE
#> 99    2007    6     regular 2007-10-06T15:30:00.000Z        FALSE
#> 100   2008    7     regular 2008-10-11T12:00:00.000Z         TRUE
#> 101   2009    7     regular 2009-10-17T12:00:00.000Z         TRUE
#> 102   2010    5     regular 2010-10-02T15:30:00.000Z         TRUE
#> 103   2011    6     regular 2011-10-08T12:00:00.000Z         TRUE
#> 104   2012    7     regular 2012-10-13T12:00:00.000Z         TRUE
#> 105   2013    7     regular 2013-10-12T12:00:00.000Z         TRUE
#> 106   2014    7     regular 2014-10-11T12:00:00.000Z         TRUE
#> 107   2015    6     regular 2015-10-10T12:00:00.000Z         TRUE
#> 108   2016    6     regular 2016-10-08T12:00:00.000Z         TRUE
#> 109   2017    7     regular 2017-10-14T15:30:00.000Z         TRUE
#> 110   2018   14     regular 2018-12-01T17:00:00.000Z         TRUE
#> 111   2018    6     regular 2018-10-06T16:00:00.000Z         TRUE
#> 112   2019    7     regular 2019-10-12T16:00:00.000Z         TRUE
#>                                        venue home_team home_score away_team
#> 1                                       <NA>     Texas         22  Oklahoma
#> 2                                       <NA>  Oklahoma          5     Texas
#> 3                                       <NA>  Oklahoma          6     Texas
#> 4                                       <NA>     Texas         40  Oklahoma
#> 5                                       <NA>  Oklahoma          2     Texas
#> 6                                       <NA>  Oklahoma          9     Texas
#> 7                                       <NA>     Texas         29  Oklahoma
#> 8                                       <NA>  Oklahoma         50     Texas
#> 9                                       <NA>     Texas         30  Oklahoma
#> 10                                      <NA>     Texas          0  Oklahoma
#> 11                                      <NA>  Oklahoma          6     Texas
#> 12                                      <NA>  Oklahoma         21     Texas
#> 13                                      <NA>     Texas         14  Oklahoma
#> 14                                      <NA>     Texas         32  Oklahoma
#> 15                                      <NA>  Oklahoma         14     Texas
#> 16                                      <NA>     Texas         21  Oklahoma
#> 17                                      <NA>  Oklahoma         14     Texas
#> 18                                      <NA>  Oklahoma         12     Texas
#> 19                                      <NA>  Oklahoma          7     Texas
#> 20                                      <NA>     Texas         26  Oklahoma
#> 21                                      <NA>     Texas         21  Oklahoma
#> 22                                      <NA>     Texas         17  Oklahoma
#> 23                                      <NA>     Texas          3  Oklahoma
#> 24                                      <NA>     Texas         17  Oklahoma
#> 25                                      <NA>  Oklahoma          9     Texas
#> 26                                      <NA>     Texas         19  Oklahoma
#> 27                                      <NA>     Texas         12  Oklahoma
#> 28                                      <NA>     Texas          6  Oklahoma
#> 29                                      <NA>     Texas          7  Oklahoma
#> 30                                      <NA>  Oklahoma         13     Texas
#> 31                                      <NA>  Oklahoma         24     Texas
#> 32                                      <NA>     Texas         19  Oklahoma
#> 33                                      <NA>     Texas         40  Oklahoma
#> 34                                      <NA>     Texas          7  Oklahoma
#> 35                                      <NA>     Texas         13  Oklahoma
#> 36                                      <NA>     Texas         20  Oklahoma
#> 37                                      <NA>     Texas         12  Oklahoma
#> 38                                      <NA>     Texas         20  Oklahoma
#> 39                                      <NA>     Texas         34  Oklahoma
#> 40                                      <NA>  Oklahoma         20     Texas
#> 41                                      <NA>  Oklahoma         20     Texas
#> 42                                      <NA>  Oklahoma         14     Texas
#> 43                                      <NA>     Texas          9  Oklahoma
#> 44                                      <NA>  Oklahoma         49     Texas
#> 45                                      <NA>  Oklahoma         19     Texas
#> 46                                      <NA>  Oklahoma         14     Texas
#> 47                                      <NA>  Oklahoma         20     Texas
#> 48                                      <NA>  Oklahoma         45     Texas
#> 49                                      <NA>  Oklahoma         21     Texas
#> 50                                      <NA>     Texas         15  Oklahoma
#> 51                                      <NA>     Texas         19  Oklahoma
#> 52                                      <NA>     Texas         24  Oklahoma
#> 53                                      <NA>     Texas         28  Oklahoma
#> 54                                      <NA>     Texas          9  Oklahoma
#> 55                                      <NA>     Texas         28  Oklahoma
#> 56                                      <NA>     Texas         28  Oklahoma
#> 57                                      <NA>     Texas         19  Oklahoma
#> 58                                      <NA>  Oklahoma         18     Texas
#> 59                                      <NA>     Texas          9  Oklahoma
#> 60                                      <NA>     Texas         26  Oklahoma
#> 61                                      <NA>     Texas         27  Oklahoma
#> 62                                      <NA>     Texas         41  Oklahoma
#> 63                                      <NA>  Oklahoma         48     Texas
#> 64                                      <NA>  Oklahoma         27     Texas
#> 65                                      <NA>  Oklahoma         52     Texas
#> 66                                      <NA>  Oklahoma         16     Texas
#> 67                                      <NA>  Oklahoma         24     Texas
#> 68                                      <NA>     Texas          6  Oklahoma
#> 69                                      <NA>     Texas         13  Oklahoma
#> 70                                      <NA>  Oklahoma         31     Texas
#> 71                                      <NA>     Texas         16  Oklahoma
#> 72                                      <NA>     Texas         20  Oklahoma
#> 73                                      <NA>     Texas         34  Oklahoma
#> 74                                      <NA>  Oklahoma         28     Texas
#> 75                                      <NA>     Texas         28  Oklahoma
#> 76                                      <NA>     Texas         15  Oklahoma
#> 77                                      <NA>  Oklahoma         14     Texas
#> 78                                      <NA>  Oklahoma         47     Texas
#> 79                                      <NA>  Oklahoma         44     Texas
#> 80                                      <NA>  Oklahoma         28     Texas
#> 81                                      <NA>     Texas         28  Oklahoma
#> 82                                      <NA>     Texas         14  Oklahoma
#> 83                                      <NA>     Texas         10  Oklahoma
#> 84                                      <NA>     Texas         34  Oklahoma
#> 85                                      <NA>  Oklahoma         38     Texas
#> 86                                      <NA>     Texas         17  Oklahoma
#> 87                                      <NA>     Texas         24  Oklahoma
#> 88                                      <NA>  Oklahoma         30     Texas
#> 89                                      <NA>     Texas         27  Oklahoma
#> 90                                      <NA>     Texas         34  Oklahoma
#> 91                                      <NA>     Texas         38  Oklahoma
#> 92                                      <NA>  Oklahoma         63     Texas
#> 93                                      <NA>     Texas          3  Oklahoma
#> 94                                      <NA>  Oklahoma         35     Texas
#> 95                                      <NA>     Texas         13  Oklahoma
#> 96  Gaylord Family Oklahoma Memorial Stadium  Oklahoma         12     Texas
#> 97                                      <NA>     Texas         45  Oklahoma
#> 98                                      <NA>  Oklahoma         10     Texas
#> 99                                      <NA>     Texas         21  Oklahoma
#> 100                                     <NA>  Oklahoma         35     Texas
#> 101                                     <NA>     Texas         16  Oklahoma
#> 102                                     <NA>  Oklahoma         28     Texas
#> 103                                     <NA>     Texas         17  Oklahoma
#> 104                                     <NA>  Oklahoma         63     Texas
#> 105                                     <NA>     Texas         36  Oklahoma
#> 106                              Cotton Bowl  Oklahoma         31     Texas
#> 107                              Cotton Bowl     Texas         24  Oklahoma
#> 108                              Cotton Bowl  Oklahoma         45     Texas
#> 109                              Cotton Bowl     Texas         24  Oklahoma
#> 110                             AT&T Stadium  Oklahoma         39     Texas
#> 111                              Cotton Bowl  Oklahoma         45     Texas
#> 112                              Cotton Bowl     Texas         27  Oklahoma
#>     away_score   winner
#> 1            6    Texas
#> 2           11    Texas
#> 3            6     <NA>
#> 4           10    Texas
#> 5            0 Oklahoma
#> 6           10    Texas
#> 7           10    Texas
#> 8            0 Oklahoma
#> 9            0    Texas
#> 10           3 Oklahoma
#> 11           3 Oklahoma
#> 12           6 Oklahoma
#> 13           6    Texas
#> 14           7    Texas
#> 15          13 Oklahoma
#> 16           7    Texas
#> 17           0 Oklahoma
#> 18           7 Oklahoma
#> 19          32    Texas
#> 20          14    Texas
#> 21           0    Texas
#> 22           7    Texas
#> 23           0    Texas
#> 24          10    Texas
#> 25           0 Oklahoma
#> 26           0    Texas
#> 27           7    Texas
#> 28           0    Texas
#> 29           7     <NA>
#> 30           0 Oklahoma
#> 31          12 Oklahoma
#> 32          16    Texas
#> 33           7    Texas
#> 34           0    Texas
#> 35           7    Texas
#> 36           0    Texas
#> 37           7    Texas
#> 38          13    Texas
#> 39          14    Texas
#> 40          14 Oklahoma
#> 41          14 Oklahoma
#> 42          13 Oklahoma
#> 43           7    Texas
#> 44          20 Oklahoma
#> 45          14 Oklahoma
#> 46           7 Oklahoma
#> 47           0 Oklahoma
#> 48           0 Oklahoma
#> 49           7 Oklahoma
#> 50          14    Texas
#> 51          12    Texas
#> 52           0    Texas
#> 53           7    Texas
#> 54           6    Texas
#> 55           7    Texas
#> 56           7    Texas
#> 57           0    Texas
#> 58           9 Oklahoma
#> 59           7    Texas
#> 60          20    Texas
#> 61          17    Texas
#> 62           9    Texas
#> 63          27 Oklahoma
#> 64           0 Oklahoma
#> 65          13 Oklahoma
#> 66          13 Oklahoma
#> 67          17 Oklahoma
#> 68           6     <NA>
#> 69           6    Texas
#> 70          10 Oklahoma
#> 71           7    Texas
#> 72          13    Texas
#> 73          14    Texas
#> 74          22 Oklahoma
#> 75          16    Texas
#> 76          15     <NA>
#> 77           7 Oklahoma
#> 78          12 Oklahoma
#> 79           9 Oklahoma
#> 80          13 Oklahoma
#> 81          24    Texas
#> 82          13    Texas
#> 83           7    Texas
#> 84          24    Texas
#> 85          17 Oklahoma
#> 86          10    Texas
#> 87          24     <NA>
#> 88          27 Oklahoma
#> 89          24    Texas
#> 90           3    Texas
#> 91          28    Texas
#> 92          14 Oklahoma
#> 93          14 Oklahoma
#> 94          24 Oklahoma
#> 95          65 Oklahoma
#> 96           0 Oklahoma
#> 97          12    Texas
#> 98          28    Texas
#> 99          28 Oklahoma
#> 100         45    Texas
#> 101         13    Texas
#> 102         20 Oklahoma
#> 103         55 Oklahoma
#> 104         21 Oklahoma
#> 105         20    Texas
#> 106         26 Oklahoma
#> 107         17    Texas
#> 108         40 Oklahoma
#> 109         29 Oklahoma
#> 110         27 Oklahoma
#> 111         48    Texas
#> 112         34 Oklahoma

cfb_team_matchup('Texas A&M','TCU')

#>    season week season_type                     date neutral_site venue
#> 1    1903    5     regular 1903-10-10T00:00:00.000Z        FALSE    NA
#> 2    1903   10     regular 1903-11-14T00:00:00.000Z        FALSE    NA
#> 3    1903   12     regular 1903-11-28T00:00:00.000Z        FALSE    NA
#> 4    1904    6     regular 1904-10-22T00:00:00.000Z        FALSE    NA
#> 5    1905    8     regular 1905-11-04T00:00:00.000Z        FALSE    NA
#> 6    1905    3     regular 1905-09-30T00:00:00.000Z        FALSE    NA
#> 7    1906    8     regular 1906-11-10T00:00:00.000Z        FALSE    NA
#> 8    1906    6     regular 1906-10-27T00:00:00.000Z        FALSE    NA
#> 9    1907    8     regular 1907-11-05T00:00:00.000Z        FALSE    NA
#> 10   1908    7     regular 1908-10-31T00:00:00.000Z        FALSE    NA
#> 11   1909    5     regular 1909-10-16T00:00:00.000Z        FALSE    NA
#> 12   1910    8     regular 1910-11-05T00:00:00.000Z        FALSE    NA
#> 13   1910    4     regular 1910-10-08T00:00:00.000Z        FALSE    NA
#> 14   1914    5     regular 1914-10-17T00:00:00.000Z        FALSE    NA
#> 15   1915    5     regular 1915-10-16T00:00:00.000Z        FALSE    NA
#> 16   1919    8     regular 1919-11-08T00:00:00.000Z        FALSE    NA
#> 17   1924    8     regular 1924-11-08T00:00:00.000Z        FALSE    NA
#> 18   1925    8     regular 1925-11-07T00:00:00.000Z        FALSE    NA
#> 19   1926    8     regular 1926-11-06T00:00:00.000Z        FALSE    NA
#> 20   1927    6     regular 1927-10-22T00:00:00.000Z        FALSE    NA
#> 21   1928    7     regular 1928-10-20T00:00:00.000Z        FALSE    NA
#> 22   1929    5     regular 1929-10-19T00:00:00.000Z        FALSE    NA
#> 23   1930    6     regular 1930-10-18T00:00:00.000Z        FALSE    NA
#> 24   1931    6     regular 1931-10-17T00:00:00.000Z        FALSE    NA
#> 25   1932    5     regular 1932-10-15T00:00:00.000Z        FALSE    NA
#> 26   1933    6     regular 1933-10-21T00:00:00.000Z        FALSE    NA
#> 27   1934    6     regular 1934-10-20T00:00:00.000Z        FALSE    NA
#> 28   1935    5     regular 1935-10-19T00:00:00.000Z        FALSE    NA
#> 29   1936    5     regular 1936-10-17T00:00:00.000Z        FALSE    NA
#> 30   1937    5     regular 1937-10-16T00:00:00.000Z        FALSE    NA
#> 31   1938    5     regular 1938-10-15T00:00:00.000Z        FALSE    NA
#> 32   1939    6     regular 1939-10-21T00:00:00.000Z        FALSE    NA
#> 33   1940    6     regular 1940-10-19T00:00:00.000Z        FALSE    NA
#> 34   1941    5     regular 1941-10-18T00:00:00.000Z        FALSE    NA
#> 35   1942    6     regular 1942-10-17T00:00:00.000Z        FALSE    NA
#> 36   1943    6     regular 1943-10-16T00:00:00.000Z        FALSE    NA
#> 37   1944    9     regular 1944-10-21T00:00:00.000Z        FALSE    NA
#> 38   1945    7     regular 1945-10-20T00:00:00.000Z        FALSE    NA
#> 39   1946    6     regular 1946-10-19T00:00:00.000Z        FALSE    NA
#> 40   1947    6     regular 1947-10-18T00:00:00.000Z        FALSE    NA
#> 41   1948    6     regular 1948-10-16T00:00:00.000Z        FALSE    NA
#> 42   1949    6     regular 1949-10-15T00:00:00.000Z        FALSE    NA
#> 43   1950    6     regular 1950-10-21T00:00:00.000Z        FALSE    NA
#> 44   1951    6     regular 1951-10-20T00:00:00.000Z        FALSE    NA
#> 45   1952    6     regular 1952-10-18T00:00:00.000Z        FALSE    NA
#> 46   1953    5     regular 1953-10-17T00:00:00.000Z        FALSE    NA
#> 47   1954    5     regular 1954-10-16T00:00:00.000Z        FALSE    NA
#> 48   1955    5     regular 1955-10-15T00:00:00.000Z        FALSE    NA
#> 49   1956    6     regular 1956-10-20T00:00:00.000Z        FALSE    NA
#> 50   1957    6     regular 1957-10-19T00:00:00.000Z        FALSE    NA
#> 51   1958    6     regular 1958-10-18T00:00:00.000Z        FALSE    NA
#> 52   1959    6     regular 1959-10-17T00:00:00.000Z        FALSE    NA
#> 53   1960    6     regular 1960-10-15T00:00:00.000Z        FALSE    NA
#> 54   1961    6     regular 1961-10-21T00:00:00.000Z        FALSE    NA
#> 55   1962    6     regular 1962-10-20T00:00:00.000Z        FALSE    NA
#> 56   1963    6     regular 1963-10-19T00:00:00.000Z        FALSE    NA
#> 57   1964    6     regular 1964-10-17T00:00:00.000Z        FALSE    NA
#> 58   1965    6     regular 1965-10-16T00:00:00.000Z        FALSE    NA
#> 59   1966    6     regular 1966-10-15T00:00:00.000Z        FALSE    NA
#> 60   1967    7     regular 1967-10-21T00:00:00.000Z        FALSE    NA
#> 61   1968    6     regular 1968-10-19T00:00:00.000Z        FALSE    NA
#> 62   1969    6     regular 1969-10-18T00:00:00.000Z        FALSE    NA
#> 63   1970    8     regular 1970-10-17T00:00:00.000Z        FALSE    NA
#> 64   1971    6     regular 1971-10-16T00:00:00.000Z        FALSE    NA
#> 65   1972    7     regular 1972-10-21T00:00:00.000Z        FALSE    NA
#> 66   1973    8     regular 1973-10-20T00:00:00.000Z        FALSE    NA
#> 67   1974    7     regular 1974-10-19T00:00:00.000Z        FALSE    NA
#> 68   1975    7     regular 1975-10-18T00:00:00.000Z        FALSE    NA
#> 69   1976   12     regular 1976-11-20T00:00:00.000Z        FALSE    NA
#> 70   1977   12     regular 1977-11-19T00:00:00.000Z        FALSE    NA
#> 71   1978   13     regular 1978-11-25T00:00:00.000Z        FALSE    NA
#> 72   1979   13     regular 1979-11-24T00:00:00.000Z        FALSE    NA
#> 73   1980   13     regular 1980-11-22T00:00:00.000Z        FALSE    NA
#> 74   1981   12     regular 1981-11-21T00:00:00.000Z        FALSE    NA
#> 75   1982   12     regular 1982-11-20T00:00:00.000Z        FALSE    NA
#> 76   1983   13     regular 1983-11-19T00:00:00.000Z        FALSE    NA
#> 77   1984   14     regular 1984-11-24T00:00:00.000Z        FALSE    NA
#> 78   1985   13     regular 1985-11-23T00:00:00.000Z        FALSE    NA
#> 79   1986   13     regular 1986-11-22T00:00:00.000Z        FALSE    NA
#> 80   1987   13     regular 1987-11-21T00:00:00.000Z        FALSE    NA
#> 81   1988   13     regular 1988-11-19T00:00:00.000Z        FALSE    NA
#> 82   1989    3     regular 1989-09-16T00:00:00.000Z        FALSE    NA
#> 83   1990   14     regular 1990-11-24T00:00:00.000Z        FALSE    NA
#> 84   1991   11     regular 1991-11-09T00:00:00.000Z        FALSE    NA
#> 85   1992   13     regular 1992-11-21T00:00:00.000Z        FALSE    NA
#> 86   1993   13     regular 1993-11-20T00:00:00.000Z        FALSE    NA
#> 87   1994   13     regular 1994-11-19T00:00:00.000Z        FALSE    NA
#> 88   1995   14     regular 1995-11-25T00:00:00.000Z        FALSE    NA
#> 89   2001    1  postseason 2001-12-28T13:30:00.000Z        FALSE    NA
#>    home_team home_score away_team away_score    winner
#> 1  Texas A&M         11       TCU          0 Texas A&M
#> 2  Texas A&M         16       TCU          0 Texas A&M
#> 3  Texas A&M         14       TCU          6 Texas A&M
#> 4  Texas A&M         29       TCU          0 Texas A&M
#> 5  Texas A&M         24       TCU         11 Texas A&M
#> 6  Texas A&M         20       TCU          0 Texas A&M
#> 7  Texas A&M         22       TCU          0 Texas A&M
#> 8  Texas A&M         42       TCU          0 Texas A&M
#> 9  Texas A&M         32       TCU          5 Texas A&M
#> 10 Texas A&M         13       TCU         10 Texas A&M
#> 11       TCU          0 Texas A&M          0      <NA>
#> 12       TCU          6 Texas A&M         23 Texas A&M
#> 13 Texas A&M         35       TCU          0 Texas A&M
#> 14 Texas A&M         40       TCU          0 Texas A&M
#> 15       TCU         10 Texas A&M         13 Texas A&M
#> 16 Texas A&M         48       TCU          0 Texas A&M
#> 17 Texas A&M         28       TCU          0 Texas A&M
#> 18       TCU          3 Texas A&M          0       TCU
#> 19       TCU         13 Texas A&M         13      <NA>
#> 20 Texas A&M          0       TCU          0      <NA>
#> 21 Texas A&M          0       TCU          6       TCU
#> 22       TCU         13 Texas A&M          7       TCU
#> 23 Texas A&M          0       TCU          3       TCU
#> 24       TCU          6 Texas A&M          0       TCU
#> 25 Texas A&M          0       TCU         17       TCU
#> 26       TCU         13 Texas A&M          7       TCU
#> 27 Texas A&M          0       TCU         13       TCU
#> 28       TCU         19 Texas A&M         14       TCU
#> 29 Texas A&M         18       TCU          7 Texas A&M
#> 30 Texas A&M          7       TCU          7      <NA>
#> 31 Texas A&M          6       TCU         34       TCU
#> 32       TCU          6 Texas A&M         20 Texas A&M
#> 33 Texas A&M         21       TCU          7 Texas A&M
#> 34       TCU          0 Texas A&M         14 Texas A&M
#> 35 Texas A&M          2       TCU          7       TCU
#> 36       TCU          0 Texas A&M         13 Texas A&M
#> 37 Texas A&M          7       TCU         13       TCU
#> 38       TCU         13 Texas A&M         12       TCU
#> 39 Texas A&M         14       TCU          0 Texas A&M
#> 40       TCU         26 Texas A&M          0       TCU
#> 41 Texas A&M         14       TCU         27       TCU
#> 42       TCU         28 Texas A&M          6       TCU
#> 43 Texas A&M         42       TCU         23 Texas A&M
#> 44       TCU         20 Texas A&M         14       TCU
#> 45       TCU          7 Texas A&M          7      <NA>
#> 46       TCU          7 Texas A&M         20 Texas A&M
#> 47 Texas A&M         20       TCU         21       TCU
#> 48       TCU         16 Texas A&M         19 Texas A&M
#> 49 Texas A&M          7       TCU          6 Texas A&M
#> 50       TCU          0 Texas A&M          7 Texas A&M
#> 51 Texas A&M          8       TCU         24       TCU
#> 52       TCU         39 Texas A&M          6       TCU
#> 53       TCU         14 Texas A&M         14      <NA>
#> 54       TCU         15 Texas A&M         14       TCU
#> 55 Texas A&M         14       TCU         20       TCU
#> 56 Texas A&M         14       TCU         14      <NA>
#> 57 Texas A&M          9       TCU         14       TCU
#> 58       TCU         17 Texas A&M          9       TCU
#> 59 Texas A&M         35       TCU          7 Texas A&M
#> 60       TCU          0 Texas A&M         20 Texas A&M
#> 61 Texas A&M         27       TCU          7 Texas A&M
#> 62       TCU         16 Texas A&M          6       TCU
#> 63 Texas A&M         15       TCU         31       TCU
#> 64       TCU         14 Texas A&M          3       TCU
#> 65 Texas A&M         10       TCU         13       TCU
#> 66       TCU         16 Texas A&M         35 Texas A&M
#> 67 Texas A&M         17       TCU          0 Texas A&M
#> 68       TCU          6 Texas A&M         14 Texas A&M
#> 69 Texas A&M         59       TCU         10 Texas A&M
#> 70       TCU         23 Texas A&M         52 Texas A&M
#> 71 Texas A&M         15       TCU          7 Texas A&M
#> 72       TCU          7 Texas A&M         30 Texas A&M
#> 73 Texas A&M         13       TCU         10 Texas A&M
#> 74       TCU          7 Texas A&M         37 Texas A&M
#> 75 Texas A&M         34       TCU         14 Texas A&M
#> 76       TCU         10 Texas A&M         20 Texas A&M
#> 77 Texas A&M         35       TCU         21 Texas A&M
#> 78       TCU          6 Texas A&M         53 Texas A&M
#> 79 Texas A&M         74       TCU         10 Texas A&M
#> 80       TCU         24 Texas A&M         42 Texas A&M
#> 81 Texas A&M         18       TCU          0 Texas A&M
#> 82       TCU          7 Texas A&M         44 Texas A&M
#> 83 Texas A&M         56       TCU         10 Texas A&M
#> 84       TCU          7 Texas A&M         44 Texas A&M
#> 85 Texas A&M         37       TCU         10 Texas A&M
#> 86       TCU          3 Texas A&M         59 Texas A&M
#> 87 Texas A&M         34       TCU         17 Texas A&M
#> 88       TCU          6 Texas A&M         38 Texas A&M
#> 89 Texas A&M         28       TCU          9 Texas A&M

cfb_team_matchup('Texas A&M','TCU', min_year = 1975)

#>    season week season_type                     date neutral_site venue
#> 1    1975    7     regular 1975-10-18T00:00:00.000Z        FALSE    NA
#> 2    1976   12     regular 1976-11-20T00:00:00.000Z        FALSE    NA
#> 3    1977   12     regular 1977-11-19T00:00:00.000Z        FALSE    NA
#> 4    1978   13     regular 1978-11-25T00:00:00.000Z        FALSE    NA
#> 5    1979   13     regular 1979-11-24T00:00:00.000Z        FALSE    NA
#> 6    1980   13     regular 1980-11-22T00:00:00.000Z        FALSE    NA
#> 7    1981   12     regular 1981-11-21T00:00:00.000Z        FALSE    NA
#> 8    1982   12     regular 1982-11-20T00:00:00.000Z        FALSE    NA
#> 9    1983   13     regular 1983-11-19T00:00:00.000Z        FALSE    NA
#> 10   1984   14     regular 1984-11-24T00:00:00.000Z        FALSE    NA
#> 11   1985   13     regular 1985-11-23T00:00:00.000Z        FALSE    NA
#> 12   1986   13     regular 1986-11-22T00:00:00.000Z        FALSE    NA
#> 13   1987   13     regular 1987-11-21T00:00:00.000Z        FALSE    NA
#> 14   1988   13     regular 1988-11-19T00:00:00.000Z        FALSE    NA
#> 15   1989    3     regular 1989-09-16T00:00:00.000Z        FALSE    NA
#> 16   1990   14     regular 1990-11-24T00:00:00.000Z        FALSE    NA
#> 17   1991   11     regular 1991-11-09T00:00:00.000Z        FALSE    NA
#> 18   1992   13     regular 1992-11-21T00:00:00.000Z        FALSE    NA
#> 19   1993   13     regular 1993-11-20T00:00:00.000Z        FALSE    NA
#> 20   1994   13     regular 1994-11-19T00:00:00.000Z        FALSE    NA
#> 21   1995   14     regular 1995-11-25T00:00:00.000Z        FALSE    NA
#> 22   2001    1  postseason 2001-12-28T13:30:00.000Z        FALSE    NA
#>    home_team home_score away_team away_score    winner
#> 1        TCU          6 Texas A&M         14 Texas A&M
#> 2  Texas A&M         59       TCU         10 Texas A&M
#> 3        TCU         23 Texas A&M         52 Texas A&M
#> 4  Texas A&M         15       TCU          7 Texas A&M
#> 5        TCU          7 Texas A&M         30 Texas A&M
#> 6  Texas A&M         13       TCU         10 Texas A&M
#> 7        TCU          7 Texas A&M         37 Texas A&M
#> 8  Texas A&M         34       TCU         14 Texas A&M
#> 9        TCU         10 Texas A&M         20 Texas A&M
#> 10 Texas A&M         35       TCU         21 Texas A&M
#> 11       TCU          6 Texas A&M         53 Texas A&M
#> 12 Texas A&M         74       TCU         10 Texas A&M
#> 13       TCU         24 Texas A&M         42 Texas A&M
#> 14 Texas A&M         18       TCU          0 Texas A&M
#> 15       TCU          7 Texas A&M         44 Texas A&M
#> 16 Texas A&M         56       TCU         10 Texas A&M
#> 17       TCU          7 Texas A&M         44 Texas A&M
#> 18 Texas A&M         37       TCU         10 Texas A&M
#> 19       TCU          3 Texas A&M         59 Texas A&M
#> 20 Texas A&M         34       TCU         17 Texas A&M
#> 21       TCU          6 Texas A&M         38 Texas A&M
#> 22 Texas A&M         28       TCU          9 Texas A&M

cfb_team_matchup('Florida State', 'Florida', min_year = 1975)

#>    season week season_type                     date neutral_site
#> 1    1975    7     regular 1975-10-18T00:00:00.000Z        FALSE
#> 2    1976    7     regular 1976-10-16T00:00:00.000Z        FALSE
#> 3    1977   14     regular 1977-12-03T00:00:00.000Z        FALSE
#> 4    1978   13     regular 1978-11-25T00:00:00.000Z        FALSE
#> 5    1979   13     regular 1979-11-24T00:00:00.000Z        FALSE
#> 6    1980   15     regular 1980-12-06T00:00:00.000Z        FALSE
#> 7    1981   13     regular 1981-11-28T00:00:00.000Z        FALSE
#> 8    1982   14     regular 1982-12-04T00:00:00.000Z        FALSE
#> 9    1983   15     regular 1983-12-03T00:00:00.000Z        FALSE
#> 10   1984   15     regular 1984-12-01T00:00:00.000Z        FALSE
#> 11   1985   14     regular 1985-11-30T00:00:00.000Z        FALSE
#> 12   1986   14     regular 1986-11-29T00:00:00.000Z        FALSE
#> 13   1987   14     regular 1987-11-28T00:00:00.000Z        FALSE
#> 14   1988   14     regular 1988-11-26T00:00:00.000Z        FALSE
#> 15   1989   14     regular 1989-12-02T00:00:00.000Z        FALSE
#> 16   1990   15     regular 1990-12-01T00:00:00.000Z        FALSE
#> 17   1991   14     regular 1991-11-30T00:00:00.000Z        FALSE
#> 18   1992   14     regular 1992-11-28T00:00:00.000Z        FALSE
#> 19   1993   14     regular 1993-11-27T00:00:00.000Z        FALSE
#> 20   1994   14     regular 1994-11-26T00:00:00.000Z        FALSE
#> 21   1994    1  postseason 1995-01-02T00:00:00.000Z         TRUE
#> 22   1995   14     regular 1995-11-25T00:00:00.000Z        FALSE
#> 23   1996   15     regular 1996-11-30T00:00:00.000Z        FALSE
#> 24   1996    1  postseason 1997-01-02T00:00:00.000Z         TRUE
#> 25   1997   14     regular 1997-11-22T00:00:00.000Z        FALSE
#> 26   1998   13     regular 1998-11-21T00:00:00.000Z        FALSE
#> 27   1999   13     regular 1999-11-20T00:00:00.000Z        FALSE
#> 28   2000   13     regular 2000-11-18T00:00:00.000Z        FALSE
#> 29   2001   13     regular 2001-11-17T20:00:00.000Z        FALSE
#> 30   2002   15     regular 2002-11-30T20:00:00.000Z        FALSE
#> 31   2003   15     regular 2003-11-29T15:30:00.000Z        FALSE
#> 32   2004   13     regular 2004-11-20T19:45:00.000Z        FALSE
#> 33   2005   13     regular 2005-11-26T15:30:00.000Z        FALSE
#> 34   2006   13     regular 2006-11-25T12:00:00.000Z        FALSE
#> 35   2007   13     regular 2007-11-24T17:00:00.000Z        FALSE
#> 36   2008   14     regular 2008-11-29T15:30:00.000Z        FALSE
#> 37   2009   13     regular 2009-11-28T15:30:00.000Z        FALSE
#> 38   2010   13     regular 2010-11-27T15:30:00.000Z        FALSE
#> 39   2011   13     regular 2011-11-26T19:00:00.000Z        FALSE
#> 40   2012   13     regular 2012-11-24T15:30:00.000Z        FALSE
#> 41   2013   14     regular 2013-11-30T12:00:00.000Z        FALSE
#> 42   2014   14     regular 2014-11-29T15:30:00.000Z        FALSE
#> 43   2015   13     regular 2015-11-28T19:30:00.000Z        FALSE
#> 44   2016   13     regular 2016-11-26T20:00:00.000Z        FALSE
#> 45   2017   13     regular 2017-11-25T12:00:00.000Z        FALSE
#> 46   2018   13     regular 2018-11-24T17:00:00.000Z        FALSE
#> 47   2019   14     regular 2019-12-01T00:30:00.000Z        FALSE
#>                                          venue     home_team home_score
#> 1                                         <NA>       Florida         34
#> 2                                         <NA> Florida State         26
#> 3                                         <NA>       Florida          9
#> 4                                         <NA> Florida State         38
#> 5                                         <NA>       Florida         16
#> 6                                         <NA> Florida State         17
#> 7                                         <NA>       Florida         35
#> 8                                         <NA> Florida State         10
#> 9                                         <NA>       Florida         53
#> 10                                        <NA> Florida State         17
#> 11                                        <NA>       Florida         38
#> 12                                        <NA> Florida State         13
#> 13                                        <NA>       Florida         14
#> 14                                        <NA> Florida State         52
#> 15                                        <NA>       Florida         17
#> 16                                        <NA> Florida State         45
#> 17                                        <NA>       Florida         14
#> 18                                        <NA> Florida State         45
#> 19                                        <NA>       Florida         21
#> 20                                        <NA>       Florida         31
#> 21                                        <NA> Florida State         23
#> 22                                        <NA>       Florida         35
#> 23                                        <NA> Florida State         24
#> 24                                        <NA>       Florida         52
#> 25                                        <NA>       Florida         32
#> 26                                        <NA> Florida State         23
#> 27                                        <NA>       Florida         23
#> 28                                        <NA> Florida State         30
#> 29                                        <NA>       Florida         37
#> 30 Bobby Bowden Field at Doak Campbell Stadium Florida State         31
#> 31                                        <NA>       Florida         34
#> 32 Bobby Bowden Field at Doak Campbell Stadium Florida State         13
#> 33                    Ben Hill Griffin Stadium       Florida         34
#> 34 Bobby Bowden Field at Doak Campbell Stadium Florida State         14
#> 35                    Ben Hill Griffin Stadium       Florida         45
#> 36 Bobby Bowden Field at Doak Campbell Stadium Florida State         15
#> 37                    Ben Hill Griffin Stadium       Florida         37
#> 38 Bobby Bowden Field at Doak Campbell Stadium Florida State         31
#> 39                    Ben Hill Griffin Stadium       Florida          7
#> 40 Bobby Bowden Field at Doak Campbell Stadium Florida State         26
#> 41                    Ben Hill Griffin Stadium       Florida          7
#> 42 Bobby Bowden Field at Doak Campbell Stadium Florida State         24
#> 43                    Ben Hill Griffin Stadium       Florida          2
#> 44 Bobby Bowden Field at Doak Campbell Stadium Florida State         31
#> 45                    Ben Hill Griffin Stadium       Florida         22
#> 46 Bobby Bowden Field at Doak Campbell Stadium Florida State         14
#> 47                    Ben Hill Griffin Stadium       Florida         40
#>        away_team away_score        winner
#> 1  Florida State          8       Florida
#> 2        Florida         33       Florida
#> 3  Florida State         37 Florida State
#> 4        Florida         21 Florida State
#> 5  Florida State         27 Florida State
#> 6        Florida         13 Florida State
#> 7  Florida State          3       Florida
#> 8        Florida         13       Florida
#> 9  Florida State         14       Florida
#> 10       Florida         27       Florida
#> 11 Florida State         14       Florida
#> 12       Florida         17       Florida
#> 13 Florida State         28 Florida State
#> 14       Florida         17 Florida State
#> 15 Florida State         24 Florida State
#> 16       Florida         30 Florida State
#> 17 Florida State          9       Florida
#> 18       Florida         24 Florida State
#> 19 Florida State         33 Florida State
#> 20 Florida State         31          <NA>
#> 21       Florida         17 Florida State
#> 22 Florida State         24       Florida
#> 23       Florida         21 Florida State
#> 24 Florida State         20       Florida
#> 25 Florida State         29       Florida
#> 26       Florida         12 Florida State
#> 27 Florida State         30 Florida State
#> 28       Florida          7 Florida State
#> 29 Florida State         13       Florida
#> 30       Florida         14 Florida State
#> 31 Florida State         38 Florida State
#> 32       Florida         20       Florida
#> 33 Florida State          7       Florida
#> 34       Florida         21       Florida
#> 35 Florida State         12       Florida
#> 36       Florida         45       Florida
#> 37 Florida State         10       Florida
#> 38       Florida          7 Florida State
#> 39 Florida State         21 Florida State
#> 40       Florida         37       Florida
#> 41 Florida State         37 Florida State
#> 42       Florida         19 Florida State
#> 43 Florida State         27 Florida State
#> 44       Florida         13 Florida State
#> 45 Florida State         38 Florida State
#> 46       Florida         41       Florida
#> 47 Florida State         17       Florida