Coach Information Search

A coach search function which provides coaching records and school history for a given coach

cfb_coaches(
  first = NULL,
  last = NULL,
  team = NULL,
  year = NULL,
  min_year = NULL,
  max_year = NULL
)

Arguments

first
(String optional): First name for the coach you are trying to look up

last
(String optional): Last name for the coach you are trying to look up

team
(String optional): Team - Select a valid team, D1 football

year
(Integer optional): Year, 4 digit format (YYYY).

min_year
(Integer optional): Minimum Year filter (inclusive), 4 digit format (YYYY).

max_year
(Integer optional): Maximum Year filter (inclusive), 4 digit format (YYYY).

Examples

cfb_coaches(first = "Nick", last = "Saban", team = 'alabama')

#> 2020-09-04 13:09:07: Scraping coaches data...

#>    first_name  last_name  school year games wins losses ties preseason_rank
#> 1     Wallace       Wade Alabama 1923    10    7      2    1             NA
#> 2     Wallace       Wade Alabama 1924     9    8      1    0             NA
#> 3     Wallace       Wade Alabama 1925    10   10      0    0             NA
#> 4     Wallace       Wade Alabama 1926    10    9      0    1             NA
#> 5     Wallace       Wade Alabama 1927    10    5      4    1             NA
#> 6     Wallace       Wade Alabama 1928     9    6      3    0             NA
#> 7     Wallace       Wade Alabama 1929     9    6      3    0             NA
#> 8     Wallace       Wade Alabama 1930    10   10      0    0             NA
#> 9       Frank     Thomas Alabama 1931    10    9      1    0             NA
#> 10      Frank     Thomas Alabama 1932    10    8      2    0             NA
#> 11      Frank     Thomas Alabama 1933     9    7      1    1             NA
#> 12      Frank     Thomas Alabama 1934    10   10      0    0             NA
#> 13      Frank     Thomas Alabama 1935     9    6      2    1             NA
#> 14      Frank     Thomas Alabama 1936     9    8      0    1             NA
#> 15      Frank     Thomas Alabama 1937    10    9      1    0             NA
#> 16      Frank     Thomas Alabama 1938     9    7      1    1             NA
#> 17      Frank     Thomas Alabama 1939     9    5      3    1             NA
#> 18      Frank     Thomas Alabama 1940     9    7      2    0             NA
#> 19      Frank     Thomas Alabama 1941    11    9      2    0             NA
#> 20      Frank     Thomas Alabama 1942    11    8      3    0             NA
#> 21      Frank     Thomas Alabama 1944     9    5      2    2             NA
#> 22      Frank     Thomas Alabama 1945    10   10      0    0             NA
#> 23      Frank     Thomas Alabama 1946    11    7      4    0             NA
#> 24        Red       Drew Alabama 1947    11    8      3    0             NA
#> 25        Red       Drew Alabama 1948    11    6      4    1             NA
#> 26        Red       Drew Alabama 1949    10    6      3    1             NA
#> 27        Red       Drew Alabama 1950    11    9      2    0             NA
#> 28        Red       Drew Alabama 1951    11    5      6    0              9
#> 29        Red       Drew Alabama 1952    12   10      2    0             NA
#> 30        Red       Drew Alabama 1953    12    6      3    3              5
#> 31        Red       Drew Alabama 1954    11    4      5    2             14
#> 32       Bear     Bryant Alabama 1958    10    5      4    1             NA
#> 33       Bear     Bryant Alabama 1959    11    7      2    2             NA
#> 34       Bear     Bryant Alabama 1960    11    8      1    2             NA
#> 35       Bear     Bryant Alabama 1961    11   11      0    0              3
#> 36       Bear     Bryant Alabama 1962    11   10      1    0              3
#> 37       Bear     Bryant Alabama 1963    11    9      2    0              3
#> 38       Bear     Bryant Alabama 1964    11   10      1    0              6
#> 39       Bear     Bryant Alabama 1965    11    9      1    1              5
#> 40       Bear     Bryant Alabama 1966    11   11      0    0              1
#> 41       Bear     Bryant Alabama 1967    11    8      2    1              2
#> 42       Bear     Bryant Alabama 1968    11    8      3    0             10
#> 43       Bear     Bryant Alabama 1969    11    6      5    0             13
#> 44       Bear     Bryant Alabama 1970    12    6      5    1             16
#> 45       Bear     Bryant Alabama 1971    12   11      1    0             16
#> 46       Bear     Bryant Alabama 1972    12   10      2    0              7
#> 47       Bear     Bryant Alabama 1973    12   11      1    0              6
#> 48       Bear     Bryant Alabama 1974    12   11      1    0              4
#> 49       Bear     Bryant Alabama 1975    12   11      1    0              2
#> 50       Bear     Bryant Alabama 1976    12    9      3    0              6
#> 51       Bear     Bryant Alabama 1977    12   11      1    0              6
#> 52       Bear     Bryant Alabama 1978    12   11      1    0              1
#> 53       Bear     Bryant Alabama 1979    12   12      0    0              2
#> 54       Bear     Bryant Alabama 1980    12   10      2    0              2
#> 55       Bear     Bryant Alabama 1981    12    9      2    1              4
#> 56       Bear     Bryant Alabama 1982    12    8      4    0              3
#> 57       Bill      Curry Alabama 1987    12    7      5    0             NA
#> 58       Bill      Curry Alabama 1988    12    9      3    0             14
#> 59       Bill      Curry Alabama 1989    12   10      2    0             16
#> 60       Gene  Stallings Alabama 1990    12    7      5    0             12
#> 61       Gene  Stallings Alabama 1991    12   11      1    0             22
#> 62       Gene  Stallings Alabama 1992    13   13      0    0              9
#> 63       Gene  Stallings Alabama 1993    13    9      3    1              2
#> 64       Gene  Stallings Alabama 1994    13   12      1    0             12
#> 65       Gene  Stallings Alabama 1995    11    8      3    0             10
#> 66       Gene  Stallings Alabama 1996    13   10      3    0             15
#> 67     Dennis Franchione Alabama 2001    12    7      5    0             25
#> 68     Dennis Franchione Alabama 2002    13   10      3    0             NA
#> 69       Nick      Saban Alabama 2007    13    7      6    0             NA
#> 70       Nick      Saban Alabama 2008    14   12      2    0             24
#> 71       Nick      Saban Alabama 2009    14   14      0    0              5
#> 72       Nick      Saban Alabama 2010    13   10      3    0              1
#> 73       Nick      Saban Alabama 2011    13   12      1    0              2
#> 74       Nick      Saban Alabama 2012    14   13      1    0              2
#> 75       Nick      Saban Alabama 2013    13   11      2    0              1
#> 76       Nick      Saban Alabama 2014    14   12      2    0              2
#> 77       Nick      Saban Alabama 2015    15   14      1    0              3
#> 78       Nick      Saban Alabama 2016    15   14      1    0              1
#> 79       Nick      Saban Alabama 2017    14   13      1    0              1
#> 80       Nick      Saban Alabama 2018    15   14      1    0              1
#> 81       Nick      Saban Alabama 2019    13   11      2    0              2
#> 82       Nick      Saban Alabama 2020     0    0      0    0              3
#>    postseason_rank
#> 1               NA
#> 2               NA
#> 3               NA
#> 4               NA
#> 5               NA
#> 6               NA
#> 7               NA
#> 8               NA
#> 9               NA
#> 10              NA
#> 11              NA
#> 12              NA
#> 13              NA
#> 14               4
#> 15               4
#> 16              13
#> 17              NA
#> 18              NA
#> 19              20
#> 20              10
#> 21              NA
#> 22               2
#> 23              NA
#> 24               6
#> 25              NA
#> 26              NA
#> 27              16
#> 28              NA
#> 29               9
#> 30              13
#> 31              NA
#> 32              NA
#> 33              10
#> 34               9
#> 35               1
#> 36               5
#> 37               8
#> 38               1
#> 39               1
#> 40               3
#> 41               8
#> 42              17
#> 43              NA
#> 44              NA
#> 45               4
#> 46               7
#> 47               4
#> 48               5
#> 49               3
#> 50              11
#> 51               2
#> 52               1
#> 53               1
#> 54               6
#> 55               7
#> 56              NA
#> 57              NA
#> 58              17
#> 59               9
#> 60              NA
#> 61               5
#> 62               1
#> 63              14
#> 64               5
#> 65              21
#> 66              11
#> 67              NA
#> 68              11
#> 69              NA
#> 70               6
#> 71               1
#> 72              10
#> 73               1
#> 74               1
#> 75               7
#> 76               4
#> 77               1
#> 78               2
#> 79               1
#> 80              NA
#> 81               8
#> 82              NA