Get player game averages for Predicted Points Added (PPA)

Get player game averages for Predicted Points Added (PPA)

cfb_metrics_ppa_players_games(
  year = NULL,
  week = NULL,
  team = NULL,
  position = NULL,
  athlete_id = NULL,
  threshold = NULL,
  excl_garbage_time = FALSE
)

Arguments

year
(Integer required): Year, 4 digit format (YYYY)

week
(Integer optional): Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier

team
(String optional): D-I Team

position
(string optional): Position abbreviation of the player you are searching for.
Position Group - options include:
* Offense: QB, RB, FB, TE, OL, G, OT, C, WR
* Defense: DB, CB, S, LB, DE, DT, NT, DL
* Special Teams: K, P, LS, PK

athlete_id
(Integer optional): Athlete ID filter for querying a single athlete
Can be found using the cfbscrapR::cfb_player_info() function.

threshold
(Integer optional): Minimum threshold of plays.

excl_garbage_time
(Logical default FALSE): Select whether to exclude Garbage Time (TRUE or FALSE)

Source

https://api.collegefootballdata.com/ppa/players/games

Value

A data frame with 9 variables:

season

integer.

week

integer.

name

character.

position

character.

team

character.

opponent

character.

avg_PPA_all

double.

avg_PPA_pass

double.

avg_PPA_rush

double.

Examples

cfb_metrics_ppa_players_games(year = 2019,week=3, team = 'TCU')

#> 2020-11-03 01:52:28: Scraping CFBData metrics PPA game-level players data...

#>   season week            name position team opponent avg_PPA_all avg_PPA_pass
#> 1   2019    3     Alex Delton       QB  TCU   Purdue      -0.861       -1.052
#> 2   2019    3      Max Duggan       QB  TCU   Purdue      -0.184       -0.234
#> 3   2019    3 Emari Demercado       RB  TCU   Purdue      -0.080           NA
#> 4   2019    3    Jalen Reagor       WR  TCU   Purdue       0.008        0.164
#> 5   2019    3       Pro Wells       TE  TCU   Purdue       0.129        0.129
#> 6   2019    3   Sewo Olonilua       RB  TCU   Purdue       0.378       -0.471
#> 7   2019    3 Darius Anderson       RB  TCU   Purdue       0.629       -0.451
#> 8   2019    3 Al'Dontre Davis       WR  TCU   Purdue       3.183        3.183
#>   avg_PPA_rush
#> 1       -0.192
#> 2       -0.085
#> 3       -0.080
#> 4       -0.201
#> 5           NA
#> 6        0.478
#> 7        0.764
#> 8           NA