Team Roster Get a teams full roster by year. If year not selected, API defaults to most recent year (2019 as of 6/23/20)

Team Roster Get a teams full roster by year. If year not selected, API defaults to most recent year (2019 as of 6/23/20)

cfb_team_roster(team, year = NULL)

Arguments

team
(String required): Team, select a valid team in D-I football

year
(Integer optional): Year, 4 digit format (YYYY)

Examples

cfb_team_roster("Florida State")

#>          id first_name    last_name weight height jersey year position
#> 1   4240036    Leonard   Warner III    241     76     35    3       LB
#> 2   4426396    Brendan         Gant    198     74     44    1       DB
#> 3   4273843      Grant      Glennon    218     75     44    2       LS
#> 4   4363032   Keyshawn       Helton    170     69     20    2       WR
#> 5   4426817     Jarvis     Brownlee    165     71     19    1       DB
#> 6   4240023       Tre'      McKitty    245     77      6    3       TE
#> 7   4363030     Warren     Thompson    201     75     11    1       WR
#> 8   4363049    Malcolm        Lamar    277     77     97    1       DE
#> 9   4240022     Khalan       Laborn    205     71      4    2       RB
#> 10  3886609       Alex   Hornibrook    220     76     12    4       QB
#> 11  4428783   Treshaun         Ward    194     70     38    1       RB
#> 12  4363041     Xavier       Peters    235     76      5    1       DE
#> 13  4240033    Ontaria       Wilson    170     72     80    2       WR
#> 14  4035619       Josh        Brown    229     75     51    4       LB
#> 15  3921955       Ryan      Roberts    300     78     56    4       OL
#> 16  4035618     Cedric         Wood    318     75     49    3       DT
#> 17  4240028  Tamorrion        Terry    203     76     15    2       WR
#> 18  4035621     Emmett         Rice    220     74     56    3       LB
#> 19  4035602 Dontavious      Jackson    229     75      5    4       LB
#> 20  4363037       A.J.       Lytton    179     71     12    2       DB
#> 21  4035606      Logan        Tyler    191     72     21    4        P
#> 22  4363045     Camren     McDonald    231     76     87    2       TE
#> 23  4240030     Hamsah  Nasirildeen    215     76     23    3       DB
#> 24  4426427    Raymond   Woodie III    198     72     31    1       DB
#> 25  4427291      Kalen      DeLoach    211     73     20    1       LB
#> 26  4428664        Tru     Thompson    315     72     90    1       DT
#> 27  4426397     Jaleel        McRae    233     74     42    1       LB
#> 28  4427325    Renardo        Green    174     72     36    1       DB
#> 29  4240042       Cory       Durden    305     77     16    2       DT
#> 30  4240021        Cam        Akers    212     71      3    3       RB
#> 31  4360468     Jaiden Lars-Woodbey    223     75      6    2       DB
#> 32  4363047     Robert       Cooper    346     74     91    2       DT
#> 33  4363044     Jordan        Young    192     74     83    1       WR
#> 34  4430153    Derrick  McLendon II    252     76     55    1       DE
#> 35  3692942     George     Campbell    183     76     15    4       WR
#> 36  4363034     Asante   Samuel Jr.    180     70     26    2       DB
#> 37  4240026     Joshua      Kaindoh    261     79     13    3       DE
#> 38  4035605     Carlos   Becker III    200     74     15    3       DB
#> 39  4035613       Kyle       Meyers    174     72     14    4       DB
#> 40  4363040     Dennis   Briggs Jr.    268     76     58    1       DE
#> 41  4360799     Jordan       Travis    208     73     13    1       QB
#> 42  4240031    DeCalon       Brooks    209     70     28    2       LB
#> 43  4035604   Janarius     Robinson    261     77     11    3       DE
#> 44  4035601    Levonta       Taylor    190     70      1    4       DB
#> 45  4035628      Keith        Gavin    212     75     89    4       WR
#> 46  4360469      Amari       Gainer    215     75     33    1       LB
#> 47  4568255      Akeem         Dent    182     73     27    1       DB
#> 48  4240029     Marvin       Wilson    311     77     21    3       DT
#> 49  4240024   Stanford  Samuels III    185     74      8    3       DB
#> 50  4240034       D.J.     Matthews    156     70      7    3       WR
#> 51  4240020      James     Blackman    195     77      1    2       QB
#> 52  4035608      Ricky       Aguayo    215     74     23    4       PK
#> 53  4363036     Isaiah       Bolden    194     74     29    1       DB
#> 54  4240027      Cyrus        Fagan    191     73     24    3       DB
#> 55  4363046  Tre'Shaun     Harrison    191     74     88    2       WR
#> 56  3917228     Tyrell       Moorer    199     71     85    4       WR
#> 57  4360470    Garrett       Murray    240     72     42    1       LS
#> 58  4363028      Nolan     McDonald    180     74     41    1       DB
#> 59  4363042      Jalen         Goss    270     79     61    1       OL
#> 60  4430678    Ricardo       Watson    247     72     54    1       LB
#> 61  4568260  Alexander     Strozier    302     74     62    4       OL
#> 62  4363029       Alex    Eleyssami    190     73     16    1       QB
#> 63  3921570      Abdul        Bello    315     78     75    4       OL
#> 64  4577584     Robert     Elder IV    255     74     69    1       OL
#> 65  4568261        Jay     Williams    300     78     74    3       OL
#> 66  4568258    Maurice        Smith    301     75     53    1       OL
#> 67  4568257      Tyler       Gilroy    235     71     45    2       LB
#> 68  4363027      Caleb         Ward    173     73     13    1       WR
#> 69  4426445    Quashon       Fuller    267     75     30    1       DE
#> 70  4035622     Andrew      Boselli    308     77     60    3       OL
#> 71  4577580       Gino      English    185     73     10    1       QB
#> 72  4577583      Ethan      Umstead    220     73     40    1       LB
#> 73  4066033     Joseph       Garcia    218     74     47    3       LB
#> 74  4066034     Tanner      Adkison    192     70     63    3       LS
#> 75  4363048   Jamarcus      Chatman    271     74     95    1       DE
#> 76  4427288      Kevon        Glenn    236     74     10    1       LB
#> 77  4577585      Caleb        Faris    175     68     81    1       WR
#> 78  4568259       Axel         Rizo    238     74     57    1       LB
#> 79  4363035   D'Marcus        Adams    180     72     28    1       WR
#> 80  4577582    Raekwon         Webb    195     68     37    3       RB
#> 81  4363043       Chaz         Neal    305     79     71    1       OL
#> 82  4568256   DeAundre       Emeric    196     69     39    2       DB
#> 83  4426452     Curtis     Fann Jr.    250     75     98    1       DE
#> 84  4426456     Dontae        Lucas    315     75     55    1       OL
#> 85  4380437     Jeremy     Czerenda    285     75     68    1       OL
#> 86  4363039  Christian      Meadows    336     76     52    1       OL
#> 87  4568263       Ryan   Fitzgerald    183     73     88    1       PK
#> 88  4035625      Jauan     Williams    310     79     73    3       OL
#> 89  4047688    Michael     Barulich    195     71     86    3       WR
#> 90  4240039      Brady        Scott    312     78     59    2       OL
#> 91  4360472  Christian    Armstrong    320     76     77    1       OL
#> 92  4429791        Ira    Henry III    317     76     78    1       OL
#> 93  4426429     Travis          Jay    185     74     18    1       DB
#> 94  4035620     Baveon      Johnson    305     75     51    3       OL
#> 95  4360474      Peter       Osimen    297     74     93    4       DT
#> 96  4577581        Tre        Young    195     72     29    1       WR
#> 97  4568262     Darius   Washington    308     76     76    1       OL
#> 98  4428307    Malcolm          Ray    278     74     99    1       DT
#> 99  4577586     Austin        White    242     77     82    1       TE
#> 100 4259260      Tommy       Martin    191     71     30    2        P
#> 101 3693281     Calvin      Brewton    184     72      4    4       DB
#> 102 4360473    Adarius         Dent    174     73     84    2       WR
#> 103 4035623       Mike       Arnold    340     77     72    3       OL
#> 104 4240035     Deonte    Sheffield    203     70     14    2       RB
#> 105 4262804     Parker     Grothaus    220     74     37    2       PK
#> 106 4035611       Gabe       Nabers    243     75     32    4       TE
#> 107 3925359     Adonis       Thomas    239     76     22    4       LB
#> 108 3921569       Cole      Minshew    330     77     70    4       OL
#> 109 4363038     Armani         Kerr    220     71     48    3       LB
#> 110 3921562     Marcus        Lewis    197     73      8    4       DB
#> 111 4363033    Anthony        Grant    197     71     10    2       RB
#>            home_city home_state home_country
#> 1         Snellville         GA          USA
#> 2           Lakeland         FL          USA
#> 3        Tallahassee         FL          USA
#> 4          Pensacola         FL          USA
#> 5      Miami Gardens         FL          USA
#> 6      Wesley Chapel         FL          USA
#> 7            Seffner         FL          USA
#> 8            Seffner         FL          USA
#> 9     Virginia Beach         VA          USA
#> 10      West Chester         PA          USA
#> 11        Plant City         FL          USA
#> 12      West Chester         OH          USA
#> 13           Ashburn         GA          USA
#> 14         Charlotte         NC          USA
#> 15        Northville         MI          USA
#> 16       Tallahassee         FL          USA
#> 17           Ashburn         GA          USA
#> 18     Miami Gardens         FL          USA
#> 19             Alief         TX          USA
#> 20    Upper Marlboro         MD          USA
#> 21              Nixa         MO          USA
#> 22        Long Beach         CA          USA
#> 23           Concord         NC          USA
#> 24          Palmetto         FL          USA
#> 25          Savannah         GA          USA
#> 26        Loganville         GA          USA
#> 27  New Smyrna Beach         FL          USA
#> 28           Orlando         FL          USA
#> 29          Newberry         FL          USA
#> 30           Clinton         MS          USA
#> 31           Fontana         CA          USA
#> 32           Lilburn         GA          USA
#> 33           Conyers         GA          USA
#> 34           Decatur         GA          USA
#> 35        Clearwater         FL          USA
#> 36           Sunrise         FL          USA
#> 37         Baltimore         MD          USA
#> 38         Kissimmee         FL          USA
#> 39       New Orleans         LA          USA
#> 40         Kissimmee         FL          USA
#> 41   West Palm Beach         FL          USA
#> 42             Tampa         FL          USA
#> 43       Panama City         FL          USA
#> 44    Virginia Beach         VA          USA
#> 45     Crawfordville         FL          USA
#> 46       Tallahassee         FL          USA
#> 47           Pahokee         FL          USA
#> 48           Houston         TX          USA
#> 49    Pembroke Pines         FL          USA
#> 50      Jacksonville         FL          USA
#> 51         South Bay         FL          USA
#> 52          Mascotte         FL          USA
#> 53             Tampa         FL          USA
#> 54     Daytona Beach         FL          USA
#> 55           Seattle         WA          USA
#> 56      Jacksonville         FL          USA
#> 57             Tampa         FL          USA
#> 58        Long Beach         CA          USA
#> 59          Valdosta         GA          USA
#> 60             Tampa         FL          USA
#> 61            Bartow         FL          USA
#> 62           Orlando         FL          USA
#> 63             Warri       <NA>      Nigeria
#> 64            Stuart         FL          USA
#> 65     Moreno Valley         CA          USA
#> 66             Miami         FL          USA
#> 67   Saint Augustine         FL          USA
#> 68         Pensacola         FL          USA
#> 69        Fort Myers         FL          USA
#> 70      Jacksonville         FL          USA
#> 71       Winter Park         FL          USA
#> 72         Lake City         FL          USA
#> 73             Miami         FL          USA
#> 74       Tallahassee         FL          USA
#> 75              Rome         GA          USA
#> 76         Riverdale         GA          USA
#> 77             Tampa         FL          USA
#> 78             Miami         FL          USA
#> 79      Ormond Beach         FL          USA
#> 80       Panama City         FL          USA
#> 81       Zephyrhills         FL          USA
#> 82           Orlando         FL          USA
#> 83         Stillmore         GA          USA
#> 84          Overtown         FL          USA
#> 85   Fort Lauderdale         FL          USA
#> 86         Montezuma         GA          USA
#> 87          Coolidge         GA          USA
#> 88        Washington         DC          USA
#> 89           Orlando         FL          USA
#> 90    Powder Springs         GA          USA
#> 91     Warner Robins         GA          USA
#> 92       Saint Louis         MO          USA
#> 93        Greenville         FL          USA
#> 94          Lakeland         FL          USA
#> 95             Tampa         FL          USA
#> 96           Atlanta         GA          USA
#> 97         Pensacola         FL          USA
#> 98     Miami Gardens         FL          USA
#> 99             Tampa         FL          USA
#> 100       Chesapeake         VA          USA
#> 101            Miami         FL          USA
#> 102          Pahokee         FL          USA
#> 103     Winter Haven         FL          USA
#> 104           Destin         FL          USA
#> 105            Logan         OH          USA
#> 106           Hahira         GA          USA
#> 107          Atlanta         GA          USA
#> 108          Pridgen         GA          USA
#> 109            Miami         FL          USA
#> 110 District Heights         MD          USA
#> 111           Buford         GA          USA