Team Roster Get a teams full roster by year. If year not selected, API defaults to most recent year (2020 as of 9/22/20)

Team Roster Get a teams full roster by year. If year not selected, API defaults to most recent year (2020 as of 9/22/20)

cfb_team_roster(team, year = NULL)

Arguments

team
(String required): Team, select a valid team in D-I football

year
(Integer optional): Year, 4 digit format (YYYY)

Source

https://api.collegefootballdata.com/roster

Value

A data frame with 12 variables:

athlete_id

character.

first_name

character.

last_name

character.

weight

integer.

height

integer.

jersey

integer.

year

integer.

position

character.

home_city

character.

home_state

character.

home_country

character.

team

character.

Examples

cfb_team_roster("Florida State")

#> 2020-11-03 01:22:29: Scraping team roster...

#>     athlete_id first_name    last_name weight height jersey year position
#> 1      4241674       Josh       Newman    200     72     21    3        S
#> 2      3921574      Kevin      Robledo    184     72     89    4       PK
#> 3      4035603      Malik        Henry    195     74     16    3       QB
#> 4      4052022       Trey       Carter    300     73     63    4       OL
#> 5      4426817     Jarvis     Brownlee    172     71     19    1       DB
#> 6      4240023       Tre'      McKitty    245     77     NA    4       TE
#> 7      4363030     Warren     Thompson    201     75     11    2       WR
#> 8      4363049    Malcolm        Lamar    277     77     93    2       DL
#> 9      4240022     Khalan       Laborn    205     71      4    2       RB
#> 10     3886609       Alex   Hornibrook    220     76     12    4       QB
#> 11     4428783   Treshaun         Ward    191     70     38    1       RB
#> 12     4363041     Xavier       Peters    235     76      5    1       DE
#> 13     4240033    Ontaria       Wilson    174     72     80    3       WR
#> 14     4035619       Josh        Brown    229     75     51    4       LB
#> 15     3921955       Ryan      Roberts    300     78     56    4       OL
#> 16     4035618     Cedric         Wood    318     75     49    3       DT
#> 17     4035621     Emmett         Rice    222     74     56    4       LB
#> 18     4035602 Dontavious      Jackson    229     75      5    4       LB
#> 19     4363037       A.J.       Lytton    179     71     12    2       DB
#> 20     4035606      Logan        Tyler    191     72     21    4        P
#> 21     4363045     Camren     McDonald    238     76     87    3       TE
#> 22     4240030     Hamsah  Nasirildeen    220     76      5    4       DB
#> 23     4428664        Tru     Thompson    301     72     90    2       DT
#> 24     4240028  Tamorrion        Terry    210     76      5    3       WR
#> 25     4426427    Raymond   Woodie III    194     72     31    2       DB
#> 26     4427291      Kalen      DeLoach    213     73     20    2       LB
#> 27     4426397     Jaleel        McRae    225     74     14    2       LB
#> 28     4577580       Gino      English    198     73     15    1       QB
#> 29     4426456     Dontae        Lucas    315     75     55    2       OL
#> 30     4035620     Baveon      Johnson    307     75     51    4       OL
#> 31     4427325    Renardo        Green    185     72     36    2       DB
#> 32     4240042       Cory       Durden    315     77     16    3       DT
#> 33     4240021        Cam        Akers    212     71      3    3       RB
#> 34     4360468     Jaiden Lars-Woodbey    225     73      6    2       DB
#> 35     4363047     Robert       Cooper    332     74     91    3       DT
#> 36     4363044     Jordan        Young    208     74     83    2       WR
#> 37     4430153    Derrick  McLendon II    251     76     55    1       DE
#> 38     3692942     George     Campbell    183     76     15    4       WR
#> 39     4363034     Asante   Samuel Jr.    184     70     26    3       DB
#> 40     4240026     Joshua      Kaindoh    265     79     13    3       DE
#> 41     4035605     Carlos   Becker III    195     74     15    4       DB
#> 42     4035613       Kyle       Meyers    174     72     14    4       DB
#> 43     4363040     Dennis   Briggs Jr.    282     76     58    2       DT
#> 44     4240031    DeCalon       Brooks    210     70     28    3       LB
#> 45     4035604   Janarius     Robinson    260     77     11    4       DE
#> 46     4035601    Levonta       Taylor    190     70      1    4       DB
#> 47     4035628      Keith        Gavin    212     75     89    4       WR
#> 48     4360469      Amari       Gainer    227     75     33    2       LB
#> 49     4568255      Akeem         Dent    191     73     27    2       DB
#> 50     4240029     Marvin       Wilson    305     77     21    4       DT
#> 51     4240024   Stanford  Samuels III    185     74      8    3       DB
#> 52     4240034       D.J.     Matthews    154     70      7    4       WR
#> 53     4240020      James     Blackman    190     77      1    3       QB
#> 54     4035608      Ricky       Aguayo    215     74     23    4       PK
#> 55     4240027      Cyrus        Fagan    195     73      4    4       DB
#> 56     4363046  Tre'Shaun     Harrison    191     74     88    2       WR
#> 57     3917228     Tyrell       Moorer    199     71     85    4       WR
#> 58     4568260  Alexander     Strozier    302     74     62    4       OL
#> 59     4363029       Alex    Eleyssami    190     73     16    1       QB
#> 60     3921570      Abdul        Bello    315     78     75    4       OL
#> 61     4066033     Joseph       Garcia    218     74     47    3       LB
#> 62     4066034     Tanner      Adkison    192     70     63    3       LS
#> 63     4363035   D'Marcus        Adams    180     72     19    2       WR
#> 64     4568256   DeAundre       Emeric    196     69     39    2       DB
#> 65     4380437     Jeremy     Czerenda    285     75     68    1       OL
#> 66     4363039  Christian      Meadows    336     76     52    1       OL
#> 67     4035625      Jauan     Williams    310     79     73    3       OL
#> 68     4047688    Michael     Barulich    195     71     86    3       WR
#> 69     4360472  Christian    Armstrong    320     76     77    1       OL
#> 70     4360474      Peter       Osimen    297     74     93    4       DT
#> 71     3693281     Calvin      Brewton    184     72      4    4       DB
#> 72     4035623       Mike       Arnold    340     77     72    3       OL
#> 73     4240035     Deonte    Sheffield    209     70     14    3       RB
#> 74     4262804     Parker     Grothaus    220     74     37    3       PK
#> 75     4035611       Gabe       Nabers    243     75     32    4       TE
#> 76     3925359     Adonis       Thomas    239     76     22    4       LB
#> 77     3921569       Cole      Minshew    330     77     70    4       OL
#> 78     3921562     Marcus        Lewis    197     73      8    4       DB
#> 79     4363033    Anthony        Grant    197     71     10    2       RB
#> 80     4360799     Jordan       Travis    200     73     13    2       QB
#> 81     4240025     Bailey      Hockman    208     74     16    3       QB
#> 82     4363027      Caleb         Ward    172     73     89    2       WR
#> 83     4426445    Quashon       Fuller    275     75     30    1       DE
#> 84     4240036    Leonard   Warner III    245     76     35    4       LB
#> 85     4426396    Brendan         Gant    195     74     44    2       DB
#> 86     4273843      Grant      Glennon    233     75     44    3       LS
#> 87     4363032   Keyshawn       Helton    171     69      6    3       WR
#> 88     4363036     Isaiah       Bolden    196     74     29    2       DB
#> 89     4360470    Garrett       Murray    218     72     42    2       LS
#> 90     4363028      Nolan     McDonald    189     74     41    2       DB
#> 91     4363042      Jalen         Goss    271     79     61    2       OL
#> 92     4430678    Ricardo       Watson    245     72     54    1       LB
#> 93     4577584     Robert     Elder IV    266     74     69    2       OL
#> 94     4568261        Jay     Williams    315     78     74    4       OL
#> 95     4568258    Maurice        Smith    286     75     53    1       OL
#> 96     4568257      Tyler       Gilroy    243     71     53    3       LB
#> 97     4035622     Andrew      Boselli    321     77     60    4       OL
#> 98     4577583      Ethan      Umstead    232     73     40    1       LB
#> 99     4363048   Jamarcus      Chatman    273     74     95    2       DE
#> 100    4427288      Kevon        Glenn    229     74     10    1       LB
#> 101    4577585      Caleb        Faris    165     68     81    1       WR
#> 102    4568259       Axel         Rizo    218     74     57    2       LB
#> 103    4577582    Raekwon         Webb    203     68     46    4       RB
#> 104    4363043       Chaz         Neal    317     79     71    2       OL
#> 105    4426452     Curtis     Fann Jr.    265     75     98    1       DE
#> 106    4568263       Ryan   Fitzgerald    199     73     88    1       PK
#> 107    4240039      Brady        Scott    317     78     59    3       OL
#> 108    4429791        Ira    Henry III    313     76     78    1       OL
#> 109    4577581        Tre        Young    194     72     29    1       WR
#> 110    4568262     Darius   Washington    301     76     76    1       OL
#> 111    4428307    Malcolm          Ray    287     74     99    1       DT
#> 112    4577586     Austin        White    242     77     82    1       TE
#> 113    4259260      Tommy       Martin    188     71     30    3        P
#> 114    4360473    Adarius         Dent    175     73     84    3       WR
#> 115    4363038     Armani         Kerr    230     71     52    4       LB
#> 116    4362233      Wyatt       Rector    236     74     19    2       TE
#> 117    4426429     Travis          Jay    191     74     18    1       DB
#> 118    4372519    Jashaun       Corbin    220     72      0    2       RB
#> 119    3917820      Meiko       Dotson    195     71      9    4       DB
#> 120    4360801    Jarrett      Jackson    295     78     48    2       DT
#> 121    4426448    Jarrian        Jones    190     72      7    2       DB
#> 122    4046366   Devontay       Taylor    315     76     58    4       OL
#> 123    4362905     Fabien       Lovett    315     76      0    2       DT
#>            home_city home_state home_country          team
#> 1         Cedar Hill         TX          USA Florida State
#> 2      Thousand Oaks         CA          USA Florida State
#> 3         Long Beach         CA          USA Florida State
#> 4        Monroeville         AL          USA Florida State
#> 5      Miami Gardens         FL          USA Florida State
#> 6      Wesley Chapel         FL          USA Florida State
#> 7            Seffner         FL          USA Florida State
#> 8            Seffner         FL          USA Florida State
#> 9     Virginia Beach         VA          USA Florida State
#> 10      West Chester         PA          USA Florida State
#> 11        Plant City         FL          USA Florida State
#> 12      West Chester         OH          USA Florida State
#> 13           Ashburn         GA          USA Florida State
#> 14         Charlotte         NC          USA Florida State
#> 15        Northville         MI          USA Florida State
#> 16       Tallahassee         FL          USA Florida State
#> 17     Miami Gardens         FL          USA Florida State
#> 18             Alief         TX          USA Florida State
#> 19    Upper Marlboro         MD          USA Florida State
#> 20              Nixa         MO          USA Florida State
#> 21        Long Beach         CA          USA Florida State
#> 22           Concord         NC          USA Florida State
#> 23        Loganville         GA          USA Florida State
#> 24           Ashburn         GA          USA Florida State
#> 25          Palmetto         FL          USA Florida State
#> 26          Savannah         GA          USA Florida State
#> 27  New Smyrna Beach         FL          USA Florida State
#> 28       Winter Park         FL          USA Florida State
#> 29          Overtown         FL          USA Florida State
#> 30          Lakeland         FL          USA Florida State
#> 31           Orlando         FL          USA Florida State
#> 32          Newberry         FL          USA Florida State
#> 33           Clinton         MS          USA Florida State
#> 34           Fontana         CA          USA Florida State
#> 35           Lilburn         GA          USA Florida State
#> 36           Conyers         GA          USA Florida State
#> 37           Decatur         GA          USA Florida State
#> 38        Clearwater         FL          USA Florida State
#> 39           Sunrise         FL          USA Florida State
#> 40         Baltimore         MD          USA Florida State
#> 41         Kissimmee         FL          USA Florida State
#> 42       New Orleans         LA          USA Florida State
#> 43         Kissimmee         FL          USA Florida State
#> 44             Tampa         FL          USA Florida State
#> 45       Panama City         FL          USA Florida State
#> 46    Virginia Beach         VA          USA Florida State
#> 47     Crawfordville         FL          USA Florida State
#> 48       Tallahassee         FL          USA Florida State
#> 49           Pahokee         FL          USA Florida State
#> 50           Houston         TX          USA Florida State
#> 51    Pembroke Pines         FL          USA Florida State
#> 52      Jacksonville         FL          USA Florida State
#> 53         South Bay         FL          USA Florida State
#> 54          Mascotte         FL          USA Florida State
#> 55     Daytona Beach         FL          USA Florida State
#> 56           Seattle         WA          USA Florida State
#> 57      Jacksonville         FL          USA Florida State
#> 58            Bartow         FL          USA Florida State
#> 59           Orlando         FL          USA Florida State
#> 60             Warri       <NA>      Nigeria Florida State
#> 61             Miami         FL          USA Florida State
#> 62       Tallahassee         FL          USA Florida State
#> 63      Ormond Beach         FL          USA Florida State
#> 64           Orlando         FL          USA Florida State
#> 65   Fort Lauderdale         FL          USA Florida State
#> 66         Montezuma         GA          USA Florida State
#> 67        Washington         DC          USA Florida State
#> 68           Orlando         FL          USA Florida State
#> 69     Warner Robins         GA          USA Florida State
#> 70             Tampa         FL          USA Florida State
#> 71             Miami         FL          USA Florida State
#> 72      Winter Haven         FL          USA Florida State
#> 73            Destin         FL          USA Florida State
#> 74             Logan         OH          USA Florida State
#> 75            Hahira         GA          USA Florida State
#> 76           Atlanta         GA          USA Florida State
#> 77           Pridgen         GA          USA Florida State
#> 78  District Heights         MD          USA Florida State
#> 79            Buford         GA          USA Florida State
#> 80   West Palm Beach         FL          USA Florida State
#> 81    Powder Springs         GA          USA Florida State
#> 82         Pensacola         FL          USA Florida State
#> 83        Fort Myers         FL          USA Florida State
#> 84        Snellville         GA          USA Florida State
#> 85          Lakeland         FL          USA Florida State
#> 86       Tallahassee         FL          USA Florida State
#> 87         Pensacola         FL          USA Florida State
#> 88             Tampa         FL          USA Florida State
#> 89             Tampa         FL          USA Florida State
#> 90        Long Beach         CA          USA Florida State
#> 91          Valdosta         GA          USA Florida State
#> 92             Tampa         FL          USA Florida State
#> 93            Stuart         FL          USA Florida State
#> 94     Moreno Valley         CA          USA Florida State
#> 95             Miami         FL          USA Florida State
#> 96   Saint Augustine         FL          USA Florida State
#> 97      Jacksonville         FL          USA Florida State
#> 98         Lake City         FL          USA Florida State
#> 99              Rome         GA          USA Florida State
#> 100        Riverdale         GA          USA Florida State
#> 101            Tampa         FL          USA Florida State
#> 102            Miami         FL          USA Florida State
#> 103      Panama City         FL          USA Florida State
#> 104      Zephyrhills         FL          USA Florida State
#> 105        Stillmore         GA          USA Florida State
#> 106         Coolidge         GA          USA Florida State
#> 107   Powder Springs         GA          USA Florida State
#> 108      Saint Louis         MO          USA Florida State
#> 109          Atlanta         GA          USA Florida State
#> 110        Pensacola         FL          USA Florida State
#> 111    Miami Gardens         FL          USA Florida State
#> 112            Tampa         FL          USA Florida State
#> 113       Chesapeake         VA          USA Florida State
#> 114          Pahokee         FL          USA Florida State
#> 115            Miami         FL          USA Florida State
#> 116         Leesburg         FL          USA Florida State
#> 117       Greenville         FL          USA Florida State
#> 118        Rockledge         FL          USA Florida State
#> 119    Daytona Beach         FL          USA Florida State
#> 120  West Palm Beach         FL          USA Florida State
#> 121          Brandon         MS          USA Florida State
#> 122           Odessa         FL          USA Florida State
#> 123     Olive Branch         MS          USA Florida State