Team Roster Get a teams full roster by year. If team is not selected, API returns rosters for every team from the selected year.

Team Roster Get a teams full roster by year. If team is not selected, API returns rosters for every team from the selected year.

cfb_team_roster(year, team = NULL)

Arguments

year
(Integer required): Year, 4 digit format (YYYY)

team
(String optional): Team, select a valid team in D-I football

Source

https://api.collegefootballdata.com/roster

Value

A data frame with 12 variables:

athlete_id

character.

first_name

character.

last_name

character.

weight

integer.

height

integer.

jersey

integer.

year

integer.

position

character.

home_city

character.

home_state

character.

home_country

character.

team

character.

Examples

cfb_team_roster(year = 2013, team = "Florida State")

#> 2020-12-21 07:45:43: Scraping team roster...

#>     athlete_id first_name        last_name          team weight height jersey
#> 1       501097      Nigel          Terrell Florida State     NA     74     43
#> 2       534576       P.J.         Williams Florida State     NA     72     26
#> 3       501075       Chad            Abram Florida State     NA     72     41
#> 4       546778       Nate          Andrews Florida State    214     72     29
#> 5       517388      James       Wilder Jr. Florida State     NA     NA     NA
#> 6       517377       Nick          O'Leary Florida State     NA     75     35
#> 7       535422    Roberto           Aguayo Florida State     NA     73     19
#> 8       546795   Jonathan        Hernandez Florida State    230     76     44
#> 9       503330      Kevin           Haplea Florida State     NA     76     33
#> 10      514119     Kelvin         Benjamin Florida State     NA     77      1
#> 11      501076   Terrence           Brooks Florida State     NA     71     31
#> 12      546810      Jalen           Ramsey Florida State     NA     73      8
#> 13      546805       E.J.       Levenberry Florida State    250     75     12
#> 14      551553      Jared          Jackson Florida State    199     74     87
#> 15      501082     Jarred          Haggins Florida State     NA     72     12
#> 16      546814   DeMarcus           Walker Florida State     NA     76     44
#> 17      546817      Jesus           Wilson Florida State     NA     70      3
#> 18      501535      Kenny             Shaw Florida State     NA     72     81
#> 19      501083  Christian            Jones Florida State     NA     76      7
#> 20      546809        Cam           Ponder Florida State     NA     NA     NA
#> 21      530282      Cason           Beatty Florida State     NA     75     38
#> 22      517381     Keelin            Smith Florida State     NA     75     37
#> 23      517387       Nick          Waisome Florida State     NA     70      6
#> 24      530299       Sean          Maguire Florida State     NA     75     10
#> 25      517364      Tyler           Hunter Florida State     NA     71      1
#> 26      501095     Telvin            Smith Florida State     NA     75     22
#> 27      530289     Ronald            Darby Florida State     NA     71      3
#> 28      530301     Reggie         Northrup Florida State     NA     73      5
#> 29      480772        Dan            Hicks Florida State     NA     76      6
#> 30      546816     Kermit        Whitfield Florida State     NA     68      8
#> 31      514136     Rashad       Greene Sr. Florida State     NA     NA     NA
#> 32      530303     Justin           Shanks Florida State     NA     74     92
#> 33      514121   Lamarcus           Brutus Florida State     NA     72     42
#> 34      546800     Isaiah            Jones Florida State    195     76     88
#> 35      546804      Bryan         LaCivita Florida State    178     72     83
#> 36      546813    Matthew           Thomas Florida State    227     76      6
#> 37      530308     Jameis          Winston Florida State     NA     76      5
#> 38      530302      Mario           Pender Florida State     NA     70      7
#> 39      501077     Shayne          Broxsie Florida State     NA     74     42
#> 40      530285       Will          Burnham Florida State     NA     72     45
#> 41      546794       Ryan            Green Florida State    204     70      7
#> 42      480781    Jacobbi         McDaniel Florida State     NA     72     55
#> 43      501553  Christian            Green Florida State     NA     74     89
#> 44      517390     Karlos         Williams Florida State     NA     73      9
#> 45      514134    Devonta          Freeman Florida State     NA     69      8
#> 46      514124       Jake            Coker Florida State     NA     77     14
#> 47      546815    Marquez            White Florida State     NA     72     27
#> 48      546801     Jeremy             Kerr Florida State     NA     78     85
#> 49      546812    Freddie        Stevenson Florida State     NA     73     23
#> 50      517382   Terrance            Smith Florida State     NA     76     24
#> 51      501084   Lamarcus           Joyner Florida State     NA     68     20
#> 52      530290      Mario      Edwards Jr. Florida State     NA     NA     NA
#> 53      546798 Ro'Derrick          Hoskins Florida State    240     75     18
#> 54     3921562     Marcus            Lewis Florida State    197     73      8
#> 55    -1010895     Wilson             Bell Florida State    320     77     78
#> 56    -1010896      Keith           Bryant Florida State    274     74     95
#> 57    -1010918      Keith            Weeks Florida State    280     76     66
#> 58    -1010920      Colin            Blake Florida State    198     75     30
#> 59    -1010923      Chris           Casher Florida State    256     76     21
#> 60    -1010933    Barrett           Kernon Florida State    247     76     69
#> 61    -1010936       Nile Lawrence-Stample Florida State    302     73     99
#> 62    -1010937     Tyrell            Lyons Florida State    219     73     22
#> 63    -1010939    Derrick    Mitchell, Jr. Florida State    305     76     11
#> 64    -1010941    Giorgio         Newberry Florida State    295     78      4
#> 65    -1010950      Danny            Adams Florida State    193     69     40
#> 66    -1010952     Austin           Barron Florida State    293     75     62
#> 67    -1010953     Marcel       Benalcazar Florida State    305     78     76
#> 68    -1010954       Brad            Bentz Florida State    265     74     93
#> 69    -1010956      Ruben           Carter Florida State    306     76     65
#> 70    -1010957       Troy             Cook Florida State    198     74     19
#> 71    -1010961      Mario     Edwards, Jr. Florida State    294     75     15
#> 72    -1010962      Ukeme           Eligwe Florida State    239     74     52
#> 73    -1010963    Cameron           Erving Florida State    308     78     75
#> 74    -1010964       John     Franklin III Florida State    180     72     11
#> 75    -1010965     Rashad         Gholston Florida State    170     70     13
#> 76    -1010966      Eddie          Goldman Florida State    320     76     90
#> 77    -1010967     Rashad           Greene Florida State    180     72     80
#> 78    -1010968      Bobby             Hart Florida State    320     76     51
#> 79    -1010970       Ryan          Hoefeld Florida State    290     75     59
#> 80    -1010971    Desmond           Hollin Florida State    285     75     43
#> 81    -1010972       Tre'          Jackson Florida State    330     76     54
#> 82    -1010973       Cody              Jay Florida State    250     73     60
#> 83    -1010975      Larry             Levy Florida State    275     74     63
#> 84    -1010977      Josue           Matias Florida State    325     78     70
#> 85    -1010979        Dan          O'Neill Florida State    175     72     83
#> 86    -1010983     Andrew           Wright Florida State    214     72     59
#> 87    -1010984      Jayon            Young Florida State    223     73     46
#> 88    -1010986   Jonathan           Akanbi Florida State     NA     NA     38
#> 89    -1010987       Ryan           Alicea Florida State    213     73     58
#> 90    -1010988     Justin           Allred Florida State    197     73     39
#> 91    -1010989       Paul           Aloise Florida State    225     71     46
#> 92    -1010990      Jared        Armstrong Florida State    185     72     23
#> 93    -1010991       Ryan           Arnold Florida State    250     75     58
#> 94    -1010992    Davarez           Bryant Florida State    247     75     82
#> 95    -1010993       Kory          Burnett Florida State    205     72     47
#> 96    -1010994      Peter        Ciappetta Florida State    211     74     43
#> 97    -1010995      Jacob            Coker Florida State    230     77     14
#> 98    -1010996     Parker           Conley Florida State    210     77     40
#> 99    -1010997     Tasean            Crews Florida State    175     71     86
#> 100   -1010998      Myles            Davis Florida State    307     74     61
#> 101   -1010999    Stephen            Davis Florida State    195     71     16
#> 102   -1011000     Gerald            Demps Florida State    200     71     23
#> 103   -1011001        Ira           Denson Florida State    330     75     55
#> 104   -1011002     Hunter           Dewitt Florida State    185     76     12
#> 105   -1011003   Reginald            Dixon Florida State    207     70     67
#> 106   -1011004     Philip           Doumar Florida State    225     73     57
#> 107   -1011005     Dorian           Earley Florida State    215     74     54
#> 108   -1011006        Ebo          Entsuah Florida State    240     71     43
#> 109   -1011007      Jacob       Fahrenkrug Florida State    326     76     60
#> 110   -1011008    Matthew             Finn Florida State    152     72     15
#> 111   -1011009  Christian         Griffith Florida State    210     78     18
#> 112   -1011010        Joe        Hernandez Florida State    210     72     73
#> 113   -1011011      Timmy         Jernigan Florida State    292     74      8
#> 114   -1011012     Pierre        Jolicouer Florida State    250     72     91
#> 115   -1011013      Jeren       Kowalewski Florida State    175     73     88
#> 116   -1011014      Larry       Lawson III Florida State    190     72      6
#> 117   -1011015   Sterling         Lovelady Florida State    291     75     53
#> 118   -1011016    Demonte       McAllister Florida State    290     74     97
#> 119   -1011017     Dorsey            Moore Florida State    205     76     16
#> 120   -1011018       Clay          Pickler Florida State    180     75     16
#> 121   -1011019    Cameron           Ponder Florida State    240     73     44
#> 122   -1011020       Jake           Romano Florida State    236     71     49
#> 123   -1011021     Junior      Saint Louis Florida State    218     74     48
#> 124   -1011022    Michael       Scheerhorn Florida State    240     76     79
#> 125   -1011023    Garrett            Scott Florida State    275     75     69
#> 126   -1011024     Nathan           Slater Florida State    223     74     45
#> 127   -1011025      Bryan            Stork Florida State    300     76     52
#> 128   -1011026    Donovan             Todd Florida State    205     71     39
#> 129   -1011027   Jonathan          Wallace Florida State    295     79     74
#> 130   -1011028   Jermaine       Washington Florida State    194     68     36
#> 131   -1011029    Levonte        Whitfield Florida State    178     67      7
#> 132   -1011030      James      Wilder, Jr. Florida State    229     74     32
#> 133   -1011031     Colton          Woodall Florida State    190     75     49
#> 134    3122936      Mavin         Saunders Florida State    256     77     89
#>     year position          home_city home_state home_country
#> 1     NA       LB             Pelham         AL          USA
#> 2     NA       DB              Ocala         FL          USA
#> 3     NA       FB           Lakeland         FL          USA
#> 4     NA       DB            Leander         TX          USA
#> 5     NA       WR               <NA>       <NA>         <NA>
#> 6     NA       TE Palm Beach Gardens         FL          USA
#> 7     NA       PK           Mascotte         FL          USA
#> 8     NA        P            Houston         TX          USA
#> 9     NA       TE          Annandale         NJ          USA
#> 10    NA       WR        Belle Glade         FL          USA
#> 11    NA       DB          Dunnellon         FL          USA
#> 12    NA       DB             Smyrna         TN          USA
#> 13    NA       LB            Chicago         IL          USA
#> 14    NA       WR            Lubbock         TX          USA
#> 15    NA       WR           Lakeland         FL          USA
#> 16    NA       DE       Jacksonville         FL          USA
#> 17    NA       WR              Miami         FL          USA
#> 18    NA       WR            Orlando         FL          USA
#> 19    NA       LB        Winter Park         FL          USA
#> 20    NA     <NA>               <NA>       <NA>         <NA>
#> 21    NA        P          Charlotte         NC          USA
#> 22    NA       DB        Fort Pierce         FL          USA
#> 23    NA       DB          Groveland         FL          USA
#> 24    NA       QB             Sparta         NJ          USA
#> 25    NA       DB           Valdosta         GA          USA
#> 26    NA       LB           Valdosta         GA          USA
#> 27    NA       CB          Oxon Hill         MD          USA
#> 28    NA       LB       Jacksonville         FL          USA
#> 29    NA       DE             Oxford         MS          USA
#> 30    NA       WR            Orlando         FL          USA
#> 31    NA     <NA>               <NA>       <NA>         <NA>
#> 32    NA       DT         Prattville         AL          USA
#> 33    NA       DB     Port St. Lucie         FL          USA
#> 34    NA       WR            Lubbock         TX          USA
#> 35    NA       WR        Gainesville         FL          USA
#> 36    NA       LB            Chicago         IL          USA
#> 37    NA       QB           Hueytown         AL          USA
#> 38    NA       RB         Cape Coral         FL          USA
#> 39    NA       TE         Monticello         FL          USA
#> 40    NA       RB             Lithia         FL          USA
#> 41    NA       RB          Ewa Beach         HI          USA
#> 42    NA       DT         Greenville         FL          USA
#> 43    NA       WR              Tampa         FL          USA
#> 44    NA       RB          Davenport         FL          USA
#> 45    NA       RB              Miami         FL          USA
#> 46    NA       QB             Mobile         AL          USA
#> 47    NA       DB             Dothan         AL          USA
#> 48    NA       TE     St. Petersburg         FL          USA
#> 49    NA       FB             Bartow         FL          USA
#> 50    NA       LB            Decatur         GA          USA
#> 51    NA       DB     Ft. Lauderdale         FL          USA
#> 52    NA       RB               <NA>       <NA>         <NA>
#> 53    NA       LB            Chicago         IL          USA
#> 54     4       DB   District Heights         MD          USA
#> 55    NA       OL             Mobile         AL          USA
#> 56    NA       DE       Delray Beach         FL          USA
#> 57    NA       OL            Valrico         FL          USA
#> 58    NA       DB        San Antonio         TX          USA
#> 59    NA       DE             Mobile         AL          USA
#> 60    NA       LS       Ormond Beach         FL          USA
#> 61    NA       DT    Fort Lauderdale         FL          USA
#> 62    NA       DB       Jacksonville         FL          USA
#> 63    NA       DT       Jacksonville         FL          USA
#> 64    NA       DT        Fort Pierce         FL          USA
#> 65    NA       LS     Ft. Lauderdale         FL          USA
#> 66    NA        C     Ft. Lauderdale         FL          USA
#> 67    NA       OL            Miramar         FL          USA
#> 68    NA       DL         Wellington         FL          USA
#> 69    NA       OL              Miami         FL          USA
#> 70    NA       QB              Miami         FL          USA
#> 71    NA       DE            Gautier         MS          USA
#> 72    NA       LB     Stone Mountain         GA          USA
#> 73    NA       OT           Moultrie         GA          USA
#> 74    NA       QB         Plantation         FL          USA
#> 75    NA       WR      Miami Gardens         FL          USA
#> 76    NA       DT         Washington         DC          USA
#> 77    NA       WR             Albany         GA          USA
#> 78    NA       OT         Lauderhill         FL          USA
#> 79    NA        C             Kenner         LA          USA
#> 80    NA       DE              Miami         FL          USA
#> 81    NA        G              Jesup         GA          USA
#> 82    NA       OL  Fort Walton Beach         FL          USA
#> 83    NA       OL      Boynton Beach         FL          USA
#> 84    NA        G         Union City         NJ          USA
#> 85    NA       WR            Jupiter         FL          USA
#> 86    NA       LB          Bradenton         FL          USA
#> 87    NA       TE      Coral Springs         FL          USA
#> 88    NA        ?                                        USA
#> 89    NA       LB         Wellington         FL          USA
#> 90    NA       DB      St. Augustine         FL          USA
#> 91    NA       LB          Palm City         FL          USA
#> 92    NA       QB        Gainesville         FL          USA
#> 93    NA       OL            Orlando         FL          USA
#> 94    NA       DE          Crestview         FL          USA
#> 95    NA       DB       Jacksonville         FL          USA
#> 96    NA       LB     Ft. Lauderdale         FL          USA
#> 97    NA       QB             Mobile         AL          USA
#> 98    NA       TE           Norcross         GA          USA
#> 99    NA       WR          Pensacola         FL          USA
#> 100   NA       OL        Gainesville         FL          USA
#> 101   NA       DB             Oviedo         FL          USA
#> 102   NA       DB           Valdosta         GA          USA
#> 103   NA       OL         Greenville         FL          USA
#> 104   NA       QB           Sarasota         FL          USA
#> 105   NA       LB        Tallahassee         FL          USA
#> 106   NA        ?            Jupiter         FL          USA
#> 107   NA       LB     St. Petersburg         FL          USA
#> 108   NA       FB           Clermont         FL          USA
#> 109   NA       OL        Robbinsdale         MN          USA
#> 110   NA       PK           Maitland         FL          USA
#> 111   NA       WR    West Palm Beach         FL          USA
#> 112   NA       LB        Miami Beach         FL          USA
#> 113   NA       DT          Lake City         FL          USA
#> 114   NA       DT              Miami         FL          USA
#> 115   NA       WR          Bradenton         FL          USA
#> 116   NA        P              Tampa         FL          USA
#> 117   NA       OL            Navarre         FL          USA
#> 118   NA       DT              Tampa         FL          USA
#> 119   NA       TE          Rockledge         FL          USA
#> 120   NA       PK          Grapevine         TX          USA
#> 121   NA       FB              Yulee         FL          USA
#> 122   NA       FB         Wellington         FL          USA
#> 123   NA       LB      Pompano Beach         FL          USA
#> 124   NA       OL               Lutz         FL          USA
#> 125   NA       OL    West Palm Beach         FL          USA
#> 126   NA       LB              Terry         MS          USA
#> 127   NA       OL         Vero Beach         FL          USA
#> 128   NA       DB            Sanford         FL          USA
#> 129   NA       OL       Jacksonville         FL          USA
#> 130   NA       WR       Jacksonville         FL          USA
#> 131   NA       WR            Orlando         FL          USA
#> 132   NA       RB              Tampa         FL          USA
#> 133   NA       DB     Ft. Lauderdale         FL          USA
#> 134    4       TE               <NA>       <NA>      Bahamas